### 2023-12-14 #6151 @editorialbot
[PRE REVIEW]: Empirical: A scientific software library for research, education, and public engagement
**Submitting author:** @mmore500 (<a href="http://orcid.org/0000-0003-4726-4479">Matthew Andres Moreno</a>)
**Repository:** https://github.com/devosoft/Empirical/
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.0.0
