### 2023-12-14 #6151 @danielskatz
@mmore500 - I see you have an open source license in the repo, but for some reason, it's not in the joss-paper branch, which is where we are looking for it.

Also, this is a very large codebase - Are you requesting that JOSS reviews all of it?
