### 2023-12-14 #6151 @danielskatz
@mmore500 - additionally, JOSS papers are expected to be 250-1000 words, give or take.  Your paper is almost 3000 words, which will need to be cut substantially before we could review it.

I'm also a little unclear if this is [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) as defined by JOSS. Could you say a bit about the kinds of researchers you expect to use this software, and if you expect them to cite it in their papers?
