### 2023-12-14 #6151 @mmore500
Hi @danielskatz

Thanks for the quick response!

We consider Empirical as best fitting under JOSS's research definition as a tool to  facilitate "execution of research experiments." Indeed, the library originally arose through a group effort to combine and organize simulation-based-experiment-specific software components and tools to facilitate their re-use across agent-based simulation experiments. You can see some of existing research work that has leveraged Empirical and cited our Zenodo software release [here](https://scholar.google.com/scholar?hl=en&as_sdt=0%2C23&q=10.5281%2Fzenodo.4141943&btnG=).

You're right that it is a large code base, reflecting our objective to provide a comprehensive collection of tools for common tasks in simulation experiments (i.e., runtime configuration, debug-friendly containers). We see library components for creating public-facing interactive web apps from C++ research software via WebAssembly as dovetailing closely to this research toolbox objective. (Among other research-specific features, the UI toolkit include special components for experiment configuration panels and real-time simulation stepping control). So, to answer your question, we are hoping to engage in a review process that considers Empirical as a software library (where end-users pick and choose relevant components to incorporate into their own simulation software) rather than as a monolithic software application.

The one part we did note for exclusion from the review is @emilydolson 's phylogeny-tracking tools located at <https://github.com/devosoft/Empirical/tree/joss-paper/include/emp/Evolve/Systematics.hpp> are hosted within the repository but considered a separate scholarly effort (linked in the submission blurb).

Noted that the JOSS manuscript branch had fallen a bit behind --- merged into it from master to bring it back up to date. Thanks for bringing that to my attention.

Also, noted on the word count. I'll confer with the coauthors to strip the word count down. I'll update this thread once that task is completed in a few days?

Thanks again for the discussion here, and let us know how to proceed.
