### 2023-12-14 #6151 @danielskatz
@mmore500 - Let me get the JOSS editors to review the scope, to make sure we think it is research software, before you work on cutting the paper
