### 2023-12-15 9c60046
Cleaned up concepts and made work with is_span.
