### 2023-12-16 b4c7bf0
Fixed isIterable usage.
