### 2023-12-19 4158e7d
Specified (and implemented) a meaningful resolution ordering for Load options.
