### 2023-12-21 8f55f9a
Set up serial pods to always be able to take multiple values at once.
