### 2023-12-21 #6151 @mmore500
Hi @arfon -

Happy to clarify! #5754 presents the phylotrack C++ library codebase, which specializes on the problem of lineage tracking in agent-based evolutionary simulation. It originated as an Empirical library component early in Empirical’s development history several years ago, but has since grown in scope and begun being distributed independently of the rest of Empirical — most notably, as Python bindings via the `phylotrackpy` PyPI package. With respect to ongoing development, @emilydolson, @jeetsukumaran and I have an independent development roadmap for phylotrack focusing on incorporation with the Python bioinformatics ecosystem.

Due to phylotrack’s header-only source, it can be used largely independently of Empirical and vice versa. It’s located within the Empirical monorepo primarily due to historical factors and convenience in maintaining revision history as well as keeping existing development infrastructure (CI, coverage, documentation build, etc.).
