### 2023-12-26 693489c
Added unit tests for external Serialize() functions.
