### 2023-12-27 c26a651
Added SerialPod::LoadValue() and functions for serializing vectors.
