### 2023-12-28 d7e0632
Removed unneeded pre-declarations from SerialPod
