### 2023-12-29 5059216
Changed concept descriptions to be in the form of a question.
