### 2023-12-29 b2cbd95
Added tests for internal SerialSave and SerialLoad
