### 2023-12-30 14f4e48
New tests for EXTERNAL SerialSave and SerialLoad added
