### 2023-12-31 c5927f7
Added a test to SerialPod with a nested class structure + other cleanup
