### 2024-01-03 8577119
Fixed TypeID to use new concepts; added a GetTypeName<T>() function.
