### 2024-01-04 3d1a62a
Added type trait variant to work on type pack (with any, all, and no options)
