+++ include/emp/meta/type_traits.hpp
@@ -1,7 +1,7 @@
 /**
  *  @note This file is part of Empirical, https://github.com/devosoft/Empirical
  *  @copyright Copyright (C) Michigan State University, MIT Software license; see doc/LICENSE.md
- *  @date 2016-2023.
+ *  @date 2016-2024.
  *
  *  @file type_traits.hpp
  *  @brief Extensions on the standard library type traits to handle Empirical classes (such as Ptr).
@@ -23,6 +23,166 @@
 
 
 namespace emp {
+
+  /// Type trait framework for working with whole TYPE PACKS.
+  template <template <typename...> typename TRAIT, typename... Ts>
+  struct has_trait_any : std::false_type {};
+  template <template <typename...> typename TRAIT, typename T, typename... EXTRA_Ts>
+  struct has_trait_any<TRAIT, T, EXTRA_Ts...>
+      : std::bool_constant<TRAIT<T>::value || has_trait_any<TRAIT, EXTRA_Ts...>::value> {};
+
+  template <template <typename...> typename TRAIT, typename... Ts>
+  struct has_trait_all : std::true_type {};
+  template <template <typename...> typename TRAIT, typename T, typename... EXTRA_Ts>
+  struct has_trait_all<TRAIT, T, EXTRA_Ts...>
+      : std::bool_constant<TRAIT<T>::value && has_trait_all<TRAIT, EXTRA_Ts...>::value> {};
+
+  template <template <typename...> typename TRAIT, typename... Ts>
+  struct has_trait_none : std::negate<emp::has_trait_any<TRAIT, Ts...>> {};
+
+
+  template <typename... Ts> using has_any_void = has_trait_any<std::is_void, Ts...>;
+  template <typename... Ts> using has_all_void = has_trait_all<std::is_void, Ts...>;
+  template <typename... Ts> using has_no_void = has_trait_none<std::is_void, Ts...>;
+
+  template <typename... Ts> using has_any_null_pointer = has_trait_any<std::is_null_pointer, Ts...>;
+  template <typename... Ts> using has_all_null_pointer = has_trait_all<std::is_null_pointer, Ts...>;
+  template <typename... Ts> using has_no_null_pointer = has_trait_none<std::is_null_pointer, Ts...>;
+
+  template <typename... Ts> using has_any_integral = has_trait_any<std::is_integral, Ts...>;
+  template <typename... Ts> using has_all_integral = has_trait_all<std::is_integral, Ts...>;
+  template <typename... Ts> using has_no_integral = has_trait_none<std::is_integral, Ts...>;
+
+  template <typename... Ts> using has_any_floating_point = has_trait_any<std::is_floating_point, Ts...>;
+  template <typename... Ts> using has_all_floating_point = has_trait_all<std::is_floating_point, Ts...>;
+  template <typename... Ts> using has_no_floating_point = has_trait_none<std::is_floating_point, Ts...>;
+
+  template <typename... Ts> using has_any_array = has_trait_any<std::is_array, Ts...>;
+  template <typename... Ts> using has_all_array = has_trait_all<std::is_array, Ts...>;
+  template <typename... Ts> using has_no_array = has_trait_none<std::is_array, Ts...>;
+
+  template <typename... Ts> using has_any_enum = has_trait_any<std::is_enum, Ts...>;
+  template <typename... Ts> using has_all_enum = has_trait_all<std::is_enum, Ts...>;
+  template <typename... Ts> using has_no_enum = has_trait_none<std::is_enum, Ts...>;
+
+  template <typename... Ts> using has_any_union = has_trait_any<std::is_union, Ts...>;
+  template <typename... Ts> using has_all_union = has_trait_all<std::is_union, Ts...>;
[…]
