### 2024-01-04 2dd1162
Renamed has_all_X traits to has_only_X in type_traits
