+++ include/emp/meta/type_traits.hpp
@@ -42,145 +42,145 @@ namespace emp {
 
 
   template <typename... Ts> using has_any_void = has_trait_any<std::is_void, Ts...>;
-  template <typename... Ts> using has_all_void = has_trait_all<std::is_void, Ts...>;
+  template <typename... Ts> using has_only_void = has_trait_all<std::is_void, Ts...>;
   template <typename... Ts> using has_no_void = has_trait_none<std::is_void, Ts...>;
 
   template <typename... Ts> using has_any_null_pointer = has_trait_any<std::is_null_pointer, Ts...>;
-  template <typename... Ts> using has_all_null_pointer = has_trait_all<std::is_null_pointer, Ts...>;
+  template <typename... Ts> using has_only_null_pointer = has_trait_all<std::is_null_pointer, Ts...>;
   template <typename... Ts> using has_no_null_pointer = has_trait_none<std::is_null_pointer, Ts...>;
 
   template <typename... Ts> using has_any_integral = has_trait_any<std::is_integral, Ts...>;
-  template <typename... Ts> using has_all_integral = has_trait_all<std::is_integral, Ts...>;
+  template <typename... Ts> using has_only_integral = has_trait_all<std::is_integral, Ts...>;
   template <typename... Ts> using has_no_integral = has_trait_none<std::is_integral, Ts...>;
 
   template <typename... Ts> using has_any_floating_point = has_trait_any<std::is_floating_point, Ts...>;
-  template <typename... Ts> using has_all_floating_point = has_trait_all<std::is_floating_point, Ts...>;
+  template <typename... Ts> using has_only_floating_point = has_trait_all<std::is_floating_point, Ts...>;
   template <typename... Ts> using has_no_floating_point = has_trait_none<std::is_floating_point, Ts...>;
 
   template <typename... Ts> using has_any_array = has_trait_any<std::is_array, Ts...>;
-  template <typename... Ts> using has_all_array = has_trait_all<std::is_array, Ts...>;
+  template <typename... Ts> using has_only_array = has_trait_all<std::is_array, Ts...>;
   template <typename... Ts> using has_no_array = has_trait_none<std::is_array, Ts...>;
 
   template <typename... Ts> using has_any_enum = has_trait_any<std::is_enum, Ts...>;
-  template <typename... Ts> using has_all_enum = has_trait_all<std::is_enum, Ts...>;
+  template <typename... Ts> using has_only_enum = has_trait_all<std::is_enum, Ts...>;
   template <typename... Ts> using has_no_enum = has_trait_none<std::is_enum, Ts...>;
 
   template <typename... Ts> using has_any_union = has_trait_any<std::is_union, Ts...>;
-  template <typename... Ts> using has_all_union = has_trait_all<std::is_union, Ts...>;
+  template <typename... Ts> using has_only_union = has_trait_all<std::is_union, Ts...>;
   template <typename... Ts> using has_no_union = has_trait_none<std::is_union, Ts...>;
 
   template <typename... Ts> using has_any_class = has_trait_any<std::is_class, Ts...>;
-  template <typename... Ts> using has_all_class = has_trait_all<std::is_class, Ts...>;
+  template <typename... Ts> using has_only_class = has_trait_all<std::is_class, Ts...>;
   template <typename... Ts> using has_no_class = has_trait_none<std::is_class, Ts...>;
 
   template <typename... Ts> using has_any_function = has_trait_any<std::is_function, Ts...>;
-  template <typename... Ts> using has_all_function = has_trait_all<std::is_function, Ts...>;
+  template <typename... Ts> using has_only_function = has_trait_all<std::is_function, Ts...>;
[…]
