### 2024-01-06 f8b4321
Updated serial pod to handle const and r-values correctly.
