### 2024-01-07 657e920
Added MakeRange overload that converts from a simple string.
