### 2024-01-15 e5496db
Extracted FlagInfo into its on class, outside of FlagManager.
