+++ include/emp/config/FlagManager.hpp
@@ -1,7 +1,7 @@
 /**
  *  @note This file is part of Empirical, https://github.com/devosoft/Empirical
  *  @copyright Copyright (C) Michigan State University, MIT Software license; see doc/LICENSE.md
- *  @date 2023.
+ *  @date 2023-24.
  *
  *  @file FlagManager.hpp
  *  @brief This file contains tools for dealing with command-line flags (from argv and argc).
@@ -41,18 +41,42 @@
 
 namespace emp {
 
+  class FlagInfo {
+  public:
+    using fun_t = std::function<void(const emp::vector<String> &)>;
+
+  private:
+    String desc;
+    size_t min_args = 0;
+    size_t max_args = 0;
+    fun_t fun;
+    char shortcut = '\0';
+
+  public:
+    FlagInfo() { }
+    FlagInfo(String desc, size_t min_args, size_t max_args,
+             fun_t fun, char shortcut='\0')
+      : desc(desc), min_args(min_args), max_args(max_args), fun(fun), shortcut(shortcut)
+    { }
+
+    FlagInfo & operator=(const FlagInfo &) = default;
+
+    const String & GetDesc() const { return desc; }
+    size_t GetMinArgs() const { return min_args; };
+    size_t GetMaxArgs() const { return max_args; };
+    fun_t GetFun() const { return fun; };
+    char GetShortcut() const { return shortcut; };
+
+    FlagInfo & SetShortcut(char in) { shortcut = in; return *this; }
+
+    template <typename... Ts>
+    void Run(Ts &&... args) { fun(std::forward<Ts>(args)...); }
+  };
+
   class FlagManager {
   private:
     emp::vector<String> args;
     emp::vector<String> extras;
-
-    struct FlagInfo {
-      String desc;
-      size_t min_args = 0;
-      size_t max_args = 0;
-      std::function<void(const emp::vector<String> &)> fun;
-      char shortcut = '\0';
-    };
 
     std::map<String, FlagInfo> flag_options;
     std::map<char, String> shortcuts;
@@ -87,17 +111,17 @@ namespace emp {
       flag_options[name] = FlagInfo{desc, 0, 0, [](const emp::vector<String> &){} };
     }
     void AddOption(String name, std::function<void()> fun, String desc="") {
-      flag_options[name] = FlagInfo{desc, 0,0, [fun](const emp::vector<String> &){fun();}};
+      flag_options[name] = FlagInfo{desc, 0, 0, [fun](const emp::vector<String> &){fun();}};
     }
     void AddOption(String name, std::function<void(String)> fun, String desc="") {
-      flag_options[name] = FlagInfo{desc, 1,1, [fun](const emp::vector<String> & in){fun(in[0]);}};
+      flag_options[name] = FlagInfo{desc, 1, 1, [fun](const emp::vector<String> & in){fun(in[0]);}};
     }
     void AddOption(String name, std::function<void(String,String)> fun, String desc="") {
-      flag_options[name] = FlagInfo{desc, 2,2, [fun](const emp::vector<String> & in){fun(in[0],in[1]);}};
+      flag_options[name] = FlagInfo{desc, 2, 2, [fun](const emp::vector<String> & in){fun(in[0],in[1]);}};
     }
     void AddOption(String name, std::function<void(const emp::vector<String> &)> fun,
                    size_t min_args=0, size_t max_args=npos, String desc="") {
-      flag_options[name] = FlagInfo{desc, min_args,max_args, fun};
+      flag_options[name] = FlagInfo{desc, min_args, max_args, fun};
     }
 
     // Allow an option to have a single-letter flag (e.g. "-h" is short for "--help")
@@ -105,7 +129,7 @@ namespace emp {
     void AddOption(char shortcut, String name, FUN_T fun, String desc="") {
[…]
