### 2024-01-15 8ab7004
Added the ability to categoize flag options in groups.
