### 2024-01-15 d8d05f9
Setup printing of flag options to respect groups.
