### 2024-01-21 64da120
Added debugging EMP_TRACK_LINE(name) macro and emp::GetDebugLine(name) to retrieve it later.
