+++ include/emp/debug/debug.hpp
@@ -1,7 +1,7 @@
 /**
  *  @note This file is part of Empirical, https://github.com/devosoft/Empirical
  *  @copyright Copyright (C) Michigan State University, MIT Software license; see doc/LICENSE.md
- *  @date 2015-2022
+ *  @date 2015-2024
  *
  *  @file debug.hpp
  *  @brief Basic tools for use in developing high-assurance code.
@@ -13,8 +13,12 @@
 
 #include <iostream>
 #include <set>
+#include <sstream>
 #include <string>
 #include <type_traits>
+#include <unordered_map>
+
+#include "../base/notify.hpp"
 
 #define EMP_DEBUG_PRINT(...) std::cout << "[" << #__VA_ARGS__ << "] = " << __VA_ARGS__ << std::endl
 
@@ -56,6 +60,34 @@ namespace emp {
   /// it is removed.  It's a useful too for printing "Ping1", "Ping2", etc, but no forgetting to
   /// remove them.
   #define emp_debug(...) { BlockRelease(true); emp::emp_debug_print(__VA_ARGS__); }
+
+
+  /// Track particular lines of code to report errors about them from elsewhere.
+  static auto & GetDebugLineMap() {
+    static std::unordered_map<std::string, std::string> line_map;
+    return line_map;
+  }
+
+  static bool HasDebugLine(std::string name) {
+    return GetDebugLineMap().find(name) != GetDebugLineMap().end();
+  }
+
+  static auto & GetDebugLine(std::string name) {
+    return GetDebugLineMap()[name];
+  }
+
+  static void AddDebugLine(std::string name, std::string file, size_t line) {
+    std::stringstream ss;
+    ss << file << ':' << line;
+    notify::TestError(HasDebugLine(name), "Adding a second debug line named '", name, "'.");
+    GetDebugLine(name) = ss.str();
+  }
+
+  #ifdef NDEBUG
+  #define EMP_TRACK_LINE(NAME)
+  #else
+  #define EMP_TRACK_LINE(NAME) emp::AddDebugLine(NAME, __FILE__, __LINE__)
+  #endif
 }
 
 #endif // #ifndef EMP_DEBUG_DEBUG_HPP_INCLUDE
