### 2024-01-21 8367990
Added tests for new code line tracking.
