+++ tests/debug/debug.cpp
@@ -1,7 +1,7 @@
 /**
  *  @note This file is part of Empirical, https://github.com/devosoft/Empirical
  *  @copyright Copyright (C) Michigan State University, MIT Software license; see doc/LICENSE.md
- *  @date 2021
+ *  @date 2024
  *
  *  @file debug.cpp
  */
@@ -12,4 +12,19 @@
 
 TEST_CASE("Test debug", "[debug]")
 {
+  EMP_TRACK_LINE("Test1");
+  EMP_TRACK_LINE("Test2");
+  EMP_TRACK_LINE("Test3");
+
+  CHECK(emp::GetDebugLine("Test1") == "debug.cpp:15");
+  CHECK(emp::GetDebugLine("Test2") == "debug.cpp:16");
+  CHECK(emp::GetDebugLine("Test3") == "debug.cpp:17");
+
+  EMP_TRACK_LINE("Test4");
+
+  // Make sure they all work a second time and test the new one...
+  CHECK(emp::GetDebugLine("Test1") == "debug.cpp:15");
+  CHECK(emp::GetDebugLine("Test2") == "debug.cpp:16");
+  CHECK(emp::GetDebugLine("Test3") == "debug.cpp:17");
+  CHECK(emp::GetDebugLine("Test4") == "debug.cpp:23");
 }
