### 2024-01-22 70d73ce
Updated emp::flip_map to allow specificatoion of output map type.
