### 2024-01-24 b0da47c
Updated lexer to have a TokenizeNext function that grabs a single token from a stream.
