+++ paper.md
@@ -46,164 +46,83 @@ breaks: false
 Empirical is a C++ library designed to promote open science and facilitate the development of scientific software that is efficient, reliable, and easily distributable to researchers and non-experts alike.
 Specifically, the library sets out to fulfill the following goals:
 
-1. **Utility:** Empirical features a wide selection of helper tools to streamline common scientific computing tasks such as configuration, data management, random number generation, and mathematical manipulations. Documentation, examples, and project templates streamline developer onboarding.
-2. **Efficiency:** Empirical tools emphasize efficiency to make heavy scientific computing workloads tractable. Where possible, computation is moved to compile-time, heap allocation is avoided, and memory-locality is respected.
-3. **Reliability:** Scientific simulation and data analysis is meaningful insofar as it is correct. In conjunction with extensive unit testing, Empirical benefits from --- and provides --- sophisticated debug-mode instrumentation tools. These include audited memory management, drop-in replacements for standard C++ containers with safety checking, and self-documenting assertions.
-4. **Distributability:** Scientific software should be easily used by any researcher, student, or citizen scientist who wants to experiment, explore, or replicate studies. As such, Empirical is highly portable, uses common data formats, facilitates flexible runtime configuration, and simplifies compilation to a performant web app.
-
-In addition to many helpful utilities to improve the scientific programming experience, the core Empirical support library comprises four major features:
-
- 1. debug-instrumented fundamental type and C++ standard library container wrappers,
- 2. implementations of general-purpose data structures and algorithms,
- 3. integrated, end-to-end frameworks for data and configuration management, and
- 4. object-oriented bindings for Emscripten/WebAssembly GUI elements.
+1. **Utility:** Empirical tools streamline common scientific computing tasks such as configuration, end-to-end data management, and mathematical manipulations.
+2. **Efficiency:** Empirical implements general-purpose data structures and algorithms that emphasize computational efficiency to support scientific computing workloads.
+3. **Reliability:** Empirical provides sophisticated debug-mode instrumentation including audited memory management and safety checked versions of standard library containers.
+4. **Distributability:** Empirical is highly portable, uses common data formats, and facilitates compile-to-web app development with object-oriented bindings for Emscripten/WebAssembly GUI elements, all with the goal of building broadly accessible scientific software.
 
 # Statement of Need
 
-Modern web-based interfaces give computational research the unique potential to embody open science objectives: they can make the scientific process more transparent with auditable and extensible code, clear and replicable methodologies, and production of accessible results [@woelfle2011open].
-In practice, however, many scientific software applications are difficult to obtain, install, or use, and produce data in proprietary formats.
-
-High quality open-science tools encourage researchers to follow effective software development practices by simplifying development and helping them improve code quality, scientific rigor, and ease of replication or extension.
-In the process, researchers get more out of their own code.
-For example, adding a GUI not only helps other users of the software, but could also help the researcher themselves gain "soft knowledge" of their system [@10.1145/3185517].
-
-Recent developments in web technology such as WebAssembly enable compilation of native source code to browser-based interactive interfaces.
-However, many scientists lack web development training and do not have the bandwidth to learn new languages and frameworks.
-Empirical catalyzes progress toward open science ideals by streamlining the development of in-browser software in C++.
-
-To be reuseable and extendable, software must remain robust at scales and in contexts that were not originally envisioned.
-Empirical seeks to make writing correct, efficient scientific software easier.
[…]
