### 2024-01-25 #6151 @mmore500
:+1: Just wanted to confirm to you that we are presently working to reduce word count, and are about halfway there. We anticipate the revised submission to be ready in about a week. Thank you!
