### 2024-01-26 32d270f
Renamed TokenInfo to TokenType and moved to _TokenType.hpp
