### 2024-01-27 97c71b2
Moved Token and TokenStream into their own helper files.
