### 2024-01-28 59adbe0
Restructured Lexer to move function definitions out of class declaration.
