+++ include/emp/compiler/Lexer.hpp
@@ -77,210 +77,259 @@ namespace emp {
     size_t GetNumTokens() const { return token_set.size(); }
 
     /// Reset the entire lexer to its start state.
-    void Reset() {
-      token_set.resize(0);
-      token_map.clear();
-      cur_token_id = MAX_ID;
-      generate_lexer = false;
-    }
+    void Reset();
 
     bool TokenOK(int id) const { return id >= 0 && id < cur_token_id; }
 
     /// Add a new token, specified by a name and the regex used to identify it.
     /// Note that token ids count down with highest IDs having priority.
-    int AddToken(const std::string & name,
-                 const std::string & regex,
-                 bool save_lexeme = true,
-                 bool save_token = true,
-                 const std::string & desc = "")
-    {
-      int id = cur_token_id--;                // Grab the next available token id.
-      generate_lexer = true;                  // Indicate the the lexer DFA needs to be rebuilt.
-      token_set.emplace_back( name, regex, id, save_lexeme, save_token, desc );
-      token_map[name] = id;
-      return id;
-    }
+    int AddToken(
+      const std::string & name,
+      const std::string & regex,
+      bool save_lexeme = true,
+      bool save_token = true,
+      const std::string & desc = "");
 
     /// Add a token to ignore, specified by a name and the regex used to identify it.
-    int IgnoreToken(const std::string & name, const std::string & regex, const std::string & desc="") {
-      int id = cur_token_id--;                // Grab the next available token id.
-      generate_lexer = true;                  // Indicate the the lexer DFA needs to be rebuilt.
-      token_set.emplace_back( name, regex, id, false, false, desc );
-      token_map[name] = id;
-      return id;
-    }
+    int IgnoreToken(
+      const std::string & name,
+      const std::string & regex,
+      const std::string & desc="");
 
     /// Get the ID associated with a token type (you provide the token name)
-    int GetTokenID(const std::string & name) const {
-      int default_id = ERROR_ID;
-      if (name.size() == 1) default_id = (int) name[0];
-      return emp::Find(token_map, name, default_id);
-    }
+    int GetTokenID(const std::string & name) const;
 
     /// Get the full information about a token (you provide the id)
-    const TokenType & GetTokenType(int id) const {
-      if (id > MAX_ID || id <= cur_token_id) return ERROR_TOKEN();
-      return token_set[(size_t)(MAX_ID - id)];
-    }
+    const TokenType & GetTokenType(int id) const;
 
     /// Get the name associated with a token type (you provide the ID)
-    std::string GetTokenName(int id) const {
-      if (id < 0) return emp::to_string("Error (", id, ")");   // Negative tokens specify errors.
-      if (id == 0) return "EOF";                               // Token zero is end-of-file.
-      if (id < 128) return emp::to_escaped_string((char) id);  // Individual characters.
-      if (id <= cur_token_id) return emp::to_string("Error (", id, ")"); // Invalid token range!
-      return GetTokenType(id).name;                            // User-defined tokens.
-    }
+    std::string GetTokenName(int id) const;
 
     /// Identify if a token should be saved.
-    bool GetSaveToken(int id) const {
-      if (id > MAX_ID || id <= cur_token_id) return true;
-      return GetTokenType(id).save_token;
-    }
+    bool GetSaveToken(int id) const;
 
     /// Create the NFA that will identify the current set of tokens in a sequence.
-    void Generate() const {
-      NFA lexer_nfa;
[…]
