### 2024-01-28 007eaac
Cleaned up comments throughout Lexer, converted to a template-controled max rule count.
