+++ tests/compiler/Lexer.cpp
@@ -21,7 +21,6 @@ TEST_CASE("Test Lexer", "[compiler]")
   REQUIRE(lx.GetTokenID("vowel") == 255);
   REQUIRE(lx.GetTokenName(255) == "vowel");
   // REQUIRE(lx.MaxTokenID() == 256);
-  REQUIRE(!lx.TokenOK(257));
 
   std::string word = "hello";
   lx.Process(word);
