+++ include/emp/compiler/Lexer.hpp
@@ -12,15 +12,17 @@
  *  Use AddToken(name, regex) to list out the relevant tokens.
  *   'name' is the unique name for this token.
  *   'regex' is the regular expression that describes this token.
- *  It will return a unique ID associated with this lexeme.
+ *  It will return a unique ID associated with this token that will be used in output Token
+ *  objects later.
  *
  *  IgnoreToken(name, regex) uses the same arguments, but is used for tokens that
- *  should be skipped over.
+ *  should be skipped over during lexical analysis
  *
  *  Names and IDs can be recovered later using GetTokenID(name) and GetTokenName(id).
  *
- *  Tokens can be retrieved either one at a time with Process(string) or Process(stream),
- *  which will return the next (non-ignored) token, removing it from the input.
+ *  Either strings or streams can be converted into their associated Tokens.
+ *    Process(in) does this one at a time (including "ignored" tokens)
+ *    which will return the next (non-ignored) token, removing it from the input.
  *
  *  Alternatively, an entire series of tokens can be processed with Tokenize().
  *
@@ -47,14 +49,15 @@
 namespace emp {
 
   /// A lexer with a set of token types (and associated regular expressions)
-  class Lexer {
+  /// MAX_ID determines the maximum number of allowed lexical rules.
+  template <int MAX_ID>
+  class Lexer_Base {
   private:
-    static constexpr int MAX_ID = 255;      ///< IDs count down so that first ones have priority.
     static constexpr int ERROR_ID = -1;     ///< Code for unknown token ID.
 
     emp::vector<TokenType> token_set;       ///< List of all active tokens types.
     emp::map<std::string, int> token_map;   ///< Map of token names to id.
-    int cur_token_id = MAX_ID;              ///< Which ID should the next new token get?
+    int cur_token_id = MAX_ID;              ///< ID for next new token (higher has priority)
     mutable bool generate_lexer = false;    ///< Do we need to regenerate the lexer?
     mutable DFA lexer_dfa;                  ///< Table driven lexer implementation.
     mutable std::string lexeme;             ///< Current state of lexeme being generated.
@@ -65,21 +68,13 @@ namespace emp {
     }
 
   public:
-    Lexer() = default;
-    Lexer(const Lexer &) = default;
-    Lexer(Lexer &&) = default;
-    ~Lexer() = default;
-
-    Lexer & operator=(const Lexer &) = default;
-    Lexer & operator=(Lexer &&) = default;
+    // All constructors and assignment operators are default.
 
     /// How many types of tokens can be identified in this Lexer?
     size_t GetNumTokens() const { return token_set.size(); }
 
     /// Reset the entire lexer to its start state.
     void Reset();
-
-    bool TokenOK(int id) const { return id >= 0 && id < cur_token_id; }
 
     /// Add a new token, specified by a name and the regex used to identify it.
     /// Note that token ids count down with highest IDs having priority.
@@ -149,32 +144,37 @@ namespace emp {
     void DebugToken(int token_id) const;
   };
 
+  using Lexer = Lexer_Base<255>;
 
   /// --------------------
   /// FUNCTION DEFINITIONS
   /// --------------------
 
-  void Lexer::Reset() {
+  template <int MAX_ID>
+  void Lexer_Base<MAX_ID>::Reset() {
     token_set.resize(0);
     token_map.clear();
     cur_token_id = MAX_ID;
     generate_lexer = false;
   }
 
-  int Lexer::AddToken(const std::string & name,
[…]
