### 2024-01-28 d0a6de2
Cleanup of comments in Lexer.
