### 2024-01-28 8cff9cf
Added a bunch more lexer tests.
