### 2024-01-31 9c7a96d
Added various special Resize functions to String.
