+++ paper.md
@@ -88,33 +88,26 @@ Browser-based delivery can yield particularly effective public-facing apps due t
 Empirical amplifies Emscripten by fleshing out its interface for interaction with browser elements.
 DOM elements are bound to corresponding C++ objects (e.g., `emp::Button` manages a `<button>` and `emp::Canvas` manages a `<canvas>`) and are easily manipulated from within C++.
 Empirical also packages collections of prefabricated web widgets (e.g., configuration managers or collapsible data displays).
-These tools radically simplify generating a mobile-friendly, web-based GUI for existing software.
+These tools simplify generating a mobile-friendly, web-based GUI for existing software.
 
 A live demo of Empirical widgets, presented alongside their source C++ code, is available [here](https://devosoft.github.io/empirical-prefab-demo/empirical-prefab-demo).
 
 ## Runtime Efficiency
 
-WebAssembly's runtime efficiency is a major driver of its growing impact on web app development.
-Just-in-time compilation engines translate critical sections into native machine code [@haas2017bringing], giving at least 50% --- and at times closer to 90% --- of native performance and providing an order of magnitude speed increase over JavaScript alone [@jangda2019not].
-Theis opens the door to entirely new possibilities for browser-based scientific computation.
-For example, [Avida-ED](https://avida-ed.msu.edu/) simulates hundreds of thousands of generations of digital organisms within the span of a class period.
+WebAssembly's runtime efficiency has driven its adoption in web development [@haas2017bringing], achieving 50% to 90% of native performance [@jangda2019not] and enabling new possibilities for browser-based scientific computation.
+For example, [Avida-ED](https://avida-ed.msu.edu/) simulates billions of digital organisms in a day.
 
-More broadly, Empirical supports runtime efficiency in scientific computing by providing optimized tools for performance-critical tasks.
-For example, `emp::BitArray` and `emp::BitVector` classes are faster drop-in replacements for their standard library equivalents (`std::bitset` and `std::vector<bool>`) while providing extensive additional functionality for rapid bit manipulations.
-Likewise,`emp::Random` wraps a cutting-edge high-performance pseudorandom number generator algorithm [@widynski2020squares].
-At a more fundamental level, Empirical's header-only design prioritizes ease of use and runtime performance at the cost of somewhat longer compilation times.
+More broadly, Empirical provides optimized tools for performance-critical tasks.
+For example, `emp::BitArray` and `emp::BitVector` are faster drop-in replacements for their standard library equivalents (`std::bitset` and `std::vector<bool>`) while providing extensive additional functionality.
+More fundamentally, Empirical's header-only design prioritizes ease of use and runtime performance, albeit at the cost of longer compilation times.
 
 ## Debugging
 
-Software bugs that slip through into production can inflict huge costs, especially in scientific contexts where the validity of generated data and analyses is paramount.
-
-In conjunction with unit tests and integration tests, runtime safety checks are commonly used to flag potential bugs.
-Most programming languages (Rust, Python, etc.) provide built-in support to detect common runtime violations, such as out-of-bounds indexing or bad type conversions.
-C++ does not, in an effort to maximize performance.
[…]
