### 2024-02-02 006faf5
Added error to FlagManager if duplicate flag is added.
