### 2024-02-06 #6151 @mmore500
We ended up not quite finishing the trim in our meeting last week, and have the last two paragraphs scheduled for our meeting Thursday Feb 8. So we will have a revised submission very soon!
