### 2024-02-08 6722f66
Added a Text_Rule class for managing each rule.
