### 2024-02-09 cbe8ff0
Added a Text_utils file for functions usful in multiple encodings.
