+++ paper.md
@@ -53,21 +53,21 @@ Specifically, the library sets out to fulfill the following goals:
 
 # Statement of Need
 
-High quality open-science tools encourage researchers to follow effective software development practices by improving code quality, scientific rigor, and ease of replication or extension.
-Empirical's debugging suite protects against common C++ programming pitfalls, such as iterator invalidation, memory leakage, and out-of-bounds indexing.
-Throughout, library design obviates trade-offs between performance and safety; compile-time switches toggle safety checks for undefined or incorrect behavior.
+High quality open-science tools improve code quality, scientific rigor, and ease of replication or extension for scientific software.
+Empirical's debugging suite combats C++ programming pitfalls, such as iterator invalidation, memory leakage, and out-of-bounds indexing.
+Throughout, library design achieves both performance and safety through compile-time toggling of checks for undefined or incorrect behavior.
 
-In practice, many scientific software applications are difficult to obtain, install, or use, and produce data in proprietary formats.
-Modern web-based interfaces give computational research the potential to better embody open science objectives by making each step of the scientific process more observable [@woelfle2011open].
-Empirical leverages recent developments in web technology to compile C++ source code to browser-based interactive interfaces.
+Unfortunately, in practice, scientific software is often difficult to obtain, install, or use.
+Modern web-based interfaces give computational research the potential to better embody open science objectives by empowering easier and more complete access [@woelfle2011open].
+Empirical leverages modern web technology to provide browser-based interactive interfaces for C++ source code.
 
 # Empirical Features
 
 ## Better Code for Scientific Software
 
-Empirical components are subjected to structured code review, unit testing (with coverage tracking), and other modern software development practices detailed [in our documentation](https://empirical.readthedocs.io/en/latest/dev/empirical-development-practices.html).
-Sustained effort invested into optimization of the library's utilities enables developers to produce safe and efficient software at far less effort, especially for new developers.
-Additionally we provide a [template project](https://github.com/devosoft/cookiecutter-empirical-project) that streamlines laying out crosscompilation boilerplate.
+Empirical components are subjected to structured code review, unit testing with coverage tracking, and other best practices detailed [in our documentation](https://empirical.readthedocs.io/en/latest/dev/empirical-development-practices.html).
+Effort invested into optimization of the library's utilities enables developer-users to more easily produce safe and efficient software, especially for new developers.
+We provide a [template project](https://github.com/devosoft/cookiecutter-empirical-project) that streamlines laying out crosscompilation boilerplate.
 
 As an example of Emprical's utility, the library provides a configuration framework that includes utilities to
 
@@ -76,87 +76,76 @@ As an example of Emprical's utility, the library provides a configuration framew
 * perform on-the-fly configuration adjustments, and
 * support independent configuration subsystems.
 
-High-quality software cannot succeed without an inclusive community of contributors.
-As detailed [in our documentation](https://empirical.readthedocs.io/en/latest/dev/empirical-development-practices.html), our development practices emphasize the importance of diverse backgrounds and perspectives in the community.
+High-quality software needs a robust, inclusive, and diverse community of users and contributors.
+Our [development practices](https://empirical.readthedocs.io/en/latest/dev/empirical-development-practices.html) reflect this priority.
 
 ## Realizing the Promise of Emscripten-based Web UIs
 
-Educational or outreach versions of scientific software can promote classroom learning and citizen science.
[…]
