### 2024-02-09 33f89a7
Add signalgp-lite citation
