+++ paper.bib
@@ -279,3 +279,20 @@ keywords = {interface design, Interactive machine learning}
   doi = {10.1038/nchem.1149},
   publisher={Nature Publishing Group UK London}
 }
+@misc{moreno2021signalgp,
+  doi = {10.48550/ARXIV.2108.00382},
+
+  url = {https://arxiv.org/abs/2108.00382},
+
+  author = {Moreno, Matthew Andres and Papa, Santiago Rodriguez and Lalejini, Alexander and Ofria, Charles},
+
+  keywords = {Neural and Evolutionary Computing (cs.NE), FOS: Computer and information sciences, FOS: Computer and information sciences},
+
+  title = {SignalGP-Lite: Event Driven Genetic Programming Library for Large-Scale Artificial Life Applications},
+
+  publisher = {arXiv},
+
+  year = {2021},
+
+  copyright = {arXiv.org perpetual, non-exclusive license}
+}
