### 2024-02-09 0b710f8
Strip trailing whitespace
