+++ paper.bib
@@ -186,13 +186,13 @@ year = {2013},
 month={11},
 doi = {10.1177/0037549712462620},
 
-URL = { 
+URL = {
         https://doi.org/10.1177/0037549712462620
-    
+
 },
-eprint = { 
+eprint = {
         https://doi.org/10.1177/0037549712462620
-    
+
 }
 ,
     abstract = { In the last decade, agent-based modeling and simulation (ABMS) has been applied to a variety of domains, demonstrating the potential of this technique to advance science, engineering, and policy analysis. However, realizing the full potential of ABMS to find breakthrough research results requires far greater computing capability than is available through current ABMS tools. The Repast for High Performance Computing (Repast HPC) project addresses this need by developing a useful and useable next-generation ABMS system explicitly focusing on larger-scale distributed computing platforms. Repast HPC is intended to smooth the path from small-scale simulations to large-scale distributed simulations through the use of a Logo-like system. This article’s contribution is its detailed presentation of the implementation of Repast HPC as a useful and usable framework, a complete ABMS platform developed explicitly for larger-scale distributed computing systems that leverages modern C++ techniques and the ReLogo language. }
@@ -244,7 +244,7 @@ series = {GECCO '18}
   url          = {https://doi.org/10.5281/zenodo.4118608}
 }
 
- @misc{world_academy_of_science_engineering_and_technology_2019, title={ Open Science Philosophy}, url={https://waset.org/files/Open-Science-Philosophy.pdf}, journal={World Academy of Science, Engineering and Technology}, year={2019}} 
+ @misc{world_academy_of_science_engineering_and_technology_2019, title={ Open Science Philosophy}, url={https://waset.org/files/Open-Science-Philosophy.pdf}, journal={World Academy of Science, Engineering and Technology}, year={2019}}
 
 @article{10.1145/3185517,
 author = {Dudley, John J. and Kristensson, Per Ola},
