### 2024-02-09 b21c59d
Try upgrading git to workaround unclean macos clone on CI
Here is the offending CI log
```
non-GNU awk detected
using gawk as awk
cat version info unavailable
using gcat as cat
find version info unavailable
using gfind as find
non-GNU grep detected
using ggrep as grep
sed version info unavailable
using gsed as sed
seq version info unavailable
using gseq as seq
non-GNU sort detected
using gsort as sort
xargs version info unavailable
using gxargs as xargs
Running tidyness enforcement tests...
See https://empirical.readthedocs.io/en/latest/dev/guide-to-testing.html#tidyness-enforcement for info o
[…]
