### 2024-02-09 4b2d73c
Try doing macos CI checkout after upgrading git
