+++ .github/workflows/CI.yml
@@ -23,13 +23,6 @@ jobs:
     name: Test Tidy Enforcement on macOS
     runs-on: macos-11
     steps:
-      - uses: actions/checkout@v4
-        with:
-          # sometimes checkout action doesn't result in clean directory?
-          fetch-depth: 0
-          ref: ${{ github.event.pull_request.head.sha || github.ref }}
-          submodules: 'recursive'
-      - run: git status --ignored
       - run: brew install coreutils findutils gnu-sed gawk grep rename sphinx-doc
       # adapted from https://stackoverflow.com/a/42878119
       - run: brew link sphinx-doc --force
@@ -43,6 +36,13 @@ jobs:
       # see https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#adding-a-system-path
       - run: echo "/usr/local/bin:$PATH" > $GITHUB_PATH
       - run: echo $PATH
+      - uses: actions/checkout@v4
+        with:
+          # sometimes checkout action doesn't result in clean directory?
+          fetch-depth: 0
+          ref: ${{ github.event.pull_request.head.sha || github.ref }}
+          submodules: 'recursive'
+      - run: git status --ignored
       - run: ./ci/test_tidy.sh
   test:
     name: Tests
