### 2024-02-09 7333687
Bypass tidy git status check in CI
because macos is having weird, apparently intractable, dirty clones
