### 2024-02-09 2b39a9d
Log git status bypass in CI
