### 2024-02-09 49e684f
Revert fix attempt for macos dirty CI checkout
