### 2024-02-09 #6151 @mmore500
Quick update — we have completed our work to bring the manuscript down to size. We should be coming in at or very near the 1,000 word mark now! 🎉

I am circulating the new draft among the co-authors and expect to confirm their approval (or final tweaks) by Tuesday. I will follow up here to confirm once that is finalized. Thank you again for working us on this process!
