### 2024-02-10 7b047bd
Changed Text::GetText() to Text::AsString()
