### 2024-02-11 9b80c9e
Started Overhaul of TextEncoding to use new rule system.
