+++ include/emp/text/TextEncoding.hpp
@@ -1,12 +1,16 @@
 /**
  *  @note This file is part of Empirical, https://github.com/devosoft/Empirical
  *  @copyright Copyright (C) Michigan State University, MIT Software license; see doc/LICENSE.md
- *  @date 2023.
+ *  @date 2023-2024.
  *
  *  @file TextEncoding.hpp
  *  @brief Plugs into emp::Text with the specification of a given encoding.
  *  @note Status: ALPHA
  *
+ *  DEVELOPER NOTES:
+ *  - Right now a close tag of "\n" is special; it will end ALL styles it is associated with.
+ *    Perhaps we should have configurable rules for how common end tags work?  For example,
+ *    there could be "no duplicates allowed", "close most recent", or "close all"
  */
 
 #ifndef EMP_TEXT_TEXTENCODING_HPP_INCLUDE
@@ -22,75 +26,294 @@
 #include "../datastructs/map_utils.hpp"
 
 #include "Text.hpp"
+#include "Text_Rule.hpp"
 
 namespace emp {
 
   class TextEncoding : public emp::TextEncoding_Interface {
   protected:
 
-    // Tags have three possible things they can do:
-    //  1: Start a new style
-    //  2: End an ongoing style, or
-    //  3: Be replaced with text (which may have a style associated)
-    struct Tag {
-      String name;                  // Unique name for this tag; default encoding symbol.
-      String pattern;               // Regular expression to identify tag.
-      size_t id = emp::MAX_SIZE_T;  // Unique ID for this tag (index in tag_map)
-      int token_id = -1;            // ID of this tag in the lexer.
-
-      // -- Starting a Style --
-      // Tags can indicate that they start a style.  Since the style may have arguments, the
-      // lexeme that identified the tag will be passed in and can be processed.
-      // Note that style zero (the default) is "no_style" and used to indicate no start style.
-      size_t start_style_id = 0;                    // No style started by default.
-      std::function<String(String)> get_style_args; // Function to convert lexeme to style arguments.
-
-      // -- Ending a Style --
-      // Tags may be used to end one or more styles.  The set of styles that can be ended is
-      // tracked, and whether multiple styles can be ended at once.  For example, a newline might
-      // end a whole set of styles in an encoding.  A "</b>" in HTML, on the other hand might only
-      // end the bold style.  A "}" in latex would end the most recent style started.
-      std::set<size_t> end_style_ids;  // Set of styles this tag can end.
-      bool multi_end = false;          // Can this tag end more than one style at a time?
-
-      // -- Text Replacement --
-      // Some tags might be used to produce a special character in the output text. For example,
-      // "&lt;" in HTML would be replaced by '<'.  A replacement character can also have a
-      // style with it.  For example, "&nbsp;" would be replaced by ' ' with the style "no_break".
-      char replace_char = '\0';     // Character to place in TEXT (Default is none).
-      size_t replace_style_id = 0;  // Style ID to use if there is a replacement char.
-      String out_encoding="";       // Use to replace char (and correct style) in encoding.
-    };
-
     struct Style {
-      String name;                      // Unique name for this style.
-      size_t id = emp::MAX_SIZE_T;      // Unique ID number (vector position) for this style.
-      std::set<size_t> open_tag_ids;    // Which tags start this style?
-      std::set<size_t> close_tag_ids;   // Which tags close this style?
-      std::set<size_t> replace_tag_ids; // Which replacements use this style?
+      String name;                 // Unique name for this style.
+      size_t id = emp::MAX_SIZE_T; // Unique ID number (vector position) for this style.
[…]
