### 2024-02-12 85f576e
Added a PrintDebug() member function to Text_Rule
