### 2024-02-13 c93a442
Updated HTMLText tests and added tests for multiple &nbsp; in a row.
