+++ tests/text/HTMLText.cpp
@@ -16,50 +16,50 @@
 TEST_CASE("Testing HTMLText", "[text]") {
   emp::HTMLText text("Test Text");
   REQUIRE(text.GetSize() == 9);
-  REQUIRE(text.GetText() == "Test Text");
+  REQUIRE(text.AString() == "Test Text");
   REQUIRE(text.Encode() == "Test Text");
 
   // Try adding style.
   text.Bold(5,9);
   REQUIRE(text.GetSize() == 9);
-  REQUIRE(text.GetText() == "Test Text");
+  REQUIRE(text.AString() == "Test Text");
   REQUIRE(text.Encode() == "Test <b>Text</b>");
 
   // Try appending.
   text << " and more Text.";
   REQUIRE(text.GetSize() == 24);
-  REQUIRE(text.GetText() == "Test Text and more Text.");
+  REQUIRE(text.AString() == "Test Text and more Text.");
   REQUIRE(text.Encode() == "Test <b>Text</b> and more Text.");
 
   // Try changing letters.
   text[21] = 's';
   text[23] = 's';
   REQUIRE(text.GetSize() == 24);
-  REQUIRE(text.GetText() == "Test Text and more Tests");
+  REQUIRE(text.AString() == "Test Text and more Tests");
   REQUIRE(text.Encode() == "Test <b>Text</b> and more Tests");
 
   // Try making a change that involves style.
   text[19] = text[5];
   REQUIRE(text.GetSize() == 24);
-  REQUIRE(text.GetText() == "Test Text and more Tests");
+  REQUIRE(text.AString() == "Test Text and more Tests");
   REQUIRE(text.Encode() == "Test <b>Text</b> and more <b>T</b>ests");
 
   text.Italic(14,18);
   REQUIRE(text.GetSize() == 24);
-  REQUIRE(text.GetText() == "Test Text and more Tests");
+  REQUIRE(text.AString() == "Test Text and more Tests");
   REQUIRE(text.Encode() == "Test <b>Text</b> and <i>more</i> <b>T</b>ests");
 
   // Try erasing the text.
   text.Resize(0);
   REQUIRE(text.GetSize() == 0);
-  REQUIRE(text.GetText() == "");
+  REQUIRE(text.AString() == "");
   REQUIRE(text.Encode() == "");
   REQUIRE(text.GetStyles().size() == 0);
 
   text << "This is <b><i>Pre-</i>formatted</b> text.";
   // text.PrintDebug();
   // REQUIRE(text.GetSize() == 27);
-  REQUIRE(text.GetText() == "This is Pre-formatted text.");
+  REQUIRE(text.AString() == "This is Pre-formatted text.");
   REQUIRE(text.Encode() == "This is <i><b>Pre-</i>formatted</b> text.");
 
   emp::Text plain_text("The <b> and </b> here should not be converted.");
@@ -77,12 +77,17 @@ TEST_CASE("Testing HTMLText", "[text]") {
 
   // But if we put it into an HTML object, the style should be reflected.
   emp::HTMLText html_text(plain_text);
-  REQUIRE(html_text.GetText() == "The <b> and </b> here should not be converted.");
+  REQUIRE(html_text.AString() == "The <b> and </b> here should not be converted.");
   REQUIRE(html_text.Encode() == "The &lt;b&gt; and &lt;/b&gt; here should <b>not</b> be converted.");
 
   emp::HTMLText text2 = "Now let's try something with a non-breaking&nbsp;space.";
-  REQUIRE(text2.GetText() == "Now let's try something with a non-breaking space.");
+  REQUIRE(text2.AString() == "Now let's try something with a non-breaking space.");
[…]
