### 2024-02-16 9b84273
Updated TextEncoding to handle non-stylized output; added more debugging.
