+++ include/emp/text/TextEncoding.hpp
@@ -39,6 +39,11 @@ namespace emp {
       std::set<size_t> rule_ids;   // Which rules manage this style?
 
       Style(String name, size_t id) : name(name), id(id) { }
+
+      void PrintDebug(std::ostream & os = std::cout) const {
+        os << "Style " << id << " (" << name << "): Rule(s) "
+           << emp::MakeEnglishList(rule_ids) << std::endl;
+      }
     };
 
     // Track information about _active_ styles.
@@ -79,8 +84,8 @@ namespace emp {
     std::map<String, size_t> name_to_style_id; // Link style names to their ID
     std::map<String, TagInfo> tag_map;         // Map of tags to the rules they may affect
 
-    // Map of special (Emphatic) strings to the rules that generate them
-    std::map<String, size_t> special_strings;
+    // Chart of characters to the rule by which they must be encoded (with no associated special style)
+    std::array<size_t, 128> special_chars;
 
     emp::Lexer lexer;                           // Lexer to process encoding.
     int text_token = -1;                        // Token to represent any non-tag text.
@@ -253,7 +258,10 @@ namespace emp {
 
       // If this rule generates internal text without a style, make it reversible.
       if (rule.internal_text.size() && !rule.base_style.size()) {
-        special_strings[rule.internal_text] = rule.id;
+        emp::notify::TestError(rule.internal_text.size() != 1,
+          "Rules that convert text with no emphatic style must be limited to one character.\n",
+          "Text = '", rule.internal_text, "'.");
+        special_chars[rule.internal_text[0]] = rule.id;
       }
     }
 
@@ -313,61 +321,16 @@ namespace emp {
       RegisterRule(rule);
     }
 
-/*
-    size_t GetTagID(const String & tag_name, const String & tag_pattern) {
-      // If we don't have this tag yet, build it.
-      if (!emp::Has(pattern_to_tag_id, tag_pattern)) {
-        Tag new_tag;
-        new_tag.name = tag_name;
-        new_tag.pattern = tag_pattern;
-        new_tag.id = tag_set.size();
-        tag_set.push_back(new_tag);
-        pattern_to_tag_id[tag_pattern] = new_tag.id;
-        return tag_set.back().id;
-      }
-      return pattern_to_tag_id[tag_pattern];
-    }
-
-    // If no pattern is provided in get, use the tag name to generate it.
-    size_t GetTagID(const String & tag_name) {
-      return GetTagID(tag_name, emp::MakeLiteral(tag_name));
-    }
-
-
-
-
-    // Get the style name with any associated args that should be used in Text.
-    String GetStyleDesc(const Style & style, const Tag & start_tag, const String & lexeme) {
-      if (!start_tag.get_style_args) return style.name;
-      return emp::MakeString(style.name, ':', start_tag.get_style_args(lexeme));
-    }
-
-
-
-
-
-
-    void _EncodeChar(const Text & text, String & out_string, const size_t char_pos) const {
-      const char c = text.GetChar(char_pos);
-      // If there is a tag associated with this character AND the associated type (if any) use tag.
-      if (char_tags[c]) {
-        const Tag & tag = tag_set[char_tags[c]];
-        const Style & style = style_set[tag.replace_style_id];
-        if (!tag.replace_style_id || text.HasStyle(style.name, char_pos)) {
-          out_string += tag.out_encoding; // Char match AND style match (or no style), so print tag
-        } else out_string += c;   // Wrong style for tag; just print character.
-      } else {
-        out_string += c;
+  private: // Pure helper functions...
[…]
