+++ paper.md
@@ -38,7 +38,7 @@ affiliations:
    index: 7
  - name: Computer Science, Grand Valley State University, USA
    index: 8
-date: 30 July 2021
+date: 13 February 2024
 bibliography: paper.bib
 breaks: false
 ---
