### 2024-02-17 9046e47
Added Bits unit tests for FindZero and GetRanges
