+++ tests/bits/Bits.cpp
@@ -1,7 +1,7 @@
 /**
  *  @note This file is part of Empirical, https://github.com/devosoft/Empirical
  *  @copyright Copyright (C) Michigan State University, MIT Software license; see doc/LICENSE.md
- *  @date 2022.
+ *  @date 2022-24.
  *
  *  @file Bits.cpp
  */
@@ -658,6 +658,20 @@ TEST_CASE("7: Test functions that analyze and manipulate ones", "[bits]") {
   CHECK(bv.FindOne(13) == 13);
   CHECK(bv.FindOne(14) == 14);
   CHECK(bv.FindOne(15) == -1);
+
+  // Also check that we find the zeroes.
+  CHECK(bv.FindZero() == 0);
+  CHECK(bv.FindZero(1) == 1);
+  CHECK(bv.FindZero(2) == 2);
+  CHECK(bv.FindZero(3) == 4);
+  CHECK(bv.FindZero(4) == 4);
+  CHECK(bv.FindZero(5) == 5);
+  CHECK(bv.FindZero(7) == 8);
+  CHECK(bv.FindZero(11) == 11);
+  CHECK(bv.FindZero(12) == 15);
+  CHECK(bv.FindZero(14) == 15);
+  CHECK(bv.FindZero(15) == 15);
+  CHECK(bv.FindZero(16) == -1);
 
   // Get all of the ones at once and make sure they're there.
   emp::vector<size_t> ones = bv.GetOnes();
@@ -1540,14 +1554,14 @@ struct TestBitArrayConstruct<> {
 TEST_CASE("17: Test BitArray Constructors", "[bits]"){
   // Create a size 50 bit vector, default to all zeros.
   emp::BitArray<50> ba1;
-  REQUIRE( ba1.GetSize() == 50 );
-  REQUIRE( ba1.CountOnes() == 0 );
-  REQUIRE( (~ba1).CountOnes() == 50 );
+  CHECK( ba1.GetSize() == 50 );
+  CHECK( ba1.CountOnes() == 0 );
+  CHECK( (~ba1).CountOnes() == 50 );
 
   // Create a size 1000 BitArray, default to all ones.
   emp::BitArray<1000> ba2(true);
-  REQUIRE( ba2.GetSize() == 1000 );
-  REQUIRE( ba2.CountOnes() == 1000 );
+  CHECK( ba2.GetSize() == 1000 );
+  CHECK( ba2.CountOnes() == 1000 );
 
   // Try a range of BitArray sizes, from 1 to 200.
   TestBitArrayConstruct<1,2,7,8,9,15,16,17,31,32,33,63,64,65,127,128,129,191,192,193,200>::Run();
@@ -1555,45 +1569,45 @@ TEST_CASE("17: Test BitArray Constructors", "[bits]"){
   // Build a relatively large BitArray.
   emp::BitArray<1000000> ba4;
   for (size_t i = 0; i < ba4.GetSize(); i += 100) ba4[i].Toggle();
-  REQUIRE( ba4.CountOnes() == 10000 );
+  CHECK( ba4.CountOnes() == 10000 );
 
   // Try out the copy constructor.
   emp::BitArray<1000000> ba5(ba4);
-  REQUIRE( ba5.GetSize() == 1000000 );
-  REQUIRE( ba5.CountOnes() == 10000 );
+  CHECK( ba5.GetSize() == 1000000 );
+  CHECK( ba5.CountOnes() == 10000 );
 
   // Construct from std::bitset.
   std::bitset<6> bit_set;
   bit_set[1] = 1;   bit_set[2] = 1;   bit_set[4] = 1;
   emp::BitArray<6> ba7(bit_set);
[…]
