### 2024-02-18 6e348ab
Cleaned up Bits::GetRanges and fixed bug in Bits::FindZero
