### 2024-02-20 5001c18
Made String::PrintEnglishList work with a wider range of containers.
