### 2024-02-22 c0e0931
Print more info with RegEx assert
