### 2024-02-22 f42458c
Update HTMLText tests to use CHECK, not REQUIRE
