+++ tests/text/HTMLText.cpp
@@ -15,78 +15,79 @@
 
 TEST_CASE("Testing HTMLText", "[text]") {
   emp::HTMLText text("Test Text");
-  REQUIRE(text.GetSize() == 9);
-  REQUIRE(text.AString() == "Test Text");
-  REQUIRE(text.Encode() == "Test Text");
+  CHECK(text.GetSize() == 9);
+  CHECK(text.AsString() == "Test Text");
+  CHECK(text.Encode() == "Test Text");
 
   // Try adding style.
   text.Bold(5,9);
-  REQUIRE(text.GetSize() == 9);
-  REQUIRE(text.AString() == "Test Text");
-  REQUIRE(text.Encode() == "Test <b>Text</b>");
+  CHECK(text.GetSize() == 9);
+  CHECK(text.AsString() == "Test Text");
+  CHECK(text.Encode() == "Test <b>Text</b>");
 
   // Try appending.
   text << " and more Text.";
-  REQUIRE(text.GetSize() == 24);
-  REQUIRE(text.AString() == "Test Text and more Text.");
-  REQUIRE(text.Encode() == "Test <b>Text</b> and more Text.");
+  CHECK(text.GetSize() == 24);
+  CHECK(text.AsString() == "Test Text and more Text.");
+  CHECK(text.Encode() == "Test <b>Text</b> and more Text.");
 
   // Try changing letters.
   text[21] = 's';
   text[23] = 's';
-  REQUIRE(text.GetSize() == 24);
-  REQUIRE(text.AString() == "Test Text and more Tests");
-  REQUIRE(text.Encode() == "Test <b>Text</b> and more Tests");
+  CHECK(text.GetSize() == 24);
+  CHECK(text.AsString() == "Test Text and more Tests");
+  CHECK(text.Encode() == "Test <b>Text</b> and more Tests");
 
   // Try making a change that involves style.
   text[19] = text[5];
-  REQUIRE(text.GetSize() == 24);
-  REQUIRE(text.AString() == "Test Text and more Tests");
-  REQUIRE(text.Encode() == "Test <b>Text</b> and more <b>T</b>ests");
+  CHECK(text.GetSize() == 24);
+  CHECK(text.AsString() == "Test Text and more Tests");
+  CHECK(text.Encode() == "Test <b>Text</b> and more <b>T</b>ests");
 
   text.Italic(14,18);
-  REQUIRE(text.GetSize() == 24);
-  REQUIRE(text.AString() == "Test Text and more Tests");
-  REQUIRE(text.Encode() == "Test <b>Text</b> and <i>more</i> <b>T</b>ests");
+  CHECK(text.GetSize() == 24);
+  CHECK(text.AsString() == "Test Text and more Tests");
+  CHECK(text.Encode() == "Test <b>Text</b> and <i>more</i> <b>T</b>ests");
 
   // Try erasing the text.
   text.Resize(0);
-  REQUIRE(text.GetSize() == 0);
-  REQUIRE(text.AString() == "");
-  REQUIRE(text.Encode() == "");
-  REQUIRE(text.GetStyles().size() == 0);
+  CHECK(text.GetSize() == 0);
+  CHECK(text.AsString() == "");
+  CHECK(text.Encode() == "");
+  CHECK(text.GetStyles().size() == 0);
 
   text << "This is <b><i>Pre-</i>formatted</b> text.";
   // text.PrintDebug();
-  // REQUIRE(text.GetSize() == 27);
-  REQUIRE(text.AString() == "This is Pre-formatted text.");
-  REQUIRE(text.Encode() == "This is <i><b>Pre-</i>formatted</b> text.");
+  // CHECK(text.GetSize() == 27);
[…]
