### 2024-02-22 129c5e9
Update Text_Rule to be a proper class; change internal text to just a char.
