### 2024-02-22 9b32ccd
Fixed missing returns in Text functions.
