### 2024-02-22 d9645e0
Lots of cleanup and bug fixes in TextEncoding
