+++ include/emp/text/TextEncoding.hpp
@@ -33,12 +33,28 @@ namespace emp {
   class TextEncoding : public emp::TextEncoding_Interface {
   protected:
 
-    struct Style {
-      String name;                 // Unique name for this style.
+    class Style {
+    private:
+      String name = "No Name";     // Unique name for this style.
       size_t id = emp::MAX_SIZE_T; // Unique ID number (vector position) for this style.
       std::set<size_t> rule_ids;   // Which rules manage this style?
 
+    public:
+      Style() { }
       Style(String name, size_t id) : name(name), id(id) { }
+
+      const String & GetName() const { return name; }
+      size_t GetID() const { return id; }
+      const std::set<size_t> & GetRules() const { return rule_ids; }
+
+      void SetID(size_t new_id) {
+        emp_assert(new_id < 1000000000, new_id, "Setting a style to have an invalid ID");
+        id = new_id;
+      }
+      void AddRuleID(size_t rule_id) {
+        emp_assert(rule_id < 1000000000, rule_id, "Adding an invalid rule ID to a style");
+        rule_ids.insert(rule_id);
+      }
 
       void PrintDebug(std::ostream & os = std::cout) const {
         os << "Style " << id << " (" << name << "): Rule(s) "
@@ -48,10 +64,11 @@ namespace emp {
 
     // Track information about _active_ styles.
     struct StyleEntry {
-      size_t style_id;
+      size_t style_id=0;
       String full_info;  // Style name and arguments, as should be stored.
       size_t rule_id=0;  // Which rule created this entry? (determines legal end)
 
+      StyleEntry() { }
       StyleEntry(size_t style_id, String full_info, size_t rule_id)
         : style_id(style_id), full_info(full_info), rule_id(rule_id) { }
       StyleEntry(const StyleEntry &) = default;
@@ -71,11 +88,19 @@ namespace emp {
       bool IsMultiClose() const { return multi_close; }
       size_t GetOpenID() const { return open_id; }
       std::set<size_t> GetCloseIDs() const { return close_ids; }
+
       void AddOpen(size_t rule_id) {
         emp_assert(open_id == 0, "Rule trying to use open tag already in use.", rule_id, open_id);
         open_id = rule_id;
       }
       void AddClose(size_t rule_id) { close_ids.insert(rule_id); }
+      void SetMultiClose(bool in_mc=true) { multi_close = in_mc; }
+
+      void PrintDebug(std::ostream & os=std::cout) const {
+        os << "open_id=" << open_id << " ; close_ids=";
+        for (size_t id : close_ids) os << id << " ";
+        os << "; multiclose=" << multi_close << std::endl;
+      }
     };
 
     emp::vector<Text_Rule> rule_set;
@@ -107,7 +132,7 @@ namespace emp {
         emp::notify::TestError(create_ok == false, "Requested style '", style_name,
                                "' does not exist; creation is disallowed.");
         style_set.emplace_back(style_name, style_set.size());
-        name_to_style_id[style_name] = style_set.back().id;
+        name_to_style_id[style_name] = style_set.back().GetID();
         return style_set.back();
       }
       const size_t style_id = name_to_style_id[style_name];
@@ -130,16 +155,17 @@ namespace emp {
     // Get a rule associated with a style that can be applied at the specific text position.
     const Text_Rule & GetRule(const Style & style, const Text & text, size_t pos) const {
       // Scan through possible rules taking the first that applies.
-      for (size_t rule_id : style.rule_ids) {
+      for (size_t rule_id : style.GetRules()) {
         const Text_Rule & rule = rule_set[rule_id];
-        // If this rule has associated internal text, make sure it's there.
[…]
