### 2024-02-23 bbfa6e1
Separated symbols from styles in Text_Rule; cleanup on comments.
