### 2024-02-24 77f5030
Pulled out a separate symbol_map in Text.hpp
