### 2024-02-24 1a9afd9
Major overhaul of TextEncoding to separate styles and symbols.
