+++ include/emp/text/TextEncoding.hpp
@@ -103,14 +103,14 @@ namespace emp {
       }
     };
 
-    emp::vector<Text_Rule> rule_set;
-    emp::vector<Style> style_set;
+    emp::vector<Text_Rule> rule_set;           // Set of all rules in this encoding
+    emp::vector<Style> style_set;              // Set of all styles known by this encoding
 
     std::map<String, size_t> name_to_style_id; // Link style names to their ID
     std::map<String, TagInfo> tag_map;         // Map of tags to the rules they may affect
 
-    // Chart of characters to the rule by which they must be encoded (with no associated special style)
-    std::array<size_t, 128> special_chars;
+    std::array<size_t, 128> special_chars;         // Chart of special ASCII chars to encoding rule
+    std::unordered_map<String, size_t> symbol_map; // Map of internal names to encoding rule
 
     emp::Lexer lexer;                           // Lexer to process encoding.
     int text_token = -1;                        // Token to represent any non-tag text.
@@ -152,13 +152,20 @@ namespace emp {
       return style_set[style_id];
     }
 
+    // Get the symbol tag associated with a given internal name.
+    const Text_Rule & GetSymbolRule(const String & internal_name) const {
+      auto it = symbol_map.find(internal_name);
+      if (it == symbol_map.end()) return rule_set[0];  // Symbol not found!
+      return rule_set[it->second];
+    }
+
     // Get a rule associated with a style that can be applied at the specific text position.
     const Text_Rule & GetRule(const Style & style, const Text & text, size_t pos) const {
       // Scan through possible rules taking the first that applies.
       for (size_t rule_id : style.GetRules()) {
         const Text_Rule & rule = rule_set[rule_id];
         // If this rule uses an internal character, make sure it's there.
-        if (rule.GeneratesText() && text.AsString()[pos] != rule.GetInternalChar()) {
+        if (rule.GetInternalChar() && text.AsString()[pos] != rule.GetInternalChar()) {
           // std::cout << "Failed on style '" << style.GetName() << "' because wrong internal text; expected '" << rule.GetInternalChar() << "' but found '" << text.AsString()[pos] << "'." << std::endl;
           continue; // Does not have correct internal char.
         }
@@ -194,6 +201,15 @@ namespace emp {
       }
     }
 
+    // Append a string that has already been otherwise processed.
+    void Append_Symbol(Text & text, const String & symbol_type, char placeholder) const {
+      text.Append_Symbol(symbol_type, placeholder);
+      // Apply all active styles.
+      for (const auto & style : active_styles) {
+        text.SetStyle(style.full_info, text.size()-1);
+      }
+    }
+
     // Helper function to append text when we know we've hit a tag.
     void Append_Tag(Text & text, int token_id, std::istream & input_stream) {
       emp_assert(Has(token_map, token_id), "Unknown token id: ", token_id);
@@ -216,7 +232,7 @@ namespace emp {
 
       if (used) return;  // If the tag was used to close, don't ALSO open.
 
-      // OPEN a new style if indicated by this tag.
+      // OPEN a new style or ADD a new symbol if indicated by this tag.
       if (info.HasOpen()) {
         const Text_Rule & rule = rule_set[info.GetOpenID()];
 
@@ -239,16 +255,13 @@ namespace emp {
         if (style_desc.size()) {
           size_t style_id = name_to_style_id[rule.GetBaseStyle()];
           active_styles.emplace_back(style_id, style_desc, rule.GetID());
+
+          emp_assert(rule.UsesText(), "Rules that start a style MUST eventually close it.");
         }
 
[…]
