### 2024-02-24 b1415b8
Overhaul on EmphaticEncoding to use new setup.
