+++ include/emp/text/EmphaticEncoding.hpp
@@ -1,7 +1,7 @@
 /**
  *  @note This file is part of Empirical, https://github.com/devosoft/Empirical
  *  @copyright Copyright (C) Michigan State University, MIT Software license; see doc/LICENSE.md
- *  @date 2022-2023.
+ *  @date 2022-2024.
  *
  *  @file EmphaticEncoding.hpp
  *  @brief Plugs into emp::Text, setting up inputs and output to be Emphatic encoded.
@@ -25,45 +25,107 @@ namespace emp {
   private:
 
     void SetupTags() {
-      SetupStyleTags("bold",        "`*", "`*");
-      SetupStyleTags("code",        "``", "``");
-      SetupStyleTags("italic",      "`/", "`/");
-      SetupStyleTags("strike",      "`~", "`~");
-      SetupStyleTags("subscript",   "`.", "`.");
-      SetupStyleTags("superscript", "`^", "`^");
-      SetupStyleTags("underline",   "`_", "`_");
-      SetupStyleTags("center",      "`=", "`=");
-      SetupStyleTags("justify",     "`|", "`|");
-      SetupStyleTags("ljustify",    "`<", "`<");
-      SetupStyleTags("rjustify",    "`>", "`>");
-      SetupStyleTags("header:1",    "`#", "`#");
-      SetupStyleTags("header:2",    "`##", "`##");
-      SetupStyleTags("header:3",    "`###", "`###");
-      SetupStyleTags("header:4",    "`####", "`####");
-      SetupStyleTags("header:5",    "`#####", "`#####");
-      SetupStyleTags("header:6",    "`######", "`######");
+      // === Styles ===
+      AddStyleToggle("bold",        "`*");
+      AddStyleToggle("code",        "`");  // Note: Longer tags will match first.
+      AddStyleToggle("code",        "``");
+      AddStyleToggle("code_block",  "```");
+      AddStyleToggle("italic",      "`/");
+      AddStyleToggle("strike",      "`~");
+      AddStyleToggle("subscript",   "`.");
+      AddStyleToggle("superscript", "`^");
+      AddStyleToggle("underline",   "`_");
+      AddStyleToggle("center",      "`=");
+      AddStyleToggle("justify",     "`|");
+      AddStyleToggle("ljustify",    "`<");
+      AddStyleToggle("rjustify",    "`>");
 
-      SetupReplaceTag("\\`",  '`');
-      SetupReplaceTag("\\\\", '\\');
-      SetupReplaceTag("\\ ",  ' ', "no_break");
+      AddStyleLine("blockquote", "`\"");
+      AddStyleLine("comment",    "`%");
+      AddStyleLine("header:1",   "`#");
+      AddStyleLine("header:2",   "`##");
+      AddStyleLine("header:3",   "`###");
+      AddStyleLine("header:4",   "`####");
+      AddStyleLine("header:5",   "`#####");
+      AddStyleLine("no_format",  "`-");
+      AddStyleLine("continue_format", "`:");
 
-      SetupReplaceTag("`-",  ' ', "bullet");
-      SetupReplaceTag("`+",  ' ', "ordered");
-      SetupReplaceTag("`>",  ' ', "indent");
-      SetupReplaceTag("`\"", ' ', "blockquote");
+      AddStyleControl("`@{", '}', "`@", "link");
+      AddStyleControl("`[", ']', "`#",  "color");
 
-      // tag_map['|'] = "continue";  // Keep previous lines setup (for set of full-line tags)
+      // === Insertions ===
 
-      // AVAILABLE TAGS: @&;:',?()]
[…]
