### 2024-02-24 a1e83d1
Overhaul on HTMLEncoding to use new setup.
