### 2024-02-26 bb50224
Refactor of Text_Rule to be more modular.
