+++ include/emp/text/Text_Rule.hpp
@@ -49,85 +49,122 @@ namespace emp {
   /// tag to be generated.  When the link ends, the close take will appear.
   /// If there is no style, the internal character will trigger this rule to apply.
 
+  // == DEVELOPER NOTES ==
+  // Types of rules we need to support:
+  //   Style (e.g., bold, font, link, etc)
+  //     Standard - open and close are individually specified
+  //     Toggles  - open and close are the same
+  //     Lines    - open is specified and close is always a newline
+  //     Control  - open may have a control sequence, but close is still simple
+  //
+  //   Object (e.g., Symbol, emoji, image, page break, etc)
+  //     Simple replacement - encoding tag becomes the object name (perhaps "type:name")
+  //     Parametered replacement - where a control parameter in the tag must be processed.
+
   class Text_Rule {
-  private:
-    size_t id = emp::MAX_SIZE_T;  ///< Unique ID for this rule (index in tag_map)
-    int open_token_id = -1;       ///< ID of open_tag_start in the lexer
-    int close_token_id = -1;      ///< ID of close_tag in the lexer
+  protected:
 
-    // - Target parsing info -
-    String open_tag_start;        ///< Beginning of open tag. (E.g. "<a" or "\bold{")
-    char open_tag_end = '\0';     ///< End of open tag (E.g., '>') or empty if no control used
-    String close_tag;             ///< Tag to indicate close of text, or empty if no text used
+    //  ====================
+    //  === HELPER TYPES ===
+    //  ====================
 
-    // - Styling info -
-    // A base style would be "font"
-    // Style argument could be "arial"
-    // This would generate a full style: "font:arial"
-    String base_style;          ///< What style is this rule associated with? (args may be added)
-    String symbol_type;         ///< Is there a special symbol associated with this rule?
-    char internal_char = '\0';  ///< What character is used to placehold internally in Text?
+    enum RuleType { UNKNOWN_RULE=0, STYLE_RULE, OBJECT_RULE };
 
-    std::function<String(String)> to_style_arg; ///< Convert CONTROL into STYLE arguments
-    std::function<String(String)> to_control;   ///< Convert STYLE into CONTROL
+    /// A helper tag type for tags with a control sequence to specify their operation.
+    struct TagInfo {
+      int token_id = -1;  ///< ID of this tag in the lexer
+      String start;   ///< Unique beginning of the tag (e.g., "<img " or "<a ")
+      char end='\0';  ///< End of tag (E.g., '>')
+
+      std::function<String(String)> to_arg_fun;     ///< Convert CONTROL into internal ARG info
+      std::function<String(String)> to_control_fun; ///< Convert ARG description into CONTROL
+
+      // Can this tag forgo a control sequence?
+      TagInfo() { }
+      TagInfo(String start, char end='\0') : start(start), end(end) { }
+      [[nodiscard]] bool IsUsed() const { return token_id >= 0; }
+      [[nodiscard]] bool IsSimple() const { return end == '\0'; }
+      [[nodiscard]] bool HasFuns() const { return to_arg_fun && to_control_fun; }
+    };
+
+
+    //  =====================
+    //  ===  MEMBER VARS  ===
+    //  =====================
+
+    size_t id = emp::MAX_SIZE_T;       ///< Unique ID for this rule (index in tag_map)
+    RuleType rule_type = UNKNOWN_RULE; ///< What type of rule is this?
+
+    TagInfo main_tag; ///< Identifier tag for an object or open tag for a style
[…]
