### 2024-02-26 5e77303
Update TextEncoding to use new Text_Rule.
