### 2024-02-26 d506afd
Updated some function names in Text_Rule to be clearer.
