### 2024-02-26 cd48b4f
Streamlined and updated names in TextEncoding.
