+++ include/emp/text/TextEncoding.hpp
@@ -111,8 +111,8 @@ namespace emp {
     std::array<size_t, 128> special_chars;     // Chart of special ASCII chars to encoding rule
 
     // Internal objects that are inserted into the text (such as symbols, images, divs, etc)
-    // must have a function that indicates how to convert them to an appropriate tag.
-    std::unordered_map<String, size_t> symbol_map;
+    // must have a rule that indicates how to convert them to an appropriate tag.
+    std::unordered_map<String, size_t> object_map;
 
     emp::Lexer lexer;                           // Lexer to process encoding.
     int text_token = -1;                        // Token to represent any non-tag text.
@@ -156,8 +156,8 @@ namespace emp {
 
     // Get the symbol tag associated with a given internal name.
     const Text_Rule & GetSymbolRule(const String & internal_name) const {
-      auto it = symbol_map.find(internal_name);
-      if (it == symbol_map.end()) {
+      auto it = object_map.find(internal_name);
+      if (it == object_map.end()) {
         notify::Warning("Unknown symbol '", internal_name, "'!  Skipping.");
         return rule_set[0];
       }
@@ -170,9 +170,8 @@ namespace emp {
       for (size_t rule_id : style.GetRules()) {
         const Text_Rule & rule = rule_set[rule_id];
         // If this rule needs an internal char for an object, make sure it's there.
-        if (rule.SetsObject() &&
-            (pos >= text.size() || text.AsString()[pos] != rule.GetInternalChar())) {
-          // std::cout << "Failed on style '" << style.GetName() << "' because wrong internal text; expected '" << rule.GetInternalChar() << "' but found '" << text.AsString()[pos] << "'." << std::endl;
+        if (rule.InsertsObject() &&
+            (pos >= text.size() || text.AsString()[pos] != rule.GetPlaceholder())) {
           continue; // Does not have correct internal char.
         }
         return rule;
@@ -216,60 +215,62 @@ namespace emp {
       }
     }
 
+    /// Helper function to append a tag part 1: Test if it is CLOSING a style.
+    /// @return Was the tag used?
+    bool Append_Tag_Close(Text & text, const TagInfo & info) {
+      if (!info.HasClose()) return false;   // If this token can't CLOSE a style, abort.
+
+      // Scan active styles from most recent to figure out which one to remove.
+      bool used = false;
+      for (size_t id = active_styles.size()-1; id < active_styles.size(); --id) {
+        // If we have found a tag to end, do so.
+        if (info.HasClose(active_styles[id].rule_id)) {
+          active_styles.erase(active_styles.begin()+id);
+          used = true; // Indicate that this tag has been used at least once.
+          if (!info.IsMultiClose()) break;
+        }
+      }
+      return used;
+    }
+
     // Helper function to append text when we know we've hit a tag.
     void Append_Tag(Text & text, int token_id, std::istream & input_stream) {
       emp_assert(Has(token_map, token_id), "Unknown token id: ", token_id);
 
       const TagInfo & info = token_map[token_id];
-      bool used = false;
 
-      // If this token might CLOSE a style, search for options.
-      if (info.HasClose()) {
-        // Scan through active styles from most recent to figure out which one to remove.
-        for (size_t id = active_styles.size()-1; id < active_styles.size(); --id) {
-          // If we have found a tag to end, do so.
-          if (info.HasClose(active_styles[id].rule_id)) {
-            active_styles.erase(active_styles.begin()+id);
-            used = true; // Indicate that this tag has been used at least once.
-            if (!info.IsMultiClose()) break;
-          }
-        }
-      }
[…]
