### 2024-02-26 f454ed5
Cleanup and renaming on Text and Text_Rule.
