### 2024-02-26 81b3316
Refactored TextEncoding into smaller functions.
