+++ include/emp/text/TextEncoding.hpp
@@ -233,48 +233,75 @@ namespace emp {
       return used;
     }
 
-    // Helper function to append text when we know we've hit a tag.
-    void Append_Tag(Text & text, int token_id, std::istream & input_stream) {
-      emp_assert(Has(token_map, token_id), "Unknown token id: ", token_id);
+    /// Helper function to append a tag part 2: Test if OPENING a style or INSERTING an object
+    /// @return Was the tag used?
+    bool Append_Tag_Open(Text & text, const Text_Rule & rule, const String & control) {
+      if (rule.SetsStyle()) {
+        String style_desc = rule.MakeStyle(control);
+        size_t style_id = name_to_style_id[rule.GetBaseStyle()];
+        active_styles.emplace_back(style_id, style_desc, rule.GetID());
+        return true;
+      }
 
-      const TagInfo & info = token_map[token_id];
+      // Do any symbol insertion required by this rule.
+      if (rule.InsertsObject()) {
+        Append_Symbol(text, rule.GetBaseObject(), rule.GetPlaceholder());
+        return true;
+      }
+
+      return false;
+    }
+
+    // Load a CONTROL sequence from an input stream.
+    String Append_LoadControl(std::istream & input_stream, const Text_Rule & rule) {
+      if (!rule.UsesControl()) return "";
+      String control;
+      char c = '\0';
+      while (input_stream) {
+        input_stream >> c;
+        if (c == rule.GetOpenTagEnd()) return control; // Stop when we find the tag end.
+        control += c;
+      }
+      notify::TestError(c != rule.GetOpenTagEnd(),
+                        "Open tag beginning with '", rule.GetOpenTagStart(),
+                        "' did not have end ('", rule.GetOpenTagEnd(), "')");
+      return "";
+    }
+
+    // Helper function to append text when we know we've hit a tag.
+    void Append_Tag(Text & text, emp::Token token, std::istream & input_stream) {
+      emp_assert(Has(token_map, token.id), "Unknown token id: ", token.id);
+
+      const TagInfo & info = token_map[token.id];
 
       if (Append_Tag_Close(text, info)) return;  // Handle close tags.
 
-      emp::notify::TestWarning(!info.HasOpen(), "No options for using tag!");
+      if (!info.HasOpen()) {
+        notify::Warning("No options for using tag '", token.lexeme, "'.  Ignoring.");
+        return;
+      }
 
       // OPEN a new style or INSERT a new object as indicated by this tag.
       const Text_Rule & rule = rule_set[info.GetOpenID()];
 
       // Collect a control sequence if needed by this rule.
-      String control;
-      if (rule.UsesControl()) {
-        char c = '\0';
-        while (input_stream) {
-          input_stream >> c;
-          if (c == rule.GetOpenTagEnd()) break; // Stop when we find the tag end.
-          control += c;
-        }
-        notify::TestError(c != rule.GetOpenTagEnd(),
-                          "Open tag beginning with '", rule.GetOpenTagStart(),
-                          "' did not have end ('", rule.GetOpenTagEnd(), "')");
-      }
+      String control = Append_LoadControl(input_stream, rule);
 
-      // Setup style opened by this rule, if any.
-      if (rule.SetsStyle()) {
-        String style_desc = rule.MakeStyle(control);
-        size_t style_id = name_to_style_id[rule.GetBaseStyle()];
-        active_styles.emplace_back(style_id, style_desc, rule.GetID());
-        emp_assert(rule.UsesText(), "Rules that start a style MUST eventually close it.", style_desc);
-      }
-
-      // Do any symbol insertion required by this rule.
-      if (rule.InsertsObject()) {
[…]
