### 2024-02-26 81c2210
Made many functions in Text nodiscard.
