+++ include/emp/text/Text.hpp
@@ -60,22 +60,23 @@ namespace emp {
 
   // A base class for any special encodings that should work with Text objects.
   struct TextEncoding_Interface {
+    using this_t = TextEncoding_Interface;
     virtual ~TextEncoding_Interface() { }
 
-    virtual String GetName() const = 0;                         // Return name of encoding.
-    virtual void Append(Text &, const String &) = 0;            // Add new text.
-    virtual String Encode(const Text &) const = 0;              // Output formatted text.
-    virtual emp::Ptr<TextEncoding_Interface> Clone() const = 0; // Copy encoding.
+    [[nodiscard]] virtual String GetName() const = 0;              // Return name of encoding.
+    virtual void Append(Text &, const String &) = 0;               // Add new text.
+    [[nodiscard]] virtual String Encode(const Text &) const = 0;   // Output formatted text.
+    [[nodiscard]] virtual emp::Ptr<this_t> Clone() const = 0;      // Copy encoding.
     virtual void PrintDebug(std::ostream &) const = 0;
   };
 
   class TextEncoding_None : public TextEncoding_Interface {
   public:
     TextEncoding_None() { }
-    String GetName() const override { return "text"; }
+    [[nodiscard]] String GetName() const override { return "text"; }
     void Append(Text & text, const String & in) override;
-    String Encode(const Text & text) const override;
-    emp::Ptr<TextEncoding_Interface> Clone() const override;
+    [[nodiscard]] String Encode(const Text & text) const override;
+    [[nodiscard]] emp::Ptr<TextEncoding_Interface> Clone() const override;
     void PrintDebug(std::ostream & os) const override;
   };
 
@@ -150,10 +151,10 @@ namespace emp {
 
 
     // GetSize() returns the number of characters IGNORING all formatting.
-    size_t GetSize() const { return text.size(); }
+    [[nodiscard]] size_t GetSize() const { return text.size(); }
 
     // Return the current text as an unformatted string.
-    const String & AsString() const { return text; }
+    [[nodiscard]] const String & AsString() const { return text; }
 
     // View the current string
     [[nodiscard]] std::string_view ViewString(size_t start=0, size_t out_size=String::npos) const {
@@ -166,17 +167,17 @@ namespace emp {
     }
 
     // Return all of the styles used in this text.
-    const std::unordered_map<String, BitVector> & GetStyleMap() const {
+    [[nodiscard]] const std::unordered_map<String, BitVector> & GetStyleMap() const {
       return style_map;
     }
 
     // Return all of the symbols in this text.
-    const std::map<size_t, String> & GetSymbols() const {
+    [[nodiscard]] const std::map<size_t, String> & GetSymbols() const {
       return symbol_layout;
     }
 
     // Return the current bit pattern for a specified style.
-    const BitVector & GetStyle(const String & style) const {
+    [[nodiscard]] const BitVector & GetStyle(const String & style) const {
       return emp::GetConstRef(style_map, style);
     }
 
@@ -186,13 +187,13 @@ namespace emp {
     /// @brief Test if this Text object is aware of how to use a specified encoding.
     /// @param name Name of the encoding to test for.
     /// @return A true/false indicate if the named encoding is known.
-    bool HasEncoding(const String & name) const {
+    [[nodiscard]] bool HasEncoding(const String & name) const {
       return emp::Has(encodings, name);
     }
 
     /// @brief Get the name of the current encoding being applied.
[…]
