### 2024-02-27 9a853af
Implement, test GetSumDistance member for Systematics
