### 2024-02-28 f7ce103
Started building a GetTextSymbolNames() helper function.
