+++ include/emp/text/Text_utils.hpp
@@ -26,6 +26,365 @@
 
 namespace emp {
 
+  /// A mapping of allowed symbol names to placeholder characters.
+  static const auto & GetTextSymbolNames() {
+    static std::map<String, char> symbol_map;
+    if (symbol_map.size() == 0) {
+      // AddReplaceTag("&iexcl;",   '!', "symbol:iexcl");      // Inverted '!'
+      // AddReplaceTag("&cent;",    'c', "symbol:cent");       // 162 ¢ : Cent
+      // AddReplaceTag("&pound;",   'L', "symbol:pound");      // 163 £ : Pound
+      // AddReplaceTag("&curren;",  '$', "symbol:curren");     // 164 ¤ : Currency
+      // AddReplaceTag("&yen;",     'Y', "symbol:yen");        // 165 ¥ : Yen
+      // AddReplaceTag("&brvbar;",  '|', "symbol:brvbar");     // 166 ¦ : Broken vertical bar
+      // AddReplaceTag("&sect;",    'S', "symbol:sect");       // 167 § : Section
+      // AddReplaceTag("&uml;",     '~', "symbol:uml");        // 168 ¨ : Spacing diaeresis
+      // AddReplaceTag("&copy;",    'c', "symbol:copy");       // 169 © : Copyright
+      // AddReplaceTag("&ordf;",    'a', "symbol:ordf");       // 170 ª : Feminine ordinal indicator
+      // AddReplaceTag("&laquo;",   '<', "symbol:laquo");      // 171 « : Opening/Left angle quotation mark
+      // AddReplaceTag("&not;",     '~', "symbol:not");        // 172 ¬ : Negation
+      // AddReplaceTag("&shy ",     '-', "symbol:shy");        // 173 ­- : Soft hyphen
+      // AddReplaceTag("&reg;",     'R', "symbol:reg");        // 174 ® : Registered trademark
+      // AddReplaceTag("&macr;",    '-', "symbol:macr");       // 175 ¯ : Spacing macron
+      // AddReplaceTag("&deg;",     'o', "symbol:deg");        // 176 ° : Degree
+      // AddReplaceTag("&plusmn;",  '+', "symbol:plusmn");     // 177 ± : Plus or minus
+      // AddReplaceTag("&sup2;",    '2', "symbol:sup2");       // 178 ² : Superscript 2
+      // AddReplaceTag("&sup3;",    '3', "symbol:sup3");       // 179 ³ : Superscript 3
+      // AddReplaceTag("&acute;",   '\'', "symbol:acute");     // 180 ´ : Spacing acute
+      // AddReplaceTag("&micro;",   'u', "symbol:micro");      // 181 µ : Micro
+      // AddReplaceTag("&para;",    'P', "symbol:para");       // 182 ¶ : Paragraph
+      // AddReplaceTag("&dot;",     '.', "symbol:dot");        // 182 · : Dot
+      // AddReplaceTag("&cedil;",   ',', "symbol:cedil");      // 184 ¸ : Spacing cedilla
+      // AddReplaceTag("&sup1;",    '1', "symbol:sup1");       // 185 ¹ : Superscript 1
+      // AddReplaceTag("&ordm;",    'o', "symbol:ordm");       // 186 º : Masculine ordinal indicator
+      // AddReplaceTag("&raquo;",   '>', "symbol:raquo");      // 187 » : Closing/Right angle quotation mark
+      // AddReplaceTag("&frac14;",  '/', "symbol:frac14");     // 188 ¼ : Fraction 1/4
+      // AddReplaceTag("&frac12;",  '/', "symbol:frac12");     // 189 ½ : Fraction 1/2
[…]
