### 2024-02-29 ca7535d
Setup a much better organization for unicode information.
