+++ include/emp/text/Text_utils.hpp
@@ -26,283 +26,330 @@
 
 namespace emp {
 
+  // Information about 
+  struct TextSymbolInfo {
+    String html_tag;    ///< Tag to use to represent this symbol in HTML
+    char placeholder;   ///< A single character that can be used as a placeholder
+    String name;        ///< General symbol name (usually based on HTML name)
+    size_t id;          ///< Unicode ID number of symbol
+    String unicode;     ///< Unicode sequence for symbol (usually two characters)
+    String desc;        ///< English description of what symbol looks like
+  };
+
   /// A mapping of allowed symbol names to placeholder characters.
   static const auto & GetTextSymbolNames() {
-    static std::map<String, char> symbol_map;
+    static std::set<TextSymbolInfo> symbol_map;
     if (symbol_map.size() == 0) {
-      // AddReplaceTag("&iexcl;",   '!', "symbol:iexcl");      // Inverted '!'
-      // AddReplaceTag("&cent;",    'c', "symbol:cent");       // 162 ¢ : Cent
-      // AddReplaceTag("&pound;",   'L', "symbol:pound");      // 163 £ : Pound
-      // AddReplaceTag("&curren;",  '$', "symbol:curren");     // 164 ¤ : Currency
-      // AddReplaceTag("&yen;",     'Y', "symbol:yen");        // 165 ¥ : Yen
-      // AddReplaceTag("&brvbar;",  '|', "symbol:brvbar");     // 166 ¦ : Broken vertical bar
-      // AddReplaceTag("&sect;",    'S', "symbol:sect");       // 167 § : Section
-      // AddReplaceTag("&uml;",     '~', "symbol:uml");        // 168 ¨ : Spacing diaeresis
-      // AddReplaceTag("&copy;",    'c', "symbol:copy");       // 169 © : Copyright
-      // AddReplaceTag("&ordf;",    'a', "symbol:ordf");       // 170 ª : Feminine ordinal indicator
-      // AddReplaceTag("&laquo;",   '<', "symbol:laquo");      // 171 « : Opening/Left angle quotation mark
-      // AddReplaceTag("&not;",     '~', "symbol:not");        // 172 ¬ : Negation
-      // AddReplaceTag("&shy ",     '-', "symbol:shy");        // 173 ­- : Soft hyphen
-      // AddReplaceTag("&reg;",     'R', "symbol:reg");        // 174 ® : Registered trademark
-      // AddReplaceTag("&macr;",    '-', "symbol:macr");       // 175 ¯ : Spacing macron
-      // AddReplaceTag("&deg;",     'o', "symbol:deg");        // 176 ° : Degree
-      // AddReplaceTag("&plusmn;",  '+', "symbol:plusmn");     // 177 ± : Plus or minus
-      // AddReplaceTag("&sup2;",    '2', "symbol:sup2");       // 178 ² : Superscript 2
-      // AddReplaceTag("&sup3;",    '3', "symbol:sup3");       // 179 ³ : Superscript 3
-      // AddReplaceTag("&acute;",   '\'', "symbol:acute");     // 180 ´ : Spacing acute
-      // AddReplaceTag("&micro;",   'u', "symbol:micro");      // 181 µ : Micro
-      // AddReplaceTag("&para;",    'P', "symbol:para");       // 182 ¶ : Paragraph
-      // AddReplaceTag("&dot;",     '.', "symbol:dot");        // 182 · : Dot
-      // AddReplaceTag("&cedil;",   ',', "symbol:cedil");      // 184 ¸ : Spacing cedilla
[…]
