### 2024-02-29 1e542a3
Reintegrate necessary changes to world.hpp
