+++ include/emp/Evolve/World.hpp
@@ -943,6 +943,7 @@ namespace emp {
 
   };
 
+  #ifndef DOXYGEN_SHOULD_SKIP_THIS
   // =============================================================
   // ===                                                       ===
   // ===  Out-of-class member function definitions from above  ===
@@ -972,7 +973,7 @@ namespace emp {
 
     // Track the new systematics info
     for (Ptr<SystematicsBase<ORG> > s : systematics) {
-      s->AddOrg(*new_org, pos, (int) update);
+      s->AddOrg(*new_org, pos);
     }
 
     SetupOrg(*new_org, pos, *random_ptr);
@@ -996,7 +997,7 @@ namespace emp {
     }
 
     for (Ptr<SystematicsBase<ORG> > s : systematics) {
-      s->RemoveOrgAfterRepro(pos, update);          // Notify systematics about organism removal
+      s->RemoveOrgAfterRepro(pos);                   // Notify systematics about organism removal
     }
 
   }
@@ -1489,6 +1490,13 @@ namespace emp {
       pop.resize(0);
       std::swap(pops[0], pops[1]);            // Move next pop into place.
 
+      // Tell systematics manager to swap next population and population
+      // Needs to happen here so that you can refer to systematics in
+      // OnPlacement functions
+      for (Ptr<SystematicsBase<ORG>> s : systematics) {
+        s->Update();
+      }
+
       // Update the active population.
       num_orgs = 0;
       for (size_t i = 0; i < pop.size(); i++) {
@@ -1498,12 +1506,7 @@ namespace emp {
       }
     }
 
-    // 3. Handle systematics and any data files that need to be printed this update.
-
-    // Tell systematics manager to swap next population and population
-    for (Ptr<SystematicsBase<ORG>> s : systematics) {
-      s->Update();
-    }
+    // 3. Handle any data files that need to be printed this update.
 
     for (auto file : files) file->Update(update);
 
@@ -1734,6 +1737,7 @@ namespace emp {
       os << std::endl;
     }
   }
+  #endif // DOXYGEN_SHOULD_SKIP_THIS
 }
 
 #endif // #ifndef EMP_EVOLVE_WORLD_HPP_INCLUDE
