+++ tests/Evolve/Systematics.cpp
@@ -1,10 +1,9 @@
-/*
- *  This file is part of Empirical, https://github.com/devosoft/Empirical
- *  Copyright (C) Michigan State University, MIT Software license; see doc/LICENSE.md
- *  date: 2021
-*/
 /**
- *  @file
+ *  @note This file is part of Empirical, https://github.com/devosoft/Empirical
+ *  @copyright Copyright (C) Michigan State University, MIT Software license; see doc/LICENSE.md
+ *  @date 2021
+ *
+ *  @file Systematics.cpp
  */
 
 #include <filesystem>
@@ -24,40 +23,37 @@
 #include "emp/Evolve/World_output.hpp"
 #include "emp/hardware/AvidaGP.hpp"
 
-
-TEST_CASE("Test Systematics", "[Evolve]")
-{
-
+TEST_CASE("Test Systematics", "[Evolve]") {
   // Taxon
   emp::Taxon<std::string> tx(0, "a");
-  REQUIRE(tx.GetID() == 0);
-  REQUIRE(tx.GetParent() == nullptr);
-  REQUIRE(tx.GetInfo() == "a");
-  REQUIRE(tx.GetNumOrgs() == 0);
-  REQUIRE(tx.GetTotOrgs() == 0);
+  CHECK(tx.GetID() == 0);
+  CHECK(tx.GetParent() == nullptr);
+  CHECK(tx.GetInfo() == "a");
+  CHECK(tx.GetNumOrgs() == 0);
+  CHECK(tx.GetTotOrgs() == 0);
   tx.AddOrg();
-  REQUIRE(tx.GetNumOrgs() == 1);
+  CHECK(tx.GetNumOrgs() == 1);
   tx.RemoveOrg();
-  REQUIRE(tx.GetNumOrgs() == 0);
-  REQUIRE(tx.GetTotOrgs() == 1);
-  REQUIRE(tx.GetTotalOffspring() == 0);
+  CHECK(tx.GetNumOrgs() == 0);
+  CHECK(tx.GetTotOrgs() == 1);
+  CHECK(tx.GetTotalOffspring() == 0);
 
   emp::Ptr< emp::Taxon<std::string, emp::datastruct::no_data> > parentPtr(&tx);
   emp::Taxon<std::string> tx_1(1, "b", parentPtr);
-  REQUIRE(tx_1.GetParent() == parentPtr);
+  CHECK(tx_1.GetParent() == parentPtr);
   tx_1.AddTotalOffspring();
-  REQUIRE(tx_1.GetTotalOffspring() == 1);
-  REQUIRE(tx.GetTotalOffspring() == 1);
+  CHECK(tx_1.GetTotalOffspring() == 1);
+  CHECK(tx.GetTotalOffspring() == 1);
 
   // Systematics
   std::function<std::string(double &)> calc_taxon = [](double & o){ return o > 50.0 ? "large" : "small"; };
   emp::Systematics<double, std::string> sys1(calc_taxon);
-  REQUIRE(sys1.GetTrackSynchronous() == false);
-  REQUIRE(sys1.GetNumAncestors() == 0);
-  REQUIRE(sys1.GetNumActive() == 0);
-  REQUIRE(sys1.GetNumOutside() == 0);
-  REQUIRE(sys1.GetTreeSize() == 0);
-  REQUIRE(sys1.GetNumTaxa() == 0);
+  CHECK(sys1.GetTrackSynchronous() == false);
+  CHECK(sys1.GetNumAncestors() == 0);
+  CHECK(sys1.GetNumActive() == 0);
+  CHECK(sys1.GetNumOutside() == 0);
[…]
