### 2024-02-29 2d7b807
Clean up optional throw
