+++ include/emp/base/_optional_throw.hpp
@@ -19,40 +19,40 @@
 namespace emp {
 
     /// Base case for assert_print...
-    inline void assert_print(std::stringstream &) { ; }
+    inline void assert_print_opt(std::stringstream &) { ; }
 
     /// Print out information about the next variable and recurse...
     template <typename T, typename... EXTRA>
-    void assert_print(std::stringstream & ss, std::string name, T && val, EXTRA &&... extra) {
+    void assert_print_opt(std::stringstream & ss, std::string name, T && val, EXTRA &&... extra) {
         if constexpr ( emp::is_streamable<std::stringstream, T>::value ) {
         ss << name << ": [" << val << "]" << std::endl;
         } else ss << name << ": (non-streamable type)" << std::endl;
-        assert_print(ss, std::forward<EXTRA>(extra)...);
+        assert_print_opt(ss, std::forward<EXTRA>(extra)...);
     }
 
     template <typename T, typename... EXTRA>
-    void assert_print_second(std::stringstream & ss, std::string name, T && val, EXTRA &&... extra) {
-        assert_print(ss, std::forward<EXTRA>(extra)...);
+    void assert_print_second_opt(std::stringstream & ss, std::string name, T && val, EXTRA &&... extra) {
+        assert_print_opt(ss, std::forward<EXTRA>(extra)...);
     }
 
     template <typename T>
-    void assert_print_second(std::stringstream & ss, std::string name, T && val) {;}
+    void assert_print_second_opt(std::stringstream & ss, std::string name, T && val) {;}
 
     template <typename T, typename... EXTRA>
-    void assert_print_first(std::stringstream & ss, std::string name, T && val, EXTRA &&... extra) {
+    void assert_print_first_opt(std::stringstream & ss, std::string name, T && val, EXTRA &&... extra) {
         if constexpr ( emp::is_streamable<std::stringstream, T>::value ) {
         ss << name << ": [" << val << "]" << std::endl;
         } else ss << name << ": (non-streamable type)" << std::endl;
-        assert_print_second(ss, std::forward<EXTRA>(extra)...);
+        assert_print_second_opt(ss, std::forward<EXTRA>(extra)...);
     }
 
-    void assert_print_first(std::stringstream & ss, int placeholder) {;}
+    void assert_print_first_opt(std::stringstream & ss, int placeholder) {;}
 
     template <typename... EXTRA>
-    void assert_throw(std::string filename, size_t line, std::string expr, std::string message, EXTRA &&... extra) {
+    void assert_throw_opt(std::string filename, size_t line, std::string expr, std::string message, EXTRA &&... extra) {
         std::stringstream ss;
         ss << "Internal Error (in " << filename << " line " << line << "): " << expr << ".\n\n Message: " << message << "\n\n";
-        assert_print_first(ss, std::forward<EXTRA>(extra)...);
+        assert_print_first_opt(ss, std::forward<EXTRA>(extra)...);
         throw(std::runtime_error(ss.str()));
     }
 }
