### 2024-02-29 d34e95a
Add bounds checking
