### 2024-02-29 98a114f
Discard changes to include/emp/datastructs/SmallFifoMap.hpp
