### 2024-02-29 a3e835a
Discard changes to include/emp/hardware/AvidaGP.hpp
