### 2024-02-29 8e61986
Discard changes to tests/tools/string_utils.cpp
