+++ tests/tools/string_utils.cpp
@@ -389,7 +389,6 @@ TEST_CASE("Another Test string_utils", "[tools]")
   );
   CHECK(slices.size() == 1);
   CHECK(slices[0] == "This is a test.");
-  // CHECK(slices[0] == "This");
 
   slices = emp::slice(
     "This is a test.",
@@ -399,7 +398,6 @@ TEST_CASE("Another Test string_utils", "[tools]")
   CHECK(slices.size() == 2);
   CHECK(slices[0] == "This");
   CHECK(slices[1] == "is a test.");
-  // CHECK(slices[1] == "is");
 
   slices = emp::slice(
     "This is a test.",
@@ -410,7 +408,6 @@ TEST_CASE("Another Test string_utils", "[tools]")
   CHECK(slices[0] == "This");
   CHECK(slices[1] == "is");
   CHECK(slices[2] == "a test.");
-  // CHECK(slices[2] == "a");
 
   // Try other ways of using slice().
   emp::slice(base_string, slices, 's');
