### 2024-02-29 5abbf48
Discard changes to tests/matching/MatchBin_logging.cpp
