+++ include/emp/text/Text_utils.hpp
@@ -348,86 +348,88 @@ namespace emp {
       symbol_map.emplace("&upsih;",   'U', "upsih",      978,  "ϒ",   "Upsilon symbol");
       symbol_map.emplace("&piv;",     'O', "piv",        982,  "ϖ",   "Pi symbol");
 
-      // AddReplaceTag("&ensp;",      ' ', "symbol:ensp");       // 8194 : En space
-      // AddReplaceTag("&emsp;",      ' ', "symbol:emsp");       // 8195 : Em space
-      // AddReplaceTag("&thinsp;",  ' ', "symbol:thinsp");     // 8201 : Thin space
-      // AddReplaceTag("&zwnj;",    ' ', "symbol:zwnj");       // 8204 : Zero width non-joiner
-      // AddReplaceTag("&zwj;",     ' ', "symbol:zwj");        // 8205 : Zero width joiner
-      // AddReplaceTag("&lrm;",     '?', "symbol:lrm");        // 8206 : Left-to-right mark
-      // AddReplaceTag("&rlm;",     '?', "symbol:rlm");        // 8207 : Right-to-left mark
-      // AddReplaceTag("&ndash;",   '-', "symbol:ndash");      // 8211 – : En dash
-      // AddReplaceTag("&mdash;",   '-', "symbol:mdash");      // 8212 — : Em dash
-      // AddReplaceTag("&lsquo;",   ' ', "symbol:lsquo");      // 8216 ‘ : Left single quotation mark
-      // AddReplaceTag("&rsquo;",   ' ', "symbol:rsquo");      // 8217 ’ : Right single quotation mark
-      // AddReplaceTag("&sbquo;",   ' ', "symbol:sbquo");      // 8218 ‚ : Single low-9 quotation mark
-      // AddReplaceTag("&ldquo;",   ' ', "symbol:ldquo");      // 8220 “ : Left double quotation mark
-      // AddReplaceTag("&rdquo;",   ' ', "symbol:rdquo");      // 8221 ” : Right double quotation mark
-      // AddReplaceTag("&bdquo;",   ' ', "symbol:bdquo");      // 8222 „ : Double low-9 quotation mark
-      // AddReplaceTag("&dagger;",  ' ', "symbol:dagger");     // 8224 † : Dagger
-      // AddReplaceTag("&Dagger;",  ' ', "symbol:Dagger");     // 8225 ‡ : Double dagger
-      // AddReplaceTag("&bull;",    ' ', "symbol:bull");       // 8226 • : Bullet
-      // AddReplaceTag("&hellip;",  ' ', "symbol:hellip");     // 8230 … : Horizontal ellipsis
-      // AddReplaceTag("&permil;",  ' ', "symbol:permil");     // 8240 ‰ : Per mille
-      // AddReplaceTag("&prime;",   ' ', "symbol:prime");      // 8242 ′ : Minutes (Degrees)
-      // AddReplaceTag("&Prime;",   ' ', "symbol:Prime");      // 8243 ″ : Seconds (Degrees)
-      // AddReplaceTag("&lsaquo;",  ' ', "symbol:lsaquo");     // 8249 ‹ : Single left angle quotation
-      // AddReplaceTag("&rsaquo;",  ' ', "symbol:rsaquo");     // 8250 › : Single right angle quotation
-      // AddReplaceTag("&oline;",   ' ', "symbol:oline");      // 8254 ‾ : Overline
-      // AddReplaceTag("&euro;",    ' ', "symbol:euro");       // 8364 € : Euro
-      // AddReplaceTag("&trade;",   ' ', "symbol:trade");      // 8482 ™ : Trademark
[…]
