### 2024-03-01 2e7856b
Lots more cleanup on symbols in Text_utils.
