+++ include/emp/text/Text_utils.hpp
@@ -28,12 +28,22 @@ namespace emp {
 
   // Information about 
   struct TextSymbolInfo {
-    String html_tag;    ///< Tag to use to represent this symbol in HTML
-    char placeholder;   ///< A single character that can be used as a placeholder
-    String name;        ///< General symbol name (usually based on HTML name)
-    size_t id;          ///< Unicode ID number of symbol
-    String unicode;     ///< Unicode sequence for symbol (usually two characters)
-    String desc;        ///< English description of what symbol looks like
+    String name = "";        ///< General symbol name (usually based on HTML name)
+    size_t id = 0;           ///< Unicode ID number of symbol
+    String unicode = "";     ///< Unicode sequence for symbol (usually two characters)
+    char placeholder = '\0'; ///< A single character that can be used as a placeholder
+    String desc = "";        ///< English description of what symbol looks like
+    String html_tag = "";    ///< Tag to use to represent this symbol in HTML
+
+    TextSymbolInfo() = default;
+    TextSymbolInfo(const TextSymbolInfo &)= default;
+    TextSymbolInfo(TextSymbolInfo &&)= default;
+    TextSymbolInfo(String name, size_t id, String unicode, char placeholder, String desc)
+      : name(name), id(id), unicode(unicode), placeholder(placeholder), desc(desc) {
+        html_tag = MakeString('&', name, ';');
+      }
+
+    auto operator<=>(const TextSymbolInfo &) const = default;
   };
 
   /// A mapping of allowed symbol names to placeholder characters.
@@ -44,392 +54,392 @@ namespace emp {
       // === UNICODE Latin-1 Supplement ===
 
       // Punctuation & Symbols
-      symbol_map.emplace("&nbsp;",    ' ',  "nbsp",      160,  " ",   "Non-breaking space");
-      symbol_map.emplace("&iexcl;",   '!',  "iexcl",     161,  "¡",   "Inverted exclamation point");
-      symbol_map.emplace("&cent;",    'c',  "cent",      162,  "¢",   "Cent");
-      symbol_map.emplace("&pound;",   'L',  "pound",     163,  "£",   "Pound");
-      symbol_map.emplace("&curren;",  '$',  "curren",    164,  "¤",   "Currency");
-      symbol_map.emplace("&yen;",     'Y',  "yen",       165,  "¥",   "Yen");
-      symbol_map.emplace("&brvbar;",  '|',  "brvbar",    166,  "¦",   "Broken vertical bar");
-      symbol_map.emplace("&sect;",    'S',  "sect",      167,  "§",   "Section");
-      symbol_map.emplace("&uml;",     '~',  "uml",       168,  "¨",   "Spacing diaeresis");
-      symbol_map.emplace("&copy;",    'c',  "copy",      169,  "©",   "Copyright");
-      symbol_map.emplace("&ordf;",    'a',  "ordf",      170,  "ª",   "Feminine ordinal indicator");
-      symbol_map.emplace("&laquo;",   '<',  "laquo",     171,  "«",   "Opening/Left angle quotation mark");
-      symbol_map.emplace("&not;",     '~',  "not",       172,  "¬",   "Negation");
-      symbol_map.emplace("&shy ",     '-',  "shy",       173,  "­-",   "Soft hyphen");
[…]
