+++ include/emp/Evolve/Systematics.hpp
@@ -5,7 +5,13 @@
 */
 /**
  *  @file
- *  @brief TODO.
+ *  @brief Track genotypes, species, clades, or lineages of organisms in a world.
+ *  
+ *  Note that this file powers the Python library phylotrackpy. The main consequence
+ *  is that we should generally prefer emp_optional_throw to emp_assert in any circumstance
+ *  where user input could trigger an error. emp_assert will trigger a segfault in Python 
+ *  (killing the whole interpreter), whereas emp_optional_throw will raise a Python exception.
+ * 
  *  @todo We should provide an option to back up systematics data to a file so that it doesn't all
  *        need to be kept in memory, especially if we're only doing post-analysis.
  *  @todo This inheritance system makes adding new systematics-related data tracking kind of a pain.
@@ -26,6 +32,7 @@
 #include <string>
 #include <unordered_set>
 
+#include "../base/assert_warning.hpp"
 #include "../base/optional_throw.hpp"
 #include "../base/Ptr.hpp"
 #include "../control/Signal.hpp"
@@ -44,7 +51,7 @@
 
 namespace emp {
 
-  template <typename A, typename B, typename C>
+  template <typename ORG, typename ORG_INFO, typename DATA_STRUCT>
   class Systematics;
 
   /// The systematics manager allows an optional second template type that
@@ -104,7 +111,7 @@ namespace emp {
       /// @param muts can contain as many strings (types of mutation) as desired, each accompanied
       /// by a number indicating how many of that mutation occurred
       /// Example: {"point_mutation":2, "insertion":1}
-      void RecordMutation(std::unordered_map<std::string, int> & muts) {
+      void RecordMutation(const std::unordered_map<std::string, int> & muts) {
         for (auto mut : muts) {
           if (Has(mut_counts, mut.first)) {
             mut_counts[mut.first] += mut.second;
@@ -137,7 +144,7 @@ namespace emp {
   /// track an evolutionary pathway)
   template <typename ORG_INFO, typename DATA_STRUCT = emp::datastruct::no_data>
   class Taxon {
-    template <typename A, typename B, typename C>
+    template <typename ORG, typename SYS_ORG_INFO, typename SYS_DATA_STRUCT>
     friend class Systematics;
   protected:
     using this_t = Taxon<ORG_INFO, DATA_STRUCT>;
@@ -300,7 +307,7 @@ namespace emp {
     size_t org_count;         ///< How many organisms are currently active?
     size_t total_depth;       ///< Sum of taxa depths for calculating average.
     size_t num_roots;         ///< How many distinct injected ancestors are currently in population?
-    int max_depth;            ///< Depth of deepest taxon. -1 means needs to be recalculated
+    mutable int max_depth;    ///< Depth of deepest taxon. -1 means needs to be recalculated
 
     size_t next_id;           ///< What ID value should the next new taxon have?
     size_t curr_update;
@@ -417,7 +424,7 @@ namespace emp {
     virtual size_t GetNumOutside() const = 0;
     virtual size_t GetTreeSize() const = 0;
     virtual size_t GetNumTaxa() const = 0;
-    virtual int GetMaxDepth() = 0;
+    virtual int GetMaxDepth() const = 0;
     virtual int GetPhylogeneticDiversity() const = 0;
     virtual double GetMeanPairwiseDistance(bool branch_only) const = 0;
     virtual double GetSumDistance() const = 0;
@@ -613,7 +620,6 @@ namespace emp {
     /// Add information about a new organism, including its stored info and parent's taxon;
     /// If you would like the systematics manager to track taxon age, you can also supply
[…]
