### 2024-03-02 dd09c54
Swap in ternary
