### 2024-03-03 de6a0c3
Setup TextEncoding to return rule object when one is added.
