### 2024-03-04 213e8d5
Switch Emphatic tests from REQUIRE to CHECK
