+++ tests/text/EmphaticText.cpp
@@ -15,62 +15,63 @@
 
 TEST_CASE("Testing EmphaticText", "[text]") {
   emp::EmphaticText text("Test Text");
-  REQUIRE(text.GetSize() == 9);
-  REQUIRE(text.AsString() == "Test Text");
-  REQUIRE(text.Encode() == "Test Text");
+  text.PrintDebug();
+  CHECK(text.GetSize() == 9);
+  CHECK(text.AsString() == "Test Text");
+  CHECK(text.Encode() == "Test Text");
 
   // Try adding style.
   text.Bold(5,9);
-  REQUIRE(text.GetSize() == 9);
-  REQUIRE(text.AsString() == "Test Text");
-  REQUIRE(text.Encode() == "Test `*Text`*");
+  CHECK(text.GetSize() == 9);
+  CHECK(text.AsString() == "Test Text");
+  CHECK(text.Encode() == "Test `*Text`*");
 
   // Try appending.
   text << " and more Text.";
-  REQUIRE(text.GetSize() == 24);
-  REQUIRE(text.AsString() == "Test Text and more Text.");
-  REQUIRE(text.Encode() == "Test `*Text`* and more Text.");
+  CHECK(text.GetSize() == 24);
+  CHECK(text.AsString() == "Test Text and more Text.");
+  CHECK(text.Encode() == "Test `*Text`* and more Text.");
 
   // Try changing letters.
   text[21] = 's';
   text[23] = 's';
-  REQUIRE(text.GetSize() == 24);
-  REQUIRE(text.AsString() == "Test Text and more Tests");
-  REQUIRE(text.Encode() == "Test `*Text`* and more Tests");
+  CHECK(text.GetSize() == 24);
+  CHECK(text.AsString() == "Test Text and more Tests");
+  CHECK(text.Encode() == "Test `*Text`* and more Tests");
 
   // Try making a change that involves style.
   text[19] = text[5];
-  REQUIRE(text.GetSize() == 24);
-  REQUIRE(text.AsString() == "Test Text and more Tests");
-  REQUIRE(text.Encode() == "Test `*Text`* and more `*T`*ests");
+  CHECK(text.GetSize() == 24);
+  CHECK(text.AsString() == "Test Text and more Tests");
+  CHECK(text.Encode() == "Test `*Text`* and more `*T`*ests");
 
   // Try erasing the text.
   text.Resize(0);
-  REQUIRE(text.GetSize() == 0);
-  REQUIRE(text.AsString() == "");
-  REQUIRE(text.Encode() == "");
-  REQUIRE(text.GetStyles().size() == 0);
+  CHECK(text.GetSize() == 0);
+  CHECK(text.AsString() == "");
+  CHECK(text.Encode() == "");
+  CHECK(text.GetStyles().size() == 0);
 
   text << "This is `/`*Pre-`/formatted`* text.";
-  // REQUIRE(text.GetSize() == 27);
-  REQUIRE(text.AsString() == "This is Pre-formatted text.");
-  REQUIRE(text.Encode() == "This is `*`/Pre-`/formatted`* text.");
+  // CHECK(text.GetSize() == 27);
+  CHECK(text.AsString() == "This is Pre-formatted text.");
+  CHECK(text.Encode() == "This is `*`/Pre-`/formatted`* text.");
 
   emp::Text plain_text("The `* and `* here should not be converted.");
-  REQUIRE(plain_text.Encode() == "The `* and `* here should not be converted.");
+  CHECK(plain_text.Encode() == "The `* and `* here should not be converted.");
 
   // Now let's bold the word "not".
   plain_text.Bold(26,29);
 
   // Should not change the output of Encode(), just internal style.
[…]
