### 2024-03-06 1359e85
Starting to work on a bi-directional map, but interface needs to be MUCH cleaner.
