### 2024-03-20 a8a00c3
Checked Bits::FindOne functions to return size_t and npos for failure.
