+++ include/emp/bits/Bits.hpp
@@ -1,7 +1,7 @@
 /**
  *  @note This file is part of Empirical, https://github.com/devosoft/Empirical
  *  @copyright Copyright (C) Michigan State University, MIT Software license; see doc/LICENSE.md
- *  @date 2022-23.
+ *  @date 2022-24.
  *
  *  @file Bits.hpp
  *  @brief A generic bit-handler to replace vector<bool>, etc +additional bitwise logic features.
@@ -159,6 +159,8 @@ namespace emp {
     constexpr void ROTR_SELF(const size_t shift_size_raw);
 
   public:
+    static constexpr size_t npos = static_cast<size_t>(-1);
+
     /// @brief Default constructor; will build the default number of bits (often 0, but not always)
     /// @param init_val Initial value of all default bits.
     Bits(bool init_val=0) { if (init_val) SetAll(); else Clear(); }
@@ -580,30 +582,30 @@ namespace emp {
     /// @param num number of bits to delete, default 1.
     void Delete(const size_t index, const size_t num=1);
 
-    /// @brief Return the position of the first one; return -1 if no ones in vector.
-    [[nodiscard]] int FindOne() const;
+    /// @brief Return the position of the first one; return npos if no ones in vector.
+    [[nodiscard]] size_t FindOne() const;
 
-    /// @brief Return the position of the first zero; return -1 if no zeroes in vector.
-    [[nodiscard]] int FindZero() const;
+    /// @brief Return the position of the first zero; return npos if no zeroes in vector.
+    [[nodiscard]] size_t FindZero() const;
 
 
     /// Deprecated: Return the position of the first one; return -1 if no ones in vector.
     [[deprecated("Renamed to more accurate FindOne()")]]
     [[nodiscard]] int FindBit() const { return FindOne(); }
 
-    /// @brief Return the position of the first one after start_pos (or -1 if none)
+    /// @brief Return the position of the first one after start_pos (or npos if none)
     /// You can loop through all 1-bit positions of Bits object "bits" with:
     ///
-    ///   for (int pos = bits.FindOne(); pos >= 0; pos = bits.FindOne(pos+1)) { ... }
+    ///   for (size_t pos = bits.FindOne(); pos < bits.size(); pos = bits.FindOne(pos+1)) { ... }
 
-    [[nodiscard]] int FindOne(const size_t start_pos) const;
+    [[nodiscard]] size_t FindOne(const size_t start_pos) const;
 
-    /// @brief Return the position of the first zero after start_pos (or -1 if none)
+    /// @brief Return the position of the first zero after start_pos (or npos if none)
     /// You can loop through all 0-bit positions of Bits object "bits" with:
     ///
-    ///   for (int pos = bits.FindZero(); pos >= 0; pos = bits.FindZero(pos+1)) { ... }
+    ///   for (size_t pos = bits.FindZero(); pos < bits.size(); pos = bits.FindZero(pos+1)) { ... }
 
-    [[nodiscard]] int FindZero(const size_t start_pos) const;
+    [[nodiscard]] size_t FindZero(const size_t start_pos) const;
 
     /// @brief Special version of FindOne takes int; common way to call.
     [[nodiscard]] int FindOne(int start_pos) const {
@@ -620,10 +622,10 @@ namespace emp {
     [[nodiscard]] int FindBit(const size_t start_pos) const;
 
     /// @brief Find the most-significant set-bit.
-    [[nodiscard]] int FindMaxOne() const;
+    [[nodiscard]] size_t FindMaxOne() const;
 
[…]
