### 2024-03-21 446fae4
Fixed warnings in Bits_Data
