### 2024-03-23 35837db
Added a test for the new Bits::ForEach function.
