### 2024-03-24 ed1bb2b
Added a test for the new Bits::ForEachPair function.
