### 2024-03-25 25eebe1
Added TestAll, TestAny, and TestNone to test one-indices in Bits
