### 2024-03-26 c5ff12b
Added tests for Bits::TestAll, TestNone, and TestAny
