+++ demos/Graphs/Steiner/Graph.hpp
@@ -1,0 +1,263 @@
+#pragma once
+
+#include "../../../include/emp/base/vector.hpp"
+#include "../../../include/emp/bits/BitVector.hpp"
+#include "../../../include/emp/debug/debug.hpp"
+
+class Graph {
+private:
+  static constexpr size_t MAX_BITS = 200;
+  using bits_t = emp::StaticBitVector<MAX_BITS>;
+
+  emp::vector<bits_t> edges;
+  bits_t keys;       // Key vertices that we still need to merge.
+  bits_t included;   // Current and former key vertices that are included in the solution.
+  bits_t unknown;    // Vertices that we still need to consider for the solution.
+
+public:
+  Graph() { }
+  Graph(size_t node_count) { Resize(node_count); }
+  Graph(const Graph &) = default;
+  Graph(Graph &&) = default;
+
+  Graph & operator=(const Graph &) = default;
+  Graph & operator=(Graph &&) = default;
+
+  bool HasKey(size_t id) const { return keys.Has(id); }
+  bool HasEdge(size_t id1, size_t id2) const { return edges[id1].Has(id2); }
+  size_t GetDegree(size_t id) const { return edges[id].CountOnes(); }
+
+  size_t GetKeyCount() const { return keys.CountOnes(); }
+  size_t GetIncludeSize() const { return included.CountOnes(); }
+  size_t GetUnknownSize() const { return unknown.CountOnes(); }
+  bool IsSolved() const { return GetKeyCount() == 1; }
+
+  bool IsUnknown(size_t id) const { return unknown[id]; }
+  bool IsIncluded(size_t id) const { return included[id]; }
+  bool IsKey(size_t id) const { return keys[id]; }
+  bool IsActive(size_t id) const { return IsUnknown(id) || IsKey(id); }
+
+  const bits_t & GetIncluded() const { return included; }
+  const bits_t & GetKeys() const { return keys; }
+  const bits_t & GetUnknown() const { return unknown; }
+
+  size_t GetNextID() const { return static_cast<size_t>(unknown.FindOne()); }
+
+  bool IsUnsolvable() const {
+    if (IsSolved()) return false;
+    for (int i = keys.FindOne(); i >= 0; i = keys.FindOne(i+1)) {
+      if (GetDegree(i) == 0) return true;
+    }
+    return false;
+  }
+
+  void Resize(size_t new_size) {
+    edges.resize(new_size);
+    for (auto & v : edges) v.resize(new_size);
+    keys.resize(new_size);
+    included.resize(new_size);
+    unknown.resize(new_size);
+    unknown.SetAll();
+  }
+
+  void AddEdge(size_t id1, size_t id2) {
+    if (id1 != id2) {
+      edges[id1].Set(id2);
+      edges[id2].Set(id1);
+    }
+  }
+
+  void RemoveEdge(size_t id1, size_t id2) {
+    edges[id1].Clear(id2);
+    edges[id2].Clear(id1);
+  }
+
+  void SetKey(size_t id) {
+    keys.Set(id);
+    included.Set(id);
+    unknown.Clear(id);
+  }
+
+  void EraseNode(size_t id) {
+    // Remove all of the edges associated with this node.
+    while (edges[id].Any()) {
+      RemoveEdge(id, edges[id].FindOne());
+    }
+
+    // Remove other records of this node.
+    keys.Clear(id); // If id was a key, it shouldn't be anymore.
+    // Leave as included.
+    unknown.Clear(id);
+  }
+
+  // Collapse the second node into the first.
+  void MergeNodes(size_t id1, size_t id2) {
+    if (id1 == id2) return; // Cannot merge a node with itself.
[…]
