### 2024-03-31 7707a5e
Added a Solver to brute-force solve graph problems.
