### 2024-03-31 6602baf
Made Bits::ForEach and ForEachPair const member functions.
