### 2024-03-31 081d8c5
Clean up function in debug.hpp
