### 2024-04-01 58a385b
Added main.cpp to run the solver.
