### 2024-04-01 238ff3c
Lots of speed-ups to Graph and Solver
