+++ demos/Graphs/Steiner/Solver.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "../../../include/emp/base/array.hpp"
 #include "../../../include/emp/base/vector.hpp"
 
 #include "Graph.hpp"
@@ -7,20 +8,24 @@
 class Solver {
 private:
   static constexpr bool VERBOSE = false;
+  static constexpr size_t MAX_BITS = 256;
+  using graph_t = Graph<MAX_BITS>;
 
-  Graph start_graph;
-  Graph best_graph;
+  graph_t start_graph;
+  graph_t best_graph;
   size_t best_size = 10000000;
 
-  void MarkSolution(const Graph & graph) {
+  emp::array<graph_t, MAX_BITS> graphs;
+
+  void MarkSolution(const graph_t & graph) {
     if (graph.GetIncludeSize() < best_size) {
       best_graph = graph;
       best_size = graph.GetIncludeSize();
-      std::cout << "New Best: " << best_size << std::endl;
+      std::cerr << "New Best: " << best_size << std::endl;
     }
   }
 
-  void PrintSolveState(const Graph & graph, size_t depth) const {
+  void PrintSolveState(const graph_t & graph, size_t depth) const {
     for (size_t i = 0; i < depth; ++i) std::cout << "  ";
     std::cout << "Solve(" << depth << "):"
               << " keys:" << graph.GetKeys()
@@ -31,7 +36,7 @@ private:
   }
 
 public:
-  Solver(const Graph & graph) : start_graph(graph) { }
+  Solver(const graph_t & graph) : start_graph(graph) { }
 
   void Debug() {
     PrintSolveState(start_graph, 0);
@@ -40,30 +45,38 @@ public:
   }
 
   void Solve() {
-    Graph graph(start_graph);
-    graph.Setup();
-    Solve(graph, 0);
+    graphs[0] = start_graph;
+    graphs[0].Setup();
+    Solve(0);
   }
 
-  void Solve(Graph graph, size_t depth) {
+  void Solve(size_t depth) {
+    graph_t & graph = graphs[depth];
     emp_debug("SOLVE: ", depth, " key=", graph.GetKeys(), " included=", graph.GetIncluded(), " unknown=", graph.GetUnknown());
 
     if constexpr (VERBOSE) PrintSolveState(graph, depth);
 
     graph.Optimize();
 
-    if (graph.IsSolved()) { MarkSolution(graph); return; }  // Solved!
-    if (graph.GetIncludeSize()+1 >= best_size) return;      // Bounded!
-    if (graph.GetUnknownSize() == 0) return;                // No solution possible.
+    if (graph.IsSolved()) { MarkSolution(graph); return; }            // Solved!
+    if (graph.TestOneNodeSolution()) { MarkSolution(graph); return; } // Solved with one new node.
+    if (graph.TestTwoNodeSolution()) { MarkSolution(graph); return; } // Solved with two new nodes.
+    if (graph.GetIncludeSize()+3 >= best_size) return;                // Bounded!
+    if (graph.GetUnknownSize() == 0) return;                          // No solution possible.
 
     size_t next_id = graph.GetNextID();
     emp_debug("  next_id=", next_id);
-    // if constexpr (VERBOSE) std::cout << "(Exclude next_id = " << next_id << ")" << std::endl;
-    graph.Exclude(next_id);
-    Solve(graph, depth+1);
+    graph_t & next_graph = graphs[depth+1];
+
     // if constexpr (VERBOSE) std::cout << "(Include next_id = " << next_id << ")" << std::endl;
-    graph.Include(next_id);
-    Solve(graph, depth+1);
+    next_graph = graph;
[…]
