### 2024-04-02 6a29591
Setup following of degree-2 paths and better bounding based on max key degree.
