### 2024-04-02 547c6f6
Added CalcReachable() and TestSolvable() to Graph.
