### 2024-04-02 a22e504
Updated solver to print 200 if unsolvable.
