### 2024-04-04 f1b130d
Some cleanup on SolveState
