### 2024-04-04 c777a84
Set Ptr constructors/destructors to =default when possible.
