### 2024-04-05 461ce91
Improved use of default functions in notify.hpp and vector.hpp
