+++ include/emp/compiler/lexer_utils.hpp
@@ -1,7 +1,7 @@
 /**
  *  @note This file is part of Empirical, https://github.com/devosoft/Empirical
  *  @copyright Copyright (C) Michigan State University, MIT Software license; see doc/LICENSE.md
- *  @date 2016-2017
+ *  @date 2016-2024.
  *
  *  @file lexer_utils.hpp
  *  @brief A set of utilities to convert between NFAs and DFAs
@@ -33,41 +33,37 @@ namespace emp {
   /// Systematic conversion of NFA to DFA...
   static inline DFA to_DFA(const NFA & nfa, int keep_invalid=false) {
     DFA dfa(1);                                  // Setup zero to be the start state.
-    std::map<std::set<size_t>, size_t> id_map;   // How do nfa state sets map to dfa states?
-    std::vector<std::set<size_t>> state_stack;   // Which states still need to be explored?
-    state_stack.emplace_back(nfa.GetStart());    // Place the starting point in the state_stack.
+    std::map<std::set<size_t>, size_t> id_map;   // Map nfa "state sets" to dfa states.
+    std::vector<std::set<size_t>> state_stack;   // States that still need to be explored.
+
+    state_stack.emplace_back(nfa.GetStart());    // Place the starting state in the state_stack.
     id_map[state_stack[0]] = 0;                  // Give starting point ID 0.
 
     // Loop through all states not full explored; remove top state and add new states.
     while (state_stack.size()) {
       // Get the next state to test.
-      std::set<size_t> cur_state = state_stack.back();
+      const std::set<size_t> cur_state = state_stack.back();
       const size_t cur_id = id_map[cur_state];
       state_stack.pop_back();
 
       // Determine if this state should be a STOP state and always use HIGHEST stop value.
       for (auto s : cur_state) dfa.AddStop(cur_id, nfa.GetStop(s));
 
-      // Run through all possible transitions
+      // Account for all possible transitions
       for (size_t sym = 0; sym < NFA::NUM_SYMBOLS; sym++) {
         std::set<size_t> next_state = nfa.GetNext(sym, cur_state);
         if (next_state.size() == 0 && !keep_invalid) continue;  // Discard invalid transitions.
 
-        // Remove NFA states with ONLY free transisions (they will all have been taken already)
-        // @CAO do more elegantly!
-        emp::vector<size_t> remove_set;
-        for (auto x : next_state) if (nfa.IsEmpty(x)) remove_set.push_back(x);
-        for (auto x : remove_set) next_state.erase(x);
+        // Remove NFA states with ONLY free transitions (they will all have been taken already)
+        std::erase_if(next_state, [&nfa](size_t x) { return nfa.IsEmpty(x); });
 
-        // Determine if we have a new state in the DFA.
-        if (id_map.find(next_state) == id_map.end()) {
-          const size_t next_id = dfa.GetSize();
-          id_map[next_state] = next_id;
-          dfa.Resize(next_id + 1);
+        // If we need a new state in the DFA, add it and put it on the stack to explore.
+        if (!id_map.contains(next_state)) {
+          id_map[next_state] = dfa.AddState();
           state_stack.emplace_back(next_state);
         }
 
-        // Setup the new connection in the DFA
+        // Set up the new connection in the DFA
         const size_t next_id = id_map[next_state];
         dfa.SetTransition(cur_id, next_id, sym);
       }
@@ -92,7 +88,7 @@ namespace emp {
   }
 
 
-  /// Merge multiple automata into one NFA (base case, single converstion)
+  /// Merge multiple automata into one NFA (base case; single conversion)
   template <typename T1>
   static NFA MergeNFA(T1 && in) {
[…]
