### 2024-04-07 8225ef1
Cleaned up config files for defaults and docs.
