+++ include/emp/config/ConfigManager.hpp
@@ -42,7 +42,7 @@ namespace emp {
   public:
     ConfigManager_Base(const std::string & _type, const std::string & _command)
       : type_keyword(_type), command_keyword(_command) { ; }
-    virtual ~ConfigManager_Base() { ; }
+    virtual ~ConfigManager_Base() = default;
 
     const std::string & GetTypeKeyword() const { return type_keyword; }
     const std::string & GetCommandKeyword() const { return command_keyword; }
@@ -66,7 +66,7 @@ namespace emp {
       , callback_fun(_fun)
     {
     }
-    ~ConfigManager() { ; }
+    ~ConfigManager() = default;
 
     bool HasObject(const std::string & obj_name) {
       return (name_map.find(obj_name) != name_map.end());
