### 2024-04-08 c99dd6b
Cleanup on SimpleParser.hpp
