### 2024-04-10 dcd1f00
Cleaned up default handling in NK.hpp
