### 2024-04-11 98c87fd
Cleaned up Evolve/Resource.hpp
