+++ include/emp/Evolve/Resource.hpp
@@ -23,9 +23,9 @@ namespace emp {
 
   class Resource {
   private:
-    double amount = 0;
-    double inflow = 0;
-    double outflow = 0;
+    double amount = 0.0;
+    double inflow = 0.0;
+    double outflow = 0.0;
 
   public:
 
@@ -33,16 +33,16 @@ namespace emp {
     Resource(double amt, double in, double out) :
       amount(amt), inflow(in), outflow(out){};
 
-    double GetAmount() {return amount;}
-    double GetInflow() {return inflow;}
-    double GetOutflow() {return outflow;}
+    [[nodiscard]] double GetAmount() const { return amount; }
+    [[nodiscard]] double GetInflow() const { return inflow; }
+    [[nodiscard]] double GetOutflow() const { return outflow; }
 
-    void SetAmount(double amt) {amount = amt;}
-    void SetInflow(double in) {inflow = in;}
-    void SetOutflow(double out) {outflow = out;}
+    void SetAmount(double amt) { amount = amt; }
+    void SetInflow(double in) { inflow = in; }
+    void SetOutflow(double out) { outflow = out; }
 
-    double Inc() {amount += inflow; return amount;}
-    double Inc(double amt) {amount += amt; return amount;}
+    double Inc() { amount += inflow; return amount; }
+    double Inc(double amt) { amount += amt; return amount; }
     double Dec() {
       amount -= amount*outflow;
       if (amount < 0) {amount = 0;}
