### 2024-04-12 22ae6bc
Clean up on StateGrid.hpp
