+++ include/emp/Evolve/StateGrid.hpp
@@ -1,7 +1,7 @@
 /**
  *  @note This file is part of Empirical, https://github.com/devosoft/Empirical
  *  @copyright Copyright (C) Michigan State University, MIT Software license; see doc/LICENSE.md
- *  @date 2017-2018.
+ *  @date 2017-2024.
  *
  *  @file StateGrid.hpp
  *  @brief StateGrid maintains a rectilinear grid that agents can traverse.
@@ -43,7 +43,7 @@ namespace emp {
     struct StateInfo {
       int state_id;        ///< Ordinal id for this state.
       char symbol;         ///< Symbol for printing this state.
-      double score_change; ///< Change ammount for organism score by stepping on this square.
+      double score_change; ///< Change amount for organism score by stepping on this square.
       std::string name;    ///< Name of this state.
       std::string desc;    ///< Explanation of this state.
 
@@ -64,32 +64,26 @@ namespace emp {
     std::map<char, size_t> symbol_map;       ///< Map of symbols to associated key ID
     std::map<std::string, size_t> name_map;  ///< Map of names to associated key ID
 
-    size_t GetKey(int state_id) const { return Find(state_map, state_id, 0); }
-    size_t GetKey(char symbol) const { return Find(symbol_map, symbol, 0); }
-    size_t GetKey(const std::string & name) const { return Find(name_map, name, 0); }
+    [[nodiscard]] size_t GetKey(int state_id) const { return Find(state_map, state_id, 0); }
+    [[nodiscard]] size_t GetKey(char symbol) const { return Find(symbol_map, symbol, 0); }
+    [[nodiscard]] size_t GetKey(const std::string & name) const { return Find(name_map, name, 0); }
   public:
-    StateGridInfo() : states(), state_map(), symbol_map(), name_map() { ; }
-    StateGridInfo(const StateGridInfo &) = default;
-    StateGridInfo(StateGridInfo &&) = default;
-    ~StateGridInfo() { ; }
-
-    StateGridInfo & operator=(const StateGridInfo &) = default;
-    StateGridInfo & operator=(StateGridInfo &&) = default;
-
     size_t GetNumStates() const { return states.size(); }
 
     // Convert from state ids...
-    char GetSymbol(int state_id) const { return states[ GetKey(state_id) ].symbol; }
-    double GetScoreChange(int state_id) const { return states[ GetKey(state_id) ].score_change; }
-    const std::string & GetName(int state_id) const { return states[ GetKey(state_id) ].name; }
-    const std::string & GetDesc(int state_id) const { return states[ GetKey(state_id) ].desc; }
+    [[nodiscard]] char GetSymbol(int state_id) const { return states[ GetKey(state_id) ].symbol; }
+    [[nodiscard]] double GetScoreChange(int state_id) const { return states[ GetKey(state_id) ].score_change; }
+    [[nodiscard]] const std::string & GetName(int state_id) const { return states[ GetKey(state_id) ].name; }
+    [[nodiscard]] const std::string & GetDesc(int state_id) const { return states[ GetKey(state_id) ].desc; }
 
     // Convert to state ids...
-    int GetState(char symbol) const {
+    [[nodiscard]] int GetState(char symbol) const {
       emp_assert( states.size() > GetKey(symbol), states.size(), symbol, (int) symbol );
       return states[ GetKey(symbol) ].state_id;
     }
-    int GetState(const std::string & name) const { return states[ GetKey(name) ].state_id; }
+    [[nodiscard]] int GetState(const std::string & name) const {
+      return states[ GetKey(name) ].state_id;
+    }
 
[…]
