### 2024-04-13 1e9d356
Clean up on Systematics.hpp
