+++ include/emp/Evolve/Systematics.hpp
@@ -63,18 +63,18 @@ namespace emp {
     }; /// The default - an empty struct
 
     struct fitness {
-        using has_fitness_t = std::true_type;
-        using has_mutations_t = std::false_type;
-        using has_phen_t = std::false_type;
+      using has_fitness_t = std::true_type;
+      using has_mutations_t = std::false_type;
+      using has_phen_t = std::false_type;
 
-        DataNode<double, data::Current, data::Range> fitness; /// This taxon's fitness (for assessing deleterious mutational steps)
-        void RecordFitness(double fit) {
-          fitness.Add(fit);
-        }
+      DataNode<double, data::Current, data::Range> fitness; /// This taxon's fitness (for assessing deleterious mutational steps)
+      void RecordFitness(double fit) {
+        fitness.Add(fit);
+      }
 
-        double GetFitness() const {
-          return fitness.GetMean();
-        }
+      [[nodiscard]] double GetFitness() const {
+        return fitness.GetMean();
+      }
     };
 
     /// Track information related to the mutational landscape
@@ -98,7 +98,7 @@ namespace emp {
       }
 
       /// @returns this taxon's fitness
-      double GetFitness() const {
+      [[nodiscard]] double GetFitness() const {
         return fitness.GetMean();
       }
 
@@ -171,52 +171,52 @@ namespace emp {
     Taxon & operator=(const Taxon &) = delete;
     Taxon & operator=(Taxon &&) = default;
 
-    bool operator<(const Taxon & other) const {
+    [[nodiscard]] bool operator<(const Taxon & other) const {
       return id < other.GetID();
     }
 
     /// Get a unique ID for this taxon; IDs are assigned sequentially, so newer taxa have higher IDs.
-    size_t GetID() const { return id; }
+    [[nodiscard]] size_t GetID() const { return id; }
 
     /// Retrieve the tracked info associated with this Taxon.
-    const info_t & GetInfo() const { return info; }
+    [[nodiscard]] const info_t & GetInfo() const { return info; }
 
     /// Retrieve a pointer to the parent Taxon.
-    Ptr<this_t> GetParent() const { return parent; }
+    [[nodiscard]] Ptr<this_t> GetParent() const { return parent; }
     void NullifyParent() {parent = nullptr;}
 
     /// Get the number of living organisms currently associated with this Taxon.
-    size_t GetNumOrgs() const { return num_orgs; }
+    [[nodiscard]] size_t GetNumOrgs() const { return num_orgs; }
 
     /// Get the total number of organisms that have ever lived associated with this Taxon
-    size_t GetTotOrgs() const { return tot_orgs; }
+    [[nodiscard]] size_t GetTotOrgs() const { return tot_orgs; }
 
     /// Get the number of taxa that were produced by organisms from this Taxon.
-    size_t GetNumOff() const { return num_offspring; }
+    [[nodiscard]] size_t GetNumOff() const { return num_offspring; }
 
     /// Get the number of taxanomic steps since the ancestral organism was injected into the World.
-    size_t GetDepth() const { return depth; }
+    [[nodiscard]] size_t GetDepth() const { return depth; }
 
     /// Get data struct associated with this taxon
-    data_t & GetData() {return data;}
+    [[nodiscard]] data_t & GetData() { return data; }
     /// Get data struct associated with this taxon
-    const data_t & GetData() const {return data;}
+    [[nodiscard]] const data_t & GetData() const { return data; }
 
     /// Get pointers to this taxon's offspring
-    std::set<Ptr<this_t> > GetOffspring() {return offspring;}
+    [[nodiscard]] std::set<Ptr<this_t> > GetOffspring() { return offspring; }
 
     /// Set this taxon's data struct to the given value
[…]
