### 2024-04-13 #6617 @editorialbot
[REVIEW]: Empirical: A scientific software library for research, education, and public engagement
**Submitting author:** @mmore500 (<a href="http://orcid.org/0000-0003-4726-4479">Matthew Andres Moreno</a>)
**Repository:** https://github.com/devosoft/Empirical/
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.1.5
