### 2024-04-14 f91de08
Removed constructors, destructors, and assignment operators from reference_vector (all were default)
