### 2024-04-15 1e06693
Cleaned up Distribution.hpp
