### 2024-04-16 a108a66
Fixed concepts hasToString and hasToDouble
