### 2024-04-17 9da6616
Cleaned up concepts.hpp
