+++ include/emp/meta/TypeID.hpp
@@ -107,7 +107,7 @@ namespace emp {
       bool init = false;                     ///< Has this info been initialized yet?
       std::string name = "[unknown type]";   ///< Unique (ideally human-readable) type name
 
-      virtual ~Info() { }
+      virtual ~Info() = default;
 
       virtual bool IsAbstract() const { return false; }
       virtual bool IsArithmetic() const { return false; }
@@ -340,65 +340,72 @@ namespace emp {
     ~TypeID() { ; }
     TypeID & operator=(const TypeID &) = default;
 
-    operator size_t() const noexcept { return (info_ptr->init) ? (size_t) info_ptr.Raw() : 0; }
-    operator bool() const noexcept { return info_ptr->init; }
-    bool operator==(TypeID in) const { return info_ptr == in.info_ptr; }
-    bool operator!=(TypeID in) const { return info_ptr != in.info_ptr; }
-    bool operator< (TypeID in) const { return info_ptr < in.info_ptr; }
-    bool operator<=(TypeID in) const { return info_ptr <=in.info_ptr; }
-    bool operator> (TypeID in) const { return info_ptr > in.info_ptr; }
-    bool operator>=(TypeID in) const { return info_ptr >= in.info_ptr; }
+    [[nodiscard]] operator size_t() const noexcept {
+      return (info_ptr->init) ? (size_t) info_ptr.Raw() : 0;
+    }
+    [[nodiscard]] operator bool() const noexcept { return info_ptr->init; }
+    [[nodiscard]] bool operator==(TypeID in) const { return info_ptr == in.info_ptr; }
+    [[nodiscard]] bool operator!=(TypeID in) const { return info_ptr != in.info_ptr; }
+    [[nodiscard]] bool operator< (TypeID in) const { return info_ptr < in.info_ptr; }
+    [[nodiscard]] bool operator<=(TypeID in) const { return info_ptr <=in.info_ptr; }
+    [[nodiscard]] bool operator> (TypeID in) const { return info_ptr > in.info_ptr; }
+    [[nodiscard]] bool operator>=(TypeID in) const { return info_ptr >= in.info_ptr; }
 
     /// Get a unique numerical ID for this TypeID object.
-    size_t GetID() const { return (size_t) info_ptr.Raw(); }
+    [[nodiscard]] size_t GetID() const { return (size_t) info_ptr.Raw(); }
 
     /// Get the name generated for this TypeID (ideally unique)
-    const std::string & GetName() const { return info_ptr->name; }
+    [[nodiscard]] const std::string & GetName() const { return info_ptr->name; }
 
     /// Update the name for ALL instances of this TypeID.
     void SetName(const std::string & in_name) { emp_assert(info_ptr); info_ptr->name = in_name; }
 
-    bool IsInitialized() const { return info_ptr->init; }
+    [[nodiscard]] bool IsInitialized() const { return info_ptr->init; }
     void SetInitialized(bool _in=true) { info_ptr->init = _in; }
 
-    bool IsAbstract() const { return info_ptr->IsAbstract(); }
-    bool IsArithmetic() const { return info_ptr->IsArithmetic(); }
-    bool IsArray() const { return info_ptr->IsArray(); }
-    bool IsClass() const { return info_ptr->IsClass(); }
-    bool IsConst() const { return info_ptr->IsConst(); }
-    bool IsEmpty() const { return info_ptr->IsEmpty(); }
-    bool IsObject() const { return info_ptr->IsObject(); }
[…]
