### 2024-04-17 devosoft/Empirical#509 @LTLA
Some of the quick start instructions need updating
(Part of openjournals/joss-reviews#6617)

The [Hello World example](https://empirical.readthedocs.io/en/latest/QuickStartGuides/1-HelloWorld.html) requires the submodules in `third-party` to be fetched. This is not done automatically on the initial `git clone`, at least not on my version of Git.

The header file names in [Base tools](https://empirical.readthedocs.io/en/latest/QuickStartGuides/2-BaseTools.html) are not correct (`.h` instead of `.hpp`).

And unrelated, but the title [here](https://empirical.readthedocs.io/en/latest/library/web/d3/d3-intro.html) has a typo ("Empirica's").
