### 2024-04-17 devosoft/Empirical#510 @LTLA
How smart is `Ptr`?
(Part of openjournals/joss-reviews#6617)

This [section of the documentation](https://empirical.readthedocs.io/en/latest/QuickStartGuides/2-BaseTools.html#base-ptr-h) suggests that `Ptr` is not smart, i.e., does not delete its contents upon destruction. I also had a look at the code and it seems that `Ptr` destruction triggers a decrement but not any deletion; an explicit call to `Delete()` is required.

This is fine if `Ptr` is intended as an exact replacement for a raw pointer (perhaps the docs could be more explicit about that). But it does seem like a missed opportunity for this library. Much of "modern C++" is moving towards smart pointers due to the ease of automatic memory management, and so it is unfortunate that `emp::Ptr` does not provide that.

On a related note, it appears that `Ptr` construction may not be thread-safe, due to the class-global `PtrTracker` that tracks all pointers. I don't see any locks to protect modifications for `New()`, etc., which may cause problems when used with OpenMP or by (Emscripten) PThreads. By comparison, `std::shared_ptr` operations are [thread-safe](https://en.cppreference.com/w/cpp/memory/shared_ptr).
