### 2024-04-17 devosoft/Empirical#511 @LTLA
Why rely on Jquery?
(Part of openjournals/joss-reviews#6617)

It seems like some of the C++ calls are converted into Javascript code that uses Jquery (see the various `$(...)` calls in `include/emp/web`). Is this really necessary? It adds a Jquery dependency to all downstream web applications that might not otherwise use it. From what I see, many of the Jquery calls in this library involve manipulating with DOM elements, so why not just use vanilla javascript (e.g., `document.getElementById`)?

In any case, if you really do need Jquery, this should be explicitly documented somewhere. I didn't see it discussed in any obvious place in the documentation.
