### 2024-04-17 #6617 @LTLA
Alright, apologies for the delay. Here's my review:

This is a very interesting library. I have long wondered how to manipulate HTML from within C++, and Empirical provides an ergonomic way of doing so. It offers the possibility of writing web applications almost completely in C++, without any need to touch Javascript or associated frameworks. I can forsee this being useful for scientific programmers who want to build a simple application based on C++ tools without requiring any web development knowledge. In this respect, Empirical targets a similar audience to Shiny (R) or dash (Python), but for C++ developers. That said, more experienced web developers will probably not use these features, as direct DOM manipulation is discouraged in frameworks like React.

The extra debugging information for the standard data structures is nice, especially for WebAssembly where debugging segfaults is even more tedious than useful. Perhaps the authors might consider checking whether a macro could be used to seamlessly switch between `emp::` and `std::`. This would allow library developers to use Empirical for debug builds, but then switch back to the STL for release, in order to reduce dependencies for their own users. For example, if I was developing a header-only library that uses Empirical for debugging, I would prefer to avoid mandating Empirical as a compile-time dependency for all applications that use my library.

The paper itself is well-written and concise. Possibly too concise, actually: it might have benefited from one or two short code examples to demonstrate the ease-of-use of the UI manipulation, bounds-checking, etc. Then readers can get a better grasp of what the library actually does, before deciding whether to proceed to the documentation itself.

On some other minor points:

- The `README.md` could be fleshed out with more motivation, e.g., by transplanting the statement of need from the paper. When I hit the repo landing page, I wasn't sure what this library was about, and when I trawled the codebase, I spent a while thinking that it was dedicated to evolution/phylo-related programming. Most people won't think that the interesting stuff is in `base` or `web` when they can instead be looking at a directory with an exciting name like `Evolve`.
- Some instructions about the dependencies would be helpful during installation, see my comments about third-party requirements and jQuery.

Related issues:

- devosoft/Empirical#509
- devosoft/Empirical#510
- devosoft/Empirical#511
