### 2024-04-18 devosoft/Empirical#511 @mercere99
You are correct -- Jquery seemed like it was going to be a big help when we first started using it in Empirical, but none of the functionality it provides is critical anymore, especially since we are working with auto-generated JavaScript code.  We have planned to remove it for a while now and need to escalate the priority of that improvement to reduce the dependencies.
