### 2024-04-18 3d7f360
Strip JQuery content
