+++ include/emp/web/Style.hpp
@@ -112,7 +112,7 @@ namespace web {
 #ifdef __EMSCRIPTEN__
       MAIN_THREAD_EM_ASM({
           var id = UTF8ToString($0);
-          emp_i.cur_obj = $( '#' + id );
+          emp_i.cur_obj = document.getElementById(id);
         }, widget_id.c_str());
 #endif
 
@@ -154,7 +154,7 @@ namespace web {
           var id = UTF8ToString($0);
           var setting = UTF8ToString($1);
           var value = UTF8ToString($2);
-          $( '#' + id ).css( setting, value);
+          document.getElementById(id).style[setting] = value;
         }, widget_id.c_str(), setting.c_str(), settings[setting].c_str());
 #else
       std::cout << "Setting '" << widget_id << "' attribute '" << setting
@@ -170,7 +170,7 @@ namespace web {
           var id = UTF8ToString($0);
           var setting = UTF8ToString($1);
           var value = UTF8ToString($2);
-          $( '#' + id ).css( setting, value);
+          document.getElementById(id).style[setting] = value;
         }, widget_id.c_str(), setting.c_str(), value.c_str());
 #else
       std::cout << "Setting '" << widget_id << "' attribute '" << setting
@@ -184,7 +184,7 @@ namespace web {
       EM_ASM_ARGS({
           var id = UTF8ToString($0);
           var name = UTF8ToString($1);
-          $( '#' + id ).addClass( name);
+          document.getElementById(id).classList.add(name);
         }, widget_id.c_str(), clss.c_str());
 #else
       std::cout << "Adding class to '" << widget_id << "': '" << clss;
@@ -196,7 +196,7 @@ namespace web {
       EM_ASM_ARGS({
           var id = UTF8ToString($0);
           var name = UTF8ToString($1);
-          $( '#' + id ).removeClass( name);
+          document.getElementById(id).classList.remove(name);
         }, widget_id.c_str(), clss.c_str());
 #else
       std::cout << "Adding class to '" << widget_id << "': '" << clss;
