### 2024-04-18 38919d6
Fixup replacement of jquery ready()
