### 2024-04-18 b8f813c
Replace jQuery .attr() with .setAttribute()
