### 2024-04-18 a7943c7
Replace call to JQuery .css()
