### 2024-04-18 f6ec2b9
Minor cleanups in a few files.
