+++ tests/meta/FunInfo.cpp
@@ -6,8 +6,6 @@
  *  @file FunInfo.cpp
  */
 
-#define CATCH_CONFIG_MAIN
-
 #include "third-party/Catch/single_include/catch2/catch.hpp"
 
 #include <sstream>
