+++ include/emp/io/StreamManager.hpp
@@ -50,7 +50,7 @@ namespace emp {
     public:
       StreamInfo(const std::string & in_name, bool in_owned) : name(in_name), owned(in_owned) { }
       StreamInfo(const StreamInfo &) = delete;
-      virtual ~StreamInfo() {}
+      virtual ~StreamInfo() = default;
 
       virtual Type GetType() const = 0;
       virtual Access GetAccess() const = 0;
@@ -188,7 +188,7 @@ namespace emp {
     }
 
   public:
-    StreamManager() { }
+    StreamManager() = default;
     StreamManager(const StreamManager &) = delete;
     // StreamManager(StreamManager &&) = default;
     ~StreamManager() {
