### 2024-04-18 71f2b87
Check for null before setting text
