### 2024-04-18 405faf9
Test for null when calling getElement
