+++ include/emp/web/Style.hpp
@@ -154,7 +154,8 @@ namespace web {
           var id = UTF8ToString($0);
           var setting = UTF8ToString($1);
           var value = UTF8ToString($2);
-          document.getElementById(id).style[setting] = value;
+          var element = document.getElementById(id);
+          if (element) element.style[setting] = value;
         }, widget_id.c_str(), setting.c_str(), settings[setting].c_str());
 #else
       std::cout << "Setting '" << widget_id << "' attribute '" << setting
@@ -170,7 +171,8 @@ namespace web {
           var id = UTF8ToString($0);
           var setting = UTF8ToString($1);
           var value = UTF8ToString($2);
-          document.getElementById(id).style[setting] = value;
+          var element = document.getElementById(id);
+          if (element) element.style[setting] = value;
         }, widget_id.c_str(), setting.c_str(), value.c_str());
 #else
       std::cout << "Setting '" << widget_id << "' attribute '" << setting
@@ -184,7 +186,8 @@ namespace web {
       EM_ASM_ARGS({
           var id = UTF8ToString($0);
           var name = UTF8ToString($1);
-          document.getElementById(id).classList.add(name);
+          var element = document.getElementById(id);
+          if (element) element.classList.add(name);
         }, widget_id.c_str(), clss.c_str());
 #else
       std::cout << "Adding class to '" << widget_id << "': '" << clss;
@@ -196,7 +199,8 @@ namespace web {
       EM_ASM_ARGS({
           var id = UTF8ToString($0);
           var name = UTF8ToString($1);
-          document.getElementById(id).classList.remove(name);
+          var element = document.getElementById(id);
+          if (element) element.classList.remove(name);
         }, widget_id.c_str(), clss.c_str());
 #else
       std::cout << "Adding class to '" << widget_id << "': '" << clss;
