+++ include/emp/web/Attributes.hpp
@@ -145,7 +145,7 @@ namespace web {
 #endif
     }
 
-    /// Apply onlay a SPECIFIC attributes setting with a specifid value!
+    /// Apply only a SPECIFIC attributes setting with a specifid value!
     static void Apply(const std::string & widget_id, const std::string & setting,
                       const std::string & value) {
 #ifdef __EMSCRIPTEN__
