### 2024-04-18 c8d4b81
Test for null before doing JS ops
