+++ include/emp/web/Widget.hpp
@@ -384,7 +384,8 @@ namespace web {
         MAIN_THREAD_EM_ASM({
             var widget_id = UTF8ToString($0);
             var out_html = UTF8ToString($1);
-            document.getElementById(widget_id).outerHTML = out_html;
+            var element = document.getElementById(widget_id);
+            if (element) element.outerHTML = out_html;
           }, id.c_str(), ss.str().c_str());
 
         // If active update style, trigger JS, and recurse to children!
@@ -498,7 +499,9 @@ namespace web {
     emp_assert(GetID() != "");  // Must have a name!
     return MAIN_THREAD_EM_ASM_DOUBLE({
       var id = UTF8ToString($0);
-      return document.getElementById(id).clientWidth;
+      var element = document.getElementById(id);
+      if (!element) return -1.0; // Check if element exists
+      return element.clientWidth;
     }, GetID().c_str());
   }
   double Widget::GetHeight(){
@@ -506,7 +509,9 @@ namespace web {
     emp_assert(GetID() != "");  // Must have a name!
     return MAIN_THREAD_EM_ASM_DOUBLE({
       var id = UTF8ToString($0);
-      return document.getElementById(id).clientHeight;
+      var element = document.getElementById(id);
+      if (!element) return -1.0; // Check if element exists
+      return element.clientHeight;
     }, GetID().c_str());
   }
   double Widget::GetInnerWidth(){
@@ -544,7 +549,9 @@ namespace web {
     emp_assert(GetID() != "");  // Must have a name!
     return MAIN_THREAD_EM_ASM_DOUBLE({
       var id = UTF8ToString($0);
-      return document.getElementById(id).offsetWidth;
+      var element = document.getElementById(id);
+      if (!element) return 0.0; // Check if element exists
+      return element.offsetWidth;
     }, GetID().c_str());
   }
   double Widget::GetOuterHeight(){
@@ -552,7 +559,9 @@ namespace web {
     emp_assert(GetID() != "");  // Must have a name!
     return MAIN_THREAD_EM_ASM_DOUBLE({
       var id = UTF8ToString($0);
-      return document.getElementById(id).offsetHeight;
+      var element = document.getElementById(id);
+      if (!element) return 0.0; // Check if element exists
+      return element.offsetHeight;
     }, GetID().c_str());
   }
 
