### 2024-04-18 66b1fa3
Replace auto in sig with template type
