### 2024-04-18 864602e
Revert "Replace auto in sig with template type"
This reverts commit 66b1fa35177145ce3b3a032aacc2bcd79f586f64.
