### 2024-04-18 56b9d68
Fix macro syntax warning
