### 2024-04-18 3b75ad8
Test js emp_i.cur_obj before using
