### 2024-04-18 devosoft/Empirical#509 @mmore500
Thanks for bringing these issues to our attention!

Just opened #514 with fixes for the three issues mentioned.
When it merges, this issue will get closed.
However, feel free to re-open if you find more lint or the issues are not adequately addressed :+1:
