### 2024-04-18 b3a0f1a
Fixup tests, allow some jquery for prefab
NOTE: disables some tests that are ok upon manual inspection
