### 2024-04-19 #6617 @mmore500
> Heya, I've started my review in between some teaching obligations. I generally dislike debugging paths and compiling C code, so I'll leave the problem-solving up to the authors:
>
> In following the installation instructions on Mac OSX (Ventura 13.2.1) I get the following missing file:
>
> `Makefile:22: WARNING: the Cookiecutter Empirical Project, which you can find at https://github.com/devosoft/cookiecutter-empirical-project, should be preferred over ProjectTemplate g++ -O3 -DNDEBUG -Wall -Wno-unused-function -std=c++20 -I../../../Empirical/include/ source/native/project_name.cpp -o project_name In file included from source/native/project_name.cpp:21: In file included from ../../../Empirical/include/emp/config/command_line.hpp:42: In file included from ../../../Empirical/include/emp/config/../tools/string_utils.hpp:35: ../../../Empirical/include/emp/config/../tools/../base/array.hpp:28:10: fatal error: '../../../third-party/cereal/include/cereal/cereal.hpp' file not found #include "../../../third-party/cereal/include/cereal/cereal.hpp" ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ 1 error generated. make: *** [project_name] Error 1`

Hi @bramvandijk88

Thanks for bringing this to our attention! We've updated the documentation to clarify a missing step of initializing the project subrepositories. If you do

`git submodule update --init --recursive --depth 1` and then try compiling again it should hopefully work!
