### 2024-04-20 b47f624
Cleanup on Text.cpp
