### 2024-04-20 5f2f488
Cleanup on Text_Rule.hpp
