+++ tests/base/notify.cpp
@@ -6,8 +6,6 @@
  *  @file notify.cpp
  */
 
-#define CATCH_CONFIG_MAIN
-
 #include "third-party/Catch/single_include/catch2/catch.hpp"
 
 #include "emp/base/notify.hpp"
