### 2024-04-21 96943ae
Updates on various Text converters.
