### 2024-04-22 #6617 @bramvandijk88
> If you do `git submodule update --init --recursive --depth 1` and then try compiling again it should hopefully work!

Installation was running for ~ 10 minutes until I ran into this:

```
rm -rf test*.out
cd hardware && make test
mkdir -p temp
c++ -std=c++20 -g -pthread -Wall -Wno-unused-function -Wno-unused-private-field -I../../include/ -I../../ -I../../third-party/cereal/include/ -DCATCH_CONFIG_MAIN event_driven_gp.cpp -o test-event_driven_gp.out
In file included from event_driven_gp.cpp:25:
In file included from ../../include/emp/hardware/EventDrivenGP.hpp:36:
In file included from ../../include/emp/hardware/../matching/MatchBin.hpp:36:
../../include/emp/hardware/../matching/../../../third-party/robin-hood-hashing/src/include/robin_hood.h:54:14: fatal error: 'sys/auxv.h' file not found
#    include <sys/auxv.h> // for getauxval
             ^~~~~~~~~~~~
1 error generated.
make[2]: *** [test-event_driven_gp] Error 1
make[1]: *** [test-hardware] Error 2
make: *** [test-native-regular] Error 2
```
