+++ include/emp/config/FlagManager.hpp
@@ -1,8 +1,9 @@
 /**
- *  @note This file is part of Empirical, https://github.com/devosoft/Empirical
- *  @copyright Copyright (C) Michigan State University, MIT Software license; see doc/LICENSE.md
- *  @date 2023.
- *
+ *  This file is part of Empirical, https://github.com/devosoft/Empirical
+ *  Copyright (C) Michigan State University, MIT Software license; see doc/LICENSE.md
+ *  date: 2023-2024
+ */
+/**
  *  @file FlagManager.hpp
  *  @brief This file contains tools for dealing with command-line flags (from argv and argc).
  *  @note Status: ALPHA
