+++ include/emp/config/SettingConfig.hpp
@@ -1,7 +1,7 @@
 /*
  *  This file is part of Empirical, https://github.com/devosoft/Empirical
  *  Copyright (C) Michigan State University, MIT Software license; see doc/LICENSE.md
- *  date: 2020
+ *  date: 2020-2024
 */
 /**
  *  @file
