### 2024-04-24 18db102
Cleaned up _is_streamable.hpp to use new concepts.
