### 2024-04-25 a6129e5
Updated Systematics.hpp to use new is_streamable
