### 2024-04-27 3954cde
Updated SimpleParser to check for either emp::vector of std::vector when dealing with incoming values.
