### 2024-04-28 4afb181
Added emp::String version of command-line helper functions.
