+++ include/emp/config/command_line.hpp
@@ -40,6 +40,7 @@
 
 #include "../base/vector.hpp"
 #include "../tools/string_utils.hpp"
+#include "../tools/String.hpp"
 
 namespace emp {
   namespace cl {
@@ -47,6 +48,15 @@ namespace emp {
     // Convert input arguments to a vector of strings for easier processing.
     emp::vector<std::string> args_to_strings(int argc, char* argv[]) {
       emp::vector<std::string> args;
+      for (size_t i = 0; i < (size_t) argc; i++) {
+        args.push_back(argv[i]);
+      }
+      return args;
+    }
+
+    // Convert input arguments to a vector of emp::String for easier processing.
+    emp::vector<emp::String> ArgsToStrings(int argc, char* argv[]) {
+      emp::vector<emp::String> args;
       for (size_t i = 0; i < (size_t) argc; i++) {
         args.push_back(argv[i]);
       }
@@ -68,14 +78,34 @@ namespace emp {
       return -1;
     }
 
+    // Search through args to find a specific value.
+    int FindArg(const emp::vector<emp::String> & args, const emp::String & pattern) {
+      for (size_t i = 0; i < args.size(); i++) {
+        if (args[i] == pattern) return (int) i;
+      }
+      return -1;
+    }
+
     // Return true/false if a specific argument is present.
     bool has_arg(const emp::vector<std::string> & args, const std::string & pattern) {
       return (find_arg(args, pattern) != -1);
     }
 
+    // Return true/false if a specific argument is present.
+    bool HasArg(const emp::vector<emp::String> & args, const emp::String & pattern) {
+      return (FindArg(args, pattern) != -1);
+    }
+
     // Return true/false if a specific argument is present and REMOVE IT.
     bool use_arg(emp::vector<std::string> & args, const std::string & pattern) {
       const int pos = find_arg(args, pattern);
+      if (pos >= 0) args.erase(args.begin()+pos);
+      return (pos != -1);
+    }
+
+    // Return true/false if a specific argument is present and REMOVE IT.
+    bool UseArg(emp::vector<emp::String> & args, const emp::String & pattern) {
+      const int pos = FindArg(args, pattern);
       if (pos >= 0) args.erase(args.begin()+pos);
       return (pos != -1);
     }
