### 2024-04-29 da59130
Added simple AsDouble(), AsInt(), and AsULL() member functions to emp::String
