### 2024-05-06 3ddc048
Bump jinja2 from 3.1.2 to 3.1.4 in /doc
Bumps [jinja2](https://github.com/pallets/jinja) from 3.1.2 to 3.1.4.
- [Release notes](https://github.com/pallets/jinja/releases)
- [Changelog](https://github.com/pallets/jinja/blob/main/CHANGES.rst)
- [Commits](https://github.com/pallets/jinja/compare/3.1.2...3.1.4)

---
updated-dependencies:
- dependency-name: jinja2
  dependency-type: indirect
...

Signed-off-by: dependabot[bot] <support@github.com>
