### 2024-05-07 devosoft/Empirical#510 @LTLA
Makes sense to me. Feel free to close, unless you want to keep it open to track the thread safety issue.
