### 2024-05-08 #6617 @mmore500
Pinging my co-maintainers and try to get a review on the de-jquerifyication PR that's open. Hopefully progress on that soon.
