+++ doc/requirements.txt
@@ -11,16 +11,15 @@ babel==2.13.1
 beautifulsoup4==4.12.2
     # via sphinx-tippy
 breathe==4.35.0
-    # via -r /tmp/requirements.in
+    # via -r requirements.in
 certifi==2023.11.17
     # via requests
 charset-normalizer==3.3.2
     # via requests
 coverxygen==1.8.1
-    # via -r /tmp/requirements.in
+    # via -r requirements.in
 docutils==0.20.1
     # via
-    #   -r /tmp/requirements.in
     #   breathe
     #   myst-parser
     #   pybtex-docutils
@@ -49,7 +48,7 @@ mdit-py-plugins==0.4.0
 mdurl==0.1.2
     # via markdown-it-py
 myst-parser==2.0.0
-    # via -r /tmp/requirements.in
+    # via -r requirements.in
 packaging==23.2
     # via sphinx
 pybtex==0.24.0
@@ -64,7 +63,7 @@ pyyaml==6.0.1
     # via
     #   myst-parser
     #   pybtex
-requests==2.31.0
+requests==2.32.0
     # via
     #   sphinx
     #   sphinx-tippy
@@ -78,7 +77,7 @@ soupsieve==2.5
     # via beautifulsoup4
 sphinx==7.2.6
     # via
-    #   -r /tmp/requirements.in
+    #   -r requirements.in
     #   breathe
     #   myst-parser
     #   sphinx-rtd-theme
@@ -91,13 +90,13 @@ sphinx==7.2.6
     #   sphinxcontrib-qthelp
     #   sphinxcontrib-serializinghtml
 sphinx-rtd-theme==2.0.0
-    # via -r /tmp/requirements.in
+    # via -r requirements.in
 sphinx-tippy==0.4.1
-    # via -r /tmp/requirements.in
+    # via -r requirements.in
 sphinxcontrib-applehelp==1.0.7
     # via sphinx
 sphinxcontrib-bibtex==2.6.1
-    # via -r /tmp/requirements.in
+    # via -r requirements.in
 sphinxcontrib-devhelp==1.0.5
     # via sphinx
 sphinxcontrib-htmlhelp==2.0.4
