### 2024-05-21 9d22e1e
Set up assert_print_first_opt to be static to prevent link errors.
