+++ include/emp/games/SudokuAnalyzer.hpp
@@ -1,0 +1,659 @@
+/**
+ *  @note This file is part of MABE, https://github.com/mercere99/MABE2
+ *  @copyright Copyright (C) Michigan State University, MIT Software license; see doc/LICENSE.md
+ *  @date 2024.
+ *
+ *  @file
+ *  @brief  Analyzes a sudoku instance to determine the solving experience for a human player.
+ *
+ *  DEVELOPER NOTES:
+ *   For the moment, we will assume that all boards are 9x9 with a standard Sudoku layout.
+ *   In the future, we should make the board configuration more flexible.
+*/
+
+#ifndef EMP_GAMES_SUDOKU_ANALYZER_HPP
+#define EMP_GAMES_SUDOKU_ANALYZER_HPP
+
+#include <array>
+#include <fstream>
+#include <istream>
+#include <set>
+#include <vector>
+
+#include "emp/base/array.hpp"
+#include "emp/base/assert.hpp"
+#include "emp/math/Random.hpp"
+#include "emp/tools/String.hpp"
+
+#include "Puzzle.hpp"
+
+namespace emp {
+
+  class SudokuAnalyzer {
+  private:
+    static constexpr size_t NUM_STATES = 9;
+    static constexpr size_t NUM_ROWS = 9;
+    static constexpr size_t NUM_COLS = 9;
+    static constexpr size_t NUM_SQUARES = 9;
+    static constexpr size_t NUM_OVERLAPS = 54;      // Multi-cell overlaps between two regions
+    static constexpr size_t NUM_CELLS = NUM_ROWS * NUM_COLS;                  // 81
+    static constexpr size_t NUM_REGIONS = NUM_ROWS + NUM_COLS + NUM_SQUARES;  // 27
+    static constexpr size_t REGIONS_PER_CELL = 3;   // Each cell is part of three regions.
+    
+    static constexpr size_t UNKNOWN_STATE = NUM_STATES; // Lower values are actual states.
+
+    emp::array<char,NUM_CELLS> value;         // Known value for cells
+    emp::array<uint32_t, NUM_CELLS> options;  // Options still available to each cell
+
+    // "members" tracks which cell ids are members of each region.
+    static constexpr int members[NUM_REGIONS][NUM_STATES] = {
+      // Rows (Overlaps: 111000000, 000111000, 000000111)
+      {  0,  1,  2,  3,  4,  5,  6,  7,  8 },  // Region 0  (Overlap with 18, 19, 20)
+      {  9, 10, 11, 12, 13, 14, 15, 16, 17 },  // Region 1  (Overlap with 18, 19, 20)
[…]
