### 2024-05-22 e461683
Fixed some types and updated functionality in SudokuAnalyzer.hpp
