+++ include/emp/games/SudokuAnalyzer.hpp
@@ -42,11 +42,14 @@ namespace emp {
     
     static constexpr size_t UNKNOWN_STATE = NUM_STATES; // Lower values are actual states.
 
+    // Which symbols are we using in this puzzle?
+    emp::array<char, NUM_STATES> symbols = {'1', '2', '3', '4', '5', '6', '7', '8', '9'};
+
     emp::array<char,NUM_CELLS> value;         // Known value for cells
     emp::array<uint32_t, NUM_CELLS> options;  // Options still available to each cell
 
     // "members" tracks which cell ids are members of each region.
-    static constexpr int members[NUM_REGIONS][NUM_STATES] = {
+    static constexpr size_t members[NUM_REGIONS][NUM_STATES] = {
       // Rows (Overlaps: 111000000, 000111000, 000000111)
       {  0,  1,  2,  3,  4,  5,  6,  7,  8 },  // Region 0  (Overlap with 18, 19, 20)
       {  9, 10, 11, 12, 13, 14, 15, 16, 17 },  // Region 1  (Overlap with 18, 19, 20)
@@ -273,7 +276,7 @@ namespace emp {
     };
 
     // Which cells are in region overlaps (horizontal or vertical)?
-    static constexpr int overlaps[NUM_OVERLAPS][3] = {
+    static constexpr size_t overlaps[NUM_OVERLAPS][3] = {
       {  0,  1,  2 }, {  3,  4,  5 }, {  6,  7,  8 },  // Overlap sets  0,  1,  2 (Row 0)
       {  9, 10, 11 }, { 12, 13, 14 }, { 15, 16, 17 },  // Overlap sets  3,  4,  5 (Row 1)
       { 18, 19, 20 }, { 21, 22, 23 }, { 24, 25, 26 },  // Overlap sets  6,  7,  8 (Row 2)
@@ -296,7 +299,7 @@ namespace emp {
     };
 
     // Which overlaps are square regions made up from? (each square is listed twice).
-    static constexpr int square_overlaps[18][3] = {
+    static constexpr size_t square_overlaps[18][3] = {
       {  0,  3,  6 }, {  1,  4,  7 }, {  2,  5,  8 },
       {  9, 12, 15 }, { 10, 13, 16 }, { 11, 14, 17 },
       { 18, 21, 24 }, { 19, 22, 25 }, { 20, 23, 26 },
@@ -307,7 +310,7 @@ namespace emp {
     };
 
     // Which region is each overlap associated with? (raw row/col first, then square id)
-    static constexpr int overlap_regions[NUM_OVERLAPS][2] = {
+    static constexpr size_t overlap_regions[NUM_OVERLAPS][2] = {
       {  0, 0 }, {  0, 1 }, {  0, 2 },
       {  1, 0 }, {  1, 1 }, {  1, 2 },
       {  2, 0 }, {  2, 1 }, {  2, 2 },
@@ -329,7 +332,7 @@ namespace emp {
       { 17, 15 }, { 17, 16 }, { 17, 17 }
     };
 
-    // static constexpr int overlap_regions[NUM_OVERLAPS][2] = {
+    // static constexpr size_t overlap_regions[NUM_OVERLAPS][2] = {
     //   {  0, 18 }, {  0, 19 }, {  0, 20 },
     //   {  1, 18 }, {  1, 19 }, {  1, 20 },
     //   {  2, 18 }, {  2, 19 }, {  2, 20 },
@@ -358,19 +361,25 @@ namespace emp {
 
     SudokuAnalyzer& operator=(const SudokuAnalyzer &) = default;
 
-    int GetValue(int cell) const { return value[cell]; }
-    uint32_t GetOptions(int cell) const { return options[cell]; }
-    int CountOptions(int cell) const {
[…]
