### 2024-05-22 f0c53d8
Setup SudokuAnalyzer example to acutally load and solve a puzzle.
