### 2024-05-22 3165424
Updated member types in PuzzleMove
