### 2024-05-22 3f53539
Cleaned up lots of types in SudokuAnalyzer
