+++ include/emp/games/SudokuAnalyzer.hpp
@@ -40,12 +40,12 @@ namespace emp {
     static constexpr size_t NUM_REGIONS = NUM_ROWS + NUM_COLS + NUM_SQUARES;  // 27
     static constexpr size_t REGIONS_PER_CELL = 3;   // Each cell is part of three regions.
     
-    static constexpr size_t UNKNOWN_STATE = NUM_STATES; // Lower values are actual states.
+    static constexpr uint8_t UNKNOWN_STATE = NUM_STATES; // Lower values are actual states.
 
     // Which symbols are we using in this puzzle?
     emp::array<char, NUM_STATES> symbols = {'1', '2', '3', '4', '5', '6', '7', '8', '9'};
 
-    emp::array<char,NUM_CELLS> value;         // Known value for cells
+    emp::array<uint8_t,NUM_CELLS> value;         // Known value for cells
     emp::array<uint32_t, NUM_CELLS> options;  // Options still available to each cell
 
     // "members" tracks which cell ids are members of each region.
@@ -85,7 +85,7 @@ namespace emp {
     };
     
     // "regions" tracks which regions each cell is a member of.
-    static constexpr int regions[NUM_CELLS][REGIONS_PER_CELL] = {
+    static constexpr size_t regions[NUM_CELLS][REGIONS_PER_CELL] = {
       // { ROW, COLUMN, BOX }
       { 0,  9, 18 }, { 0, 10, 18 }, { 0, 11, 18 },  // Cells  0- 2
       { 0, 12, 19 }, { 0, 13, 19 }, { 0, 14, 19 },  // Cells  3- 5
@@ -119,7 +119,7 @@ namespace emp {
     };
 
     // Which *other* cells is each cell linked to by at least one region?
-    static constexpr int links[NUM_CELLS][20] = {
+    static constexpr size_t links[NUM_CELLS][20] = {
       { 1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 18, 19, 20, 27, 36, 45, 54, 63, 72 }, // Cell 0
       { 0,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 18, 19, 20, 28, 37, 46, 55, 64, 73 }, // Cell 1
       { 0,  1,  3,  4,  5,  6,  7,  8,  9, 10, 11, 18, 19, 20, 29, 38, 47, 56, 65, 74 },
@@ -204,8 +204,8 @@ namespace emp {
     };
     
     // Given a binary representation of options, which bit position is the first available?
-    static constexpr int next_opt[512] = {
-      -1, 0, 1, 0, 2, 0, 1, 0, 3, 0, 1, 0, 2, 0, 1, 0,
+    static constexpr uint8_t next_opt[512] = {
+      UNKNOWN_STATE, 0, 1, 0, 2, 0, 1, 0, 3, 0, 1, 0, 2, 0, 1, 0,
       4, 0, 1, 0, 2, 0, 1, 0, 3, 0, 1, 0, 2, 0, 1, 0,
       5, 0, 1, 0, 2, 0, 1, 0, 3, 0, 1, 0, 2, 0, 1, 0,
       4, 0, 1, 0, 2, 0, 1, 0, 3, 0, 1, 0, 2, 0, 1, 0,
@@ -240,7 +240,7 @@ namespace emp {
     };
 
     // Given a binary representation of state options for a cell, how many are available?
-    static constexpr int opts_count[512] = {
+    static constexpr size_t opts_count[512] = {
       0, 1, 1, 2, 1, 2, 2, 3, 1, 2, 2, 3, 2, 3, 3, 4,
       1, 2, 2, 3, 2, 3, 3, 4, 2, 3, 3, 4, 3, 4, 4, 5,
[…]
