### 2024-05-23 devosoft/Empirical#510 @mmore500
Closing this and opening an issue specifically on the thread safe issue 👍
