### 2024-05-23 #6617 @mmore500
Closed the last reviewer issue on the tracker! Added the remaining sub-issue of `emp::Ptr` debug mode thread safety to the medium-term roadmap, as the thread safety issue arises specifically in the context of debug-mode telemetry as opposed to user production code.

Let me know if there’s other action we should take on our end to progress this review process!
