### 2024-05-23 4477b3b
Added IsFixedSize() to Bits_Data to change Bits behavior based on this criterion
