### 2024-05-23 74e9096
Switch Bits to emp::String; unified String handling; added FindUniqueOnes() function
