+++ include/emp/bits/Bits.hpp
@@ -67,6 +67,7 @@
 #include "../math/Random.hpp"
 #include "../math/Range.hpp"
 #include "../meta/type_traits.hpp"
+#include "../tools/String.hpp"
 
 #include "Bits_Data.hpp"
 #include "_bitset_helpers.hpp"
@@ -188,11 +189,11 @@ namespace emp {
     template <size_t NUM_BITS>
     explicit Bits(const std::bitset<NUM_BITS> & bitset);
 
-    /// @brief Constructor to generate a Bits from a string of '0's and '1's.
-    Bits(const std::string & bitstring);
+    /// @brief Constructor to generate a Bits from a string of '0's and '1's or indices
+    Bits(const emp::String & bitstring) { Set(bitstring); }
 
     /// @brief Constructor to generate a Bits from a literal string of '0's and '1's.
-    Bits(const char * bitstring) : Bits(std::string(bitstring)) {}
+    Bits(const char * bitstring) : Bits(emp::String(bitstring)) {}
 
     /// @brief Constructor to generate a random set of bits in the default size.
     /// @param random Random number generator to use.
@@ -251,10 +252,10 @@ namespace emp {
     Bits & operator=(const std::bitset<NUM_BITS> & bitset) &;
 
     /// @brief Assignment operator from a string of '0's and '1's.
-    Bits & operator=(const std::string & bitstring) &;
+    Bits & operator=(const emp::String & bitstring) & { return Set(bitstring); }
 
     /// @brief Assignment operator from a literal string of '0's and '1's.
-    Bits & operator=(const char * bitstring) & { return operator=(std::string(bitstring)); }
+    Bits & operator=(const char * bitstring) & { return Set(emp::String(bitstring)); }
 
     /// @brief Assignment from another Bits object without changing size.
     template <typename DATA2_T, bool ZERO_LEFT2>
@@ -307,6 +308,9 @@ namespace emp {
 
     /// @brief Update the bit value at the specified index.
     Bits & Set(size_t index, bool value=true);
+
+    /// @brief Set to the value in the bitstring.
+    Bits & Set(emp::String str);
 
     /// @brief Set all bits to 1.
     Bits & SetAll();
@@ -745,20 +749,20 @@ namespace emp {
     [[nodiscard]] char GetAsChar(size_t id) const { return Get(id) ? '1' : '0'; }
 
     /// @brief Convert this Bits to a vector string [index 0 based on ZERO_LEFT]
-    [[nodiscard]] std::string ToString() const;
+    [[nodiscard]] emp::String ToString() const;
 
     /// @brief Convert this Bits to an array-based string [index 0 on left]
-    [[nodiscard]] std::string ToArrayString() const;
+    [[nodiscard]] emp::String ToArrayString() const;
 
     /// @brief Convert this Bits to a numerical string [index 0 on right]
-    [[nodiscard]] std::string ToBinaryString() const;
+    [[nodiscard]] emp::String ToBinaryString() const;
 
     /// @brief Convert this Bits to a series of IDs
-    [[nodiscard]] std::string ToIDString(const std::string & spacer=" ") const;
+    [[nodiscard]] emp::String ToIDString(const emp::String & spacer=" ") const;
 
     /// @brief Convert this Bits to a series of IDs with ranges condensed.
-    [[nodiscard]] std::string ToRangeString(const std::string & spacer=",",
-                                            const std::string & ranger="-") const;
+    [[nodiscard]] emp::String ToRangeString(const emp::String & spacer=",",
[…]
