### 2024-05-23 9e3a39a
Added PopIfAny() member function to String to pop single char.
