+++ include/emp/games/SudokuAnalyzer.hpp
@@ -22,6 +22,7 @@
 
 #include "emp/base/array.hpp"
 #include "emp/base/assert.hpp"
+#include "emp/bits/BitSet.hpp"
 #include "emp/math/Random.hpp"
 #include "emp/tools/String.hpp"
 
@@ -42,295 +43,403 @@ namespace emp {
     
     static constexpr uint8_t UNKNOWN_STATE = NUM_STATES; // Lower values are actual states.
 
-    // Which symbols are we using in this puzzle?
-    emp::array<char, NUM_STATES> symbols = {'1', '2', '3', '4', '5', '6', '7', '8', '9'};
+    using grid_bits_t = BitSet<NUM_CELLS>;
+    using region_bits_t = BitSet<NUM_REGIONS>;
 
-    emp::array<uint8_t,NUM_CELLS> value;         // Known value for cells
-    emp::array<uint32_t, NUM_CELLS> options;  // Options still available to each cell
+    // Track which cells are in each region.
+    static const emp::array<grid_bits_t, NUM_REGIONS> & RegionMap() {
+      static emp::array<grid_bits_t, NUM_REGIONS> bit_regions = BuildRegionMap();
+      return bit_regions;
+    }
+    static emp::array<grid_bits_t, NUM_REGIONS> BuildRegionMap() {
+      emp::array<grid_bits_t, NUM_REGIONS> bit_regions;
 
-    // "members" tracks which cell ids are members of each region.
-    static constexpr size_t members[NUM_REGIONS][NUM_STATES] = {
       // Rows (Overlaps: 111000000, 000111000, 000000111)
-      {  0,  1,  2,  3,  4,  5,  6,  7,  8 },  // Region 0  (Overlap with 18, 19, 20)
-      {  9, 10, 11, 12, 13, 14, 15, 16, 17 },  // Region 1  (Overlap with 18, 19, 20)
-      { 18, 19, 20, 21, 22, 23, 24, 25, 26 },  // Region 2  (Overlap with 18, 19, 20)
-      { 27, 28, 29, 30, 31, 32, 33, 34, 35 },  // Region 3  (Overlap with 21, 22, 23)
-      { 36, 37, 38, 39, 40, 41, 42, 43, 44 },  // Region 4  (Overlap with 21, 22, 23)
-      { 45, 46, 47, 48, 49, 50, 51, 52, 53 },  // Region 5  (Overlap with 21, 22, 23)
-      { 54, 55, 56, 57, 58, 59, 60, 61, 62 },  // Region 6  (Overlap with 24, 25, 26)
-      { 63, 64, 65, 66, 67, 68, 69, 70, 71 },  // Region 7  (Overlap with 24, 25, 26)
-      { 72, 73, 74, 75, 76, 77, 78, 79, 80 },  // Region 8  (Overlap with 24, 25, 26)
+      bit_regions[0 ] = "{  0,  1,  2,  3,  4,  5,  6,  7,  8 }"; // Overlap with 18, 19, 20
+      bit_regions[1 ] = "{  9, 10, 11, 12, 13, 14, 15, 16, 17 }"; // Overlap with 18, 19, 20
+      bit_regions[2 ] = "{ 18, 19, 20, 21, 22, 23, 24, 25, 26 }"; // Overlap with 18, 19, 20
[…]
