### 2024-05-23 1ff4ab1
Track region overlaps and use in solving.
