### 2024-05-23 a0580d0
Remove unused code from SudokuAnalyzer
