+++ include/emp/games/SudokuAnalyzer.hpp
@@ -181,312 +181,6 @@ namespace emp {
     emp::array<grid_bits_t, NUM_STATES> bit_options;
     grid_bits_t is_set;
 
-
-    // // "members" tracks which cell ids are members of each region.
-    // static constexpr size_t members[NUM_REGIONS][NUM_STATES] = {
-    //   // Rows (Overlaps: 111000000, 000111000, 000000111)
-    //   {  0,  1,  2,  3,  4,  5,  6,  7,  8 },  // Region 0  (Overlap with 18, 19, 20)
-    //   {  9, 10, 11, 12, 13, 14, 15, 16, 17 },  // Region 1  (Overlap with 18, 19, 20)
-    //   { 18, 19, 20, 21, 22, 23, 24, 25, 26 },  // Region 2  (Overlap with 18, 19, 20)
-    //   { 27, 28, 29, 30, 31, 32, 33, 34, 35 },  // Region 3  (Overlap with 21, 22, 23)
-    //   { 36, 37, 38, 39, 40, 41, 42, 43, 44 },  // Region 4  (Overlap with 21, 22, 23)
-    //   { 45, 46, 47, 48, 49, 50, 51, 52, 53 },  // Region 5  (Overlap with 21, 22, 23)
-    //   { 54, 55, 56, 57, 58, 59, 60, 61, 62 },  // Region 6  (Overlap with 24, 25, 26)
-    //   { 63, 64, 65, 66, 67, 68, 69, 70, 71 },  // Region 7  (Overlap with 24, 25, 26)
-    //   { 72, 73, 74, 75, 76, 77, 78, 79, 80 },  // Region 8  (Overlap with 24, 25, 26)
-
-    //   // Columns (Overlaps: 111000000, 000111000, 000000111)
-    //   { 0,  9, 18, 27, 36, 45, 54, 63, 72 },   // Region 9  (Overlap with 18, 21, 24)
-    //   { 1, 10, 19, 28, 37, 46, 55, 64, 73 },   // Region 10 (Overlap with 18, 21, 24)
-    //   { 2, 11, 20, 29, 38, 47, 56, 65, 74 },   // Region 11 (Overlap with 18, 21, 24)
-    //   { 3, 12, 21, 30, 39, 48, 57, 66, 75 },   // Region 12 (Overlap with 19, 22, 25)
-    //   { 4, 13, 22, 31, 40, 49, 58, 67, 76 },   // Region 13 (Overlap with 19, 22, 25)
-    //   { 5, 14, 23, 32, 41, 50, 59, 68, 77 },   // Region 14 (Overlap with 19, 22, 25)
-    //   { 6, 15, 24, 33, 42, 51, 60, 69, 78 },   // Region 15 (Overlap with 20, 23, 26)
-    //   { 7, 16, 25, 34, 43, 52, 61, 70, 79 },   // Region 16 (Overlap with 20, 23, 26)
-    //   { 8, 17, 26, 35, 44, 53, 62, 71, 80 },   // Region 17 (Overlap with 20, 23, 26)
-
[…]
