+++ paper.bib
@@ -296,3 +296,23 @@ keywords = {interface design, Interactive machine learning}
 
   copyright = {arXiv.org perpetual, non-exclusive license}
 }
+@misc{moreno2022best,
+  doi = {10.48550/ARXIV.2211.10897},
+  url = {https://arxiv.org/abs/2211.10897},
+  author = {Moreno,  Matthew Andres and Ofria,  Charles},
+  keywords = {Distributed,  Parallel,  and Cluster Computing (cs.DC),  FOS: Computer and information sciences,  FOS: Computer and information sciences},
+  title = {Best-Effort Communication Improves Performance and Scales Robustly on Conventional Hardware},
+  publisher = {arXiv},
+  year = {2022},
+  copyright = {arXiv.org perpetual,  non-exclusive license}
+}
+@misc{dolson2024phylotrack,
+  doi = {10.48550/ARXIV.2405.09389},
+  url = {https://arxiv.org/abs/2405.09389},
+  author = {Dolson,  Emily and Rodriguez-Papa,  Santiago and Moreno,  Matthew Andres},
+  keywords = {Populations and Evolution (q-bio.PE),  Neural and Evolutionary Computing (cs.NE),  FOS: Biological sciences,  FOS: Biological sciences,  FOS: Computer and information sciences,  FOS: Computer and information sciences},
+  title = {Phylotrack: C++ and Python libraries for in silico phylogenetic tracking},
+  publisher = {arXiv},
+  year = {2024},
+  copyright = {arXiv.org perpetual,  non-exclusive license}
+}
