### 2024-05-24 12d44b5
Fix bumpversion regex
