+++ .bumpversion.cfg
@@ -4,13 +4,11 @@ commit = True
 tag = True
 
 [bumpversion:file:third-party/package.json]
-search = version='{current_version}'
-replace = version='{new_version}'
+search = "version": "{current_version}"
+replace = "version": "{new_version}"
 
-[bumpversion:file:docs/conf.py]
-search = version='{current_version}'
-replace = version='{new_version}'
+[bumpversion:file:doc/conf.py]
 
 [bumpversion:file:CITATION.cff]
-search = version='{current_version}'
-replace = version='{new_version}'
+search = version: {current_version}
+replace = version: {new_version}
