### 2024-05-24 #6617 @mmore500
Hi @mahfuz05062 --- I have completed the author tasks after review is complete. The release version accompanying the paper is `v1.1.0` and the Zenodo DOI is <https://doi.org/10.5281/zenodo.11270974>
