+++ paper.md
@@ -167,7 +167,7 @@ Like Emscripten, Cheerp provides primarily low-level APIs for browser interactio
 * [ecology in evolutionary computation explorer](https://emilydolson.github.io/ecology_of_evolutionary_computation/web/interaction_networks.html) [@dolson2018evological]: interactive visualization of ecological interaction networks in evolutionary computation.
 * [Symbulation](https://anyaevostinar.github.io/SymbulationEmp/web/symbulation.html) [@vostinar2017suicide]: agent-based model for evolution of parasitism, mutualism, and commensalism.
 * [SignalGP](https://github.com/amlalejini/signalgp) [@Lalejini2018-GECCO; @moreno2021signalgp]: an event-driven genetic programming substrate.
-* [PhylotrackPy](https://github.com/emilydolson/phylotrackpy) [@dolson2024phylotrackpy]: a phylogeny-tracking tool for agent-based evolution, closely integrated with Empirical codebase.
+* [PhylotrackPy](https://github.com/emilydolson/phylotrackpy) [@dolson2024phylotrack]: a phylogeny-tracking tool for agent-based evolution, closely integrated with Empirical codebase.
 * [Model of cancer evolution on an oxygen gradient](http://emilydolson.github.io/memic_model/web/memic_model.html).
 
 # Acknowledgements
