### 2024-05-26 43358d5
Added lots of helpers for analyzing collections of bits.
