### 2024-05-26 6cfcbbe
Lots of cleanup on SudokuAnalyzer() + start of Swordfish implementation.
