+++ include/emp/games/SudokuAnalyzer.hpp
@@ -40,6 +40,7 @@ namespace emp {
     static constexpr size_t NUM_CELLS = NUM_ROWS * NUM_COLS;                  // 81
     static constexpr size_t NUM_REGIONS = NUM_ROWS + NUM_COLS + NUM_SQUARES;  // 27
     static constexpr size_t REGIONS_PER_CELL = 3;   // Each cell is part of three regions.
+    static constexpr size_t NO_REGION = static_cast<size_t>(-1);
 
     // Calculate multi-cell overlaps between regions; each row/col overlaps with 3 square regions.
     static constexpr size_t NUM_OVERLAPS = (NUM_ROWS + NUM_COLS) * 3 ;        // 54
@@ -95,6 +96,9 @@ namespace emp {
       return bit_regions;
     }
     static inline const grid_bits_t & RegionMap(size_t id) { return RegionMap()[id]; }
+    static inline const grid_bits_t & RowMap(size_t id) { return RegionMap()[id]; }
+    static inline const grid_bits_t & ColMap(size_t id) { return RegionMap()[id+NUM_ROWS]; }
+    static inline const grid_bits_t & BoxMap(size_t id) { return RegionMap()[id+NUM_ROWS+NUM_COLS]; }
 
 
     // Track which region ids each cell belongs to.
@@ -116,7 +120,9 @@ namespace emp {
 
       return cell_regions;
     }
-
+    static const region_bits_t CellMemberships(size_t cell_id) {
+      return CellMemberships()[cell_id];
+    }
 
     // Track which cells are in the same regions as each other cell.
     // E.g., Cell 23 is at CellLinks()[23], with 1's in all positions with neighbor cell.
@@ -329,6 +335,17 @@ namespace emp {
       }      
     }
 
+    // Scan for contradictions or lack of options that would make puzzle unsolvable.
+    bool IsUnsolvable() {
+      // Identify which cells are either set or still have options.
+      grid_bits_t has_options = is_set;
+      for (uint8_t state = 0; state < NUM_STATES; ++state) {
+        has_options |= bit_options[state];
+      }
+
+      return !has_options.All();
+    }
+
     // Use a brute-force approach to completely solve this puzzle.
     // Return true if solvable, false if unsolvable.
     // @CAO - Change to use a stack of backup states, remove recursion, and track set counts for faster ID of failure (i.e., no 3's left, but only 8 of 9 set)
@@ -426,7 +443,7 @@ namespace emp {
     
     // If K cells in a region are all limited to the same K states, eliminate those states
     // from all other cells in the same region.
-    emp::vector<PuzzleMove> Solve_FindLimitedCells() {
+    emp::vector<PuzzleMove> Solve_FindLimitedCells2() {
       std::cout << "TESTING 4:LimitedCells" << std::endl;
 
       emp::vector<PuzzleMove> moves;
@@ -470,8 +487,9 @@ namespace emp {
     
     // Eliminate all other possibilities from K cells if they are the only
     // ones that can possess K states in a single region.
-    emp::vector<PuzzleMove> Solve_FindLimitedStates() {
-      std::cout << "TESTING 4:LimitedStates" << std::endl;
+    emp::vector<PuzzleMove> Solve_FindLimitedStates2() {
+      std::cout << "TESTING 5:LimitedStates" << std::endl;
+      Print();
 
       emp::vector<PuzzleMove> moves;
 
@@ -479,6 +497,7 @@ namespace emp {
       for (uint8_t state1 = 0; state1 < NUM_STATES-1; ++state1) {
         for (uint8_t state2 = state1+1; state2 < NUM_STATES; ++state2) {
           grid_bits_t both_states = bit_options[state1] & bit_options[state2];
[…]
