### 2024-05-27 c1f019f
Closer to working Swordfist technique in Sudoku.
