### 2024-05-27 5da6aef
Fleshed out PuzzleMove, and added PuzzleSolveFun and PuzzleAnalyzer structs.
