### 2024-05-27 409cb6e
Moved SudokuInfo into its own struct and used is and PuzzleAnalyze as base classes for SudokuAnalyzer.
