### 2024-05-28 30c537c
Place many common puzzle-solving functions into GridPuzzleAnalyzer base class...
