### 2024-05-28 96de028
Use GridPuzzleAnalyzer base class for SudokuAnalyzer.
