+++ include/emp/games/SudokuAnalyzer.hpp
@@ -30,7 +30,8 @@
 
 namespace emp {
 
-  struct SudokuInfo {
+  class SudokuAnalyzer : public GridPuzzleAnalyzer<9, 9, 9> {
+  private:
     static constexpr size_t NUM_STATES = 9;
     static constexpr size_t NUM_ROWS = 9;
     static constexpr size_t NUM_COLS = 9;
@@ -42,18 +43,14 @@ namespace emp {
     static constexpr size_t NO_REGION = static_cast<size_t>(-1);
 
     // Calculate multi-cell overlaps between regions; each row/col overlaps with 3 square regions.
-    static constexpr size_t NUM_OVERLAPS = (NUM_ROWS + NUM_COLS) * 3 ;        // 54
-    
-    static constexpr uint8_t UNKNOWN_STATE = NUM_STATES; // Lower values are actual states.
-  };
+    static constexpr size_t NUM_OVERLAPS = (NUM_ROWS + NUM_COLS) * 3 ;        // 54    
 
-  class SudokuAnalyzer
-    : public SudokuInfo, public PuzzleAnalyzer<SudokuInfo::NUM_CELLS, SudokuInfo::NUM_STATES> {
-  private:
+    using base_t = GridPuzzleAnalyzer<NUM_ROWS, NUM_COLS, NUM_STATES>;
     using grid_bits_t = BitSet<NUM_CELLS>;
     using state_bits_t = BitSet<NUM_STATES>;
     using region_bits_t = BitSet<NUM_REGIONS>;
     using region_pair_t = std::pair<size_t, size_t>;
+    using symbol_set_t = emp::array<char, NUM_STATES>;
 
     // Track which cells are in each region.
     static const emp::array<grid_bits_t, NUM_REGIONS> & RegionMap() {
@@ -180,66 +177,28 @@ namespace emp {
       return out;
     }
 
-    ////////////////////////////////////
-    //                                //
-    //    --- Member Variables ---    //
-    //                                //
-    ////////////////////////////////////
 
-
-    // Which symbols are we using in this puzzle? (default to standard)
-    emp::array<char, NUM_STATES> symbols = {'1', '2', '3', '4', '5', '6', '7', '8', '9'};
-
-    emp::array<uint8_t,NUM_CELLS> value;      // Known value for cells
-
-    // Track the available options by state, across the whole puzzle.
-    // E.g., symbol 5 is at bit_options[5] and has 81 bits indicating if each cell can be '5'.
-    emp::array<grid_bits_t, NUM_STATES> bit_options;
-    grid_bits_t is_set;
-
-    
   public:
-    SudokuAnalyzer() { Clear(); }
+    SudokuAnalyzer() {
+      symbols = symbol_set_t{'1', '2', '3', '4', '5', '6', '7', '8', '9'};
+      Clear();
+    }
     SudokuAnalyzer(const SudokuAnalyzer &) = default;
     ~SudokuAnalyzer() { ; }
 
     SudokuAnalyzer& operator=(const SudokuAnalyzer &) = default;
 
-    uint8_t GetValue(size_t cell) const { return value[cell]; }
-    // uint32_t GetOptions(size_t cell) const { return options[cell]; }
-    size_t SymbolToState(char symbol) const {
-      for (size_t i=0; i < NUM_STATES; ++i) {
-        if (symbols[i] == symbol) return i;
+    // Set the value of an individual cell; remove option from linked cells.
+    // Return true/false based on whether progress was made toward solving the puzzle.
+    bool Set(size_t cell, uint8_t state) override {
+      if (base_t::Set(cell, state)) {
+        // Make sure this state is also blocked from all linked cells.
+        bit_options[state] &= ~CellLinks(cell);
[…]
