### 2024-05-28 ae1b13f
Cleanup and extra functionality on Bits.
