### 2024-05-28 b129120
Move CalcProfile into Puzzle.hpp + cleanup.
