### 2024-05-28 cf4cd73
Swordfish-2 now works for SudokuAnalyzer plus more integration with PuzzleAnalyzer base.
