+++ include/emp/games/SudokuAnalyzer.hpp
@@ -180,6 +180,15 @@ namespace emp {
 
   public:
     SudokuAnalyzer() {
+      // Set up the solver functions
+      AddSolveFunction("CellLastState",  1.0, [this](){ return Solve_FindLastCellState(); });
+      AddSolveFunction("RegionLastCell", 2.0, [this](){ return Solve_FindLastRegionState(); });
+      AddSolveFunction("RegionOverlap",  3.0, [this](){ return Solve_FindRegionOverlap(); });
+      AddSolveFunction("LimitedCells2",  4.0, [this](){ return Solve_FindLimitedCells2(); });
+      AddSolveFunction("LimitedStates2", 5.0, [this](){ return Solve_FindLimitedStates2(); });
+      AddSolveFunction("Swordfish2-RC",  6.0, [this](){ return Solve_FindSwordfish2_ROW_COL(); });
+      AddSolveFunction("Swordfish2-Box", 7.0, [this](){ return Solve_FindSwordfish2_BOX(); });
+
       symbols = symbol_set_t{'1', '2', '3', '4', '5', '6', '7', '8', '9'};
       Clear();
     }
@@ -225,11 +234,11 @@ namespace emp {
     }
 
     // Print the current state of the puzzle, including all options available.
-    void Print(std::ostream & out=std::cout) {
+    void Print(std::ostream & out=std::cout) override {
       out << " +-----------------------+-----------------------+-----------------------+"
           << std::endl;;
       for (size_t r = 0; r < 9; r++) {       // Puzzle row
-        for (size_t s = 0; s < 9; s+=3) {    // Subset row
+        for (uint8_t s = 0; s < 9; s+=3) {    // Subset row
           for (size_t c = 0; c < 9; c++) {   // Puzzle col
             size_t id = r*9+c;
             if (c%3==0) out << " |";
@@ -290,7 +299,6 @@ namespace emp {
 
     // If there's only one state a cell can be, pick it!
     emp::vector<PuzzleMove> Solve_FindLastCellState() {
-      std::cout << "TESTING 1:LastCState" << std::endl;
       emp::vector<PuzzleMove> moves;
 
       grid_bits_t unique_cells = emp::FindUniqueOnes(bit_options);
@@ -306,8 +314,6 @@ namespace emp {
 
     // If there's only one cell that can have a certain state in a region, choose it!
     emp::vector<PuzzleMove> Solve_FindLastRegionState() {
-      std::cout << "TESTING 2:LastRState" << std::endl;
-
       emp::vector<PuzzleMove> moves;
 
       // Loop through each state for each region testing.
@@ -326,7 +332,6 @@ namespace emp {
     // If only cells that can have a state in region A are all also in region
     // B, no other cell in region B can have that state as a possibility.
     emp::vector<PuzzleMove> Solve_FindRegionOverlap() {
-      std::cout << "TESTING 3:RegionOverlap" << std::endl;
       emp::vector<PuzzleMove> moves;
 
       for (auto [r1, r2] : RegionOverlaps()) {
@@ -357,8 +362,6 @@ namespace emp {
     // If K cells in a region are all limited to the same K states, eliminate those states
     // from all other cells in the same region.
     emp::vector<PuzzleMove> Solve_FindLimitedCells2() {
-      std::cout << "TESTING 4:LimitedCells" << std::endl;
-
       emp::vector<PuzzleMove> moves;
 
       // Identify which sites have exactly two options.
@@ -401,9 +404,6 @@ namespace emp {
     // Eliminate all other possibilities from K cells if they are the only
     // ones that can possess K states in a single region.
     emp::vector<PuzzleMove> Solve_FindLimitedStates2() {
[…]
