### 2024-05-28 a5fd8d6
Add tools for full-color ascii printing of puzzle boards.
