### 2024-05-28 3a7a532
Added Bits too for finding groups of three ones.
