### 2024-05-28 29134c3
Added color printing and moves for 3-cell versions of FindLimited{Cells,States}
