+++ include/emp/games/SudokuAnalyzer.hpp
@@ -188,6 +188,8 @@ namespace emp {
       AddSolveFunction("LimitedStates2", 5.0, [this](){ return Solve_FindLimitedStates2(); });
       AddSolveFunction("Swordfish2-RC",  6.0, [this](){ return Solve_FindSwordfish2_ROW_COL(); });
       AddSolveFunction("Swordfish2-Box", 7.0, [this](){ return Solve_FindSwordfish2_BOX(); });
+      AddSolveFunction("LimitedCells3",  8.0, [this](){ return Solve_FindLimitedCells3(); });
+      AddSolveFunction("LimitedStates3", 9.0, [this](){ return Solve_FindLimitedStates3(); });
 
       symbols = symbol_set_t{'1', '2', '3', '4', '5', '6', '7', '8', '9'};
       Clear();
@@ -234,7 +236,7 @@ namespace emp {
     }
 
     // Print the current state of the puzzle, including all options available.
-    void Print(std::ostream & out=std::cout) override {
+    void Print(bool color=true, std::ostream & out=std::cout) override {
       out << " +-----------------------+-----------------------+-----------------------+"
           << std::endl;;
       for (size_t r = 0; r < 9; r++) {       // Puzzle row
@@ -244,12 +246,12 @@ namespace emp {
             if (c%3==0) out << " |";
             else out << "  ";
             if (value[id] == UNKNOWN_STATE) {
-              out << " " << (char) (HasOption(id,s)   ? symbols[s] : '.')
-                  << " " << (char) (HasOption(id,s+1) ? symbols[s+1] : '.')
-                  << " " << (char) (HasOption(id,s+2) ? symbols[s+2] : '.');
+              out << " " << (HasOption(id,s)   ? ColorSymbol(s) : ".")
+                  << " " << (HasOption(id,s+1) ? ColorSymbol(s+1) : ".")
+                  << " " << (HasOption(id,s+2) ? ColorSymbol(s+2) : ".");
             } else {
               if (s==0) out << "      ";
-              if (s==3) out << "   " << symbols[value[id]] << "  ";
+              if (s==3) out << "   " << ColorSymbol(value[id], true) << "  ";
               if (s==6) out << "      ";
               // if (s==0) out << " /   \\";
               // if (s==3) out << " | " << symbols[value[id]] << " |";
@@ -359,7 +361,7 @@ namespace emp {
       return moves;
     }
     
-    // If K cells in a region are all limited to the same K states, eliminate those states
+    // If 2 cells in a region are all limited to the same 2 states, eliminate those states
     // from all other cells in the same region.
     emp::vector<PuzzleMove> Solve_FindLimitedCells2() {
       emp::vector<PuzzleMove> moves;
@@ -396,11 +398,67 @@ namespace emp {
         }
       }
 
+      return moves;
+    }
+
+
+    // If 3 cells in a region are all limited to the same 3 states, eliminate those states
+    // from all other cells in the same region.
+    emp::vector<PuzzleMove> Solve_FindLimitedCells3() {
+      emp::vector<PuzzleMove> moves;
+
+      // Identify which sites have exactly two or three options.
+      grid_bits_t two_ones = FindTwoOnes(bit_options);
+      grid_bits_t three_ones = FindThreeOnes(bit_options);
+      grid_bits_t valid = two_ones | three_ones;
+
+      // Try all (9*8*7/6=84) triples of states and measure which sites have these options.
+      for (uint8_t state1 = 0; state1 < NUM_STATES-1; ++state1) {
+        for (uint8_t state2 = state1+1; state2 < NUM_STATES; ++state2) {
+          for (uint8_t state3 = state2+1; state3 < NUM_STATES; ++state3) {
+
[…]
