### 2024-05-29 15ce585
Added Swordfish 3 and 4
