+++ include/emp/games/SudokuAnalyzer.hpp
@@ -190,6 +190,8 @@ namespace emp {
       AddSolveFunction("Swordfish2-Box", 7.0, [this](){ return Solve_FindSwordfish2_BOX(); });
       AddSolveFunction("LimitedCells3",  8.0, [this](){ return Solve_FindLimitedCells3(); });
       AddSolveFunction("LimitedStates3", 9.0, [this](){ return Solve_FindLimitedStates3(); });
+      AddSolveFunction("Swordfish3-RC",  10.0, [this](){ return Solve_FindSwordfish3_ROW_COL(); });
+      AddSolveFunction("Swordfish4-RC",  11.0, [this](){ return Solve_FindSwordfish4_ROW_COL(); });
 
       symbols = symbol_set_t{'1', '2', '3', '4', '5', '6', '7', '8', '9'};
       Clear();
@@ -459,8 +461,8 @@ namespace emp {
     }
 
 
-    // Eliminate all other possibilities from K cells if they are the only
-    // ones that can possess K states in a single region.
+    // Eliminate all other possibilities from 2 cells if they are the only
+    // ones that can possess 2 states in a single region.
     emp::vector<PuzzleMove> Solve_FindLimitedStates2() {
       emp::vector<PuzzleMove> moves;
 
@@ -505,6 +507,8 @@ namespace emp {
       return moves;
     }
 
+    // Eliminate all other possibilities from 3 cells if they are the only
+    // ones that can possess 3 states in a single region.
     emp::vector<PuzzleMove> Solve_FindLimitedStates3() {
       emp::vector<PuzzleMove> moves;
 
@@ -535,33 +539,19 @@ namespace emp {
               for (uint8_t block_state = 0; block_state < NUM_STATES; ++block_state) {
                 if (block_state == state1 || block_state == state2 || block_state == state3) continue;
                 if (bit_options[block_state].Has(pos1)) {
-                  std::cout << "STATES " << (state1+1) << (state2+1) << (state3+1)
-                    << " at sites "
-                    << CellToCoords(pos1) << ", "
-                    << CellToCoords(pos2) << ", "
-                    << CellToCoords(pos3) << ", "
-                    << "blocking state " << (block_state+1) << " at pos1"
-                    << std::endl;
+                  // std::cout << "STATES " << (state1+1) << (state2+1) << (state3+1)
+                  //   << " at sites "
+                  //   << CellToCoords(pos1) << ", "
+                  //   << CellToCoords(pos2) << ", "
+                  //   << CellToCoords(pos3) << ", "
+                  //   << "blocking state " << (block_state+1) << " at pos1"
+                  //   << std::endl;
                   moves.push_back(PuzzleMove{PuzzleMove::BLOCK_STATE, pos1, block_state});                
                 }
                 if (bit_options[block_state].Has(pos2)) {
-                  std::cout << "STATES " << (state1+1) << (state2+1) << (state3+1)
-                    << " at sites "
-                    << CellToCoords(pos1) << ", "
-                    << CellToCoords(pos2) << ", "
-                    << CellToCoords(pos3) << ", "
-                    << "blocking state " << (block_state+1) << " at pos2"
-                    << std::endl;
                   moves.push_back(PuzzleMove{PuzzleMove::BLOCK_STATE, pos2, block_state});                
                 }
                 if (bit_options[block_state].Has(pos3)) {
-                  std::cout << "STATES " << (state1+1) << (state2+1) << (state3+1)
-                    << " at sites "
-                    << CellToCoords(pos1) << ", "
-                    << CellToCoords(pos2) << ", "
-                    << CellToCoords(pos3) << ", "
-                    << "blocking state " << (block_state+1) << " at pos3"
[…]
