### 2024-05-29 c518038
Added a ComboSet object that steps through all combinations in a compile-time specified range
