### 2024-05-30 e784149
Add JOSS badge
