### 2024-05-31 910580b
Setup more sophisticated scoring in Puzzle.hpp
