+++ include/emp/games/Puzzle.hpp
@@ -61,6 +61,8 @@ namespace emp {
     std::vector<Slice> slices; // Actual solving profile.
     FinalState final_state = UNSOLVED;
 
+    size_t size() const { return slices.size(); }
+
     bool IsSolved() const { return final_state == SOLVED; }
     bool IsUnsolved() const { return final_state == UNSOLVED; }
     bool IsUnsolvable() const { return final_state == UNSOLVABLE; }
@@ -73,7 +75,32 @@ namespace emp {
     void SetUnsolved()   { final_state = UNSOLVED; }
     void SetUnsolvable() { final_state = UNSOLVABLE; }
     
-    void Clear() { slices.resize(0); }
+    void Clear() { slices.resize(0); final_state = UNSOLVED; }
+
+    size_t CountTypes() const {
+      BitVector types;
+      for (Slice slice : slices) {
+        if (types.size() <= slice.level) types.resize(slice.level+8);
+        types.Set(slice.level);
+      }
+      return types.CountOnes();
+    }
+
+    size_t CountMoves(size_t id) const {
+      size_t count = 0;
+      for (Slice slice : slices) {
+        if (slice.level == id) ++count;
+      }
+      return count;
+    }
+
+    double CalcScore() const {
+      double score = 0.0;
+      for (Slice slice : slices) {
+        score += slice.level + 1;
+      }
+      return score;
+    }
 
     String ToString() const {
       String out;
@@ -116,6 +143,10 @@ namespace emp {
 
     uint8_t GetValue(size_t cell) const { return value[cell]; }
 
+    bool HasValue(size_t cell) const { return GetValue() != UNKNOWN_STATE; }
+
+    size_t GetNumSolveFuns() const { return solve_funs.size(); }
+
     uint8_t SymbolToState(char symbol) const {
       for (uint8_t i=0; i < NUM_STATES; ++i) {
         if (symbols[i] == symbol) return i;
@@ -123,13 +154,13 @@ namespace emp {
       return UNKNOWN_STATE;
     }
 
-    // Convert a cell ID to its coordinates.
+    /// Convert a cell ID to its coordinates.
     emp::String CellToCoords(size_t id) {
       return emp::MakeString(id%NUM_COLS, ',', id/NUM_COLS);
     }
 
     /// Test if a cell is allowed to be a particular state.
-    bool HasOption(size_t cell, uint8_t state) {
+    bool HasOption(size_t cell, uint8_t state) const {
       emp_assert(cell < 81, cell);
       emp_assert(state >= 0 && state < NUM_STATES, state);
       return bit_options[state].Has(cell);
@@ -145,7 +176,7 @@ namespace emp {
     bool IsSet(size_t cell) const { return value[cell] != UNKNOWN_STATE; }
     bool IsSolved() const { return is_set.All(); }
     
-    // Clear out the old solution info when starting a new solve attempt.
+    /// Clear out the old solution info when starting a new solve attempt.
     void Clear() {
       value.fill(UNKNOWN_STATE);
       for (auto & val_options : bit_options) {
@@ -163,8 +194,8 @@ namespace emp {
       solve_funs.push_back(PuzzleSolveFun{name, difficulty, fun});
     }
 
-    // Set the value of an individual cell
-    // Return true/false based on whether progress was made toward solving the puzzle.
+    /// Set the value of an individual cell
+    /// Return true/false based on whether progress was made toward solving the puzzle.
     virtual bool Set(size_t cell, uint8_t state) {
       emp_assert(cell < NUM_CELLS);   // Make sure cell is in a valid range.
       emp_assert(state < NUM_STATES); // Make sure state is in a valid range.
@@ -181,32 +212,65 @@ namespace emp {
       return true;
     }
 
-    // Remove a symbol option from a particular cell.
[…]
