### 2024-06-01 #6617 @mahfuz05062
@mmore500 I found several issues with the citation/reference. See below:

- Is there a reson for using {} in the reference for "Not So Fast: Analyzing the Performance of WebAssembly vs. Native Code". It seems unusual.
- The link you provided for "Suicide, signals, and symbionts: Evolving cooperation in agent-based systems" requires a login to the university of michigan portal. Maybe remove that link and mention that it's a Dissertation.
- You have two references for shiny. For the second one, you added a link for "pyodide.org". I believe this is meant for pyodide. Please fix this.
