### 2024-06-01 df00e6f
Fix jangda bibtex entry
