+++ paper.bib
@@ -11,15 +11,18 @@
 }
 
 @inproceedings{jangda2019not,
-  title={Not so fast: Analyzing the performance of webassembly vs. native code},
-  author={Jangda, Abhinav and Powers, Bobby and Berger, Emery D and Guha, Arjun},
-  booktitle={2019 $\{$USENIX$\}$ Annual Technical Conference ($\{$USENIX$\}$$\{$ATC$\}$ 19)},
-  pages={107--120},
-  year={2019},
-  month={7},
-  isbn={9781939133038},
-  url={https://www.usenix.org/conference/atc19/presentation/jangda},
-  publisher={USENIX Association}
+  author = {Jangda, Abhinav and Powers, Bobby and Berger, Emery D. and Guha, Arjun},
+  title = {Not so fast: analyzing the performance of webassembly vs. native code},
+  year = {2019},
+  isbn = {9781939133038},
+  publisher = {USENIX Association},
+  address = {USA},
+  booktitle = {Proceedings of the 2019 USENIX Conference on Usenix Annual Technical Conference},
+  pages = {107-120},
+  numpages = {14},
+  location = {Renton, WA, USA},
+  series = {USENIX ATC '19},
+  url = {https://www.usenix.org/conference/atc19/presentation/jangda},
 }
 
 @article{kazi2000techniques,
