+++ paper.bib
@@ -107,12 +107,16 @@
 }
 
 @Manual{developers2021pyodide,
-  title= {shiny: Web Application Framework for R},
-  author = {Michael Droettboom and Pyodide Developers},
-  year = {2021},
-  note = {Python package version 0.0.17},
-  url = {https://pyodide.org/}
+  author       = {Michael Droettboom and the Pyodide development team},
+  title        = {pyodide/pyodide},
+  month        = aug,
+  year         = 2021,
+  publisher    = {Zenodo},
+  version      = {0.26.0},
+  doi          = {10.5281/zenodo.5156931},
+  url          = {https://doi.org/10.5281/zenodo.5156931}
 }
+
 
 @article{helikar2012cell,
   title={The cell collective: toward an open and collaborative approach to systems biology},
