### 2024-06-01 a95ee77
Fix vostinar disseration url, note is thesis
