+++ paper.bib
@@ -132,13 +132,13 @@
   publisher={Springer}
 }
 
-@book{vostinar2017suicide,
+@phdthesis{vostinar2017suicide,
   title={Suicide, signals, and symbionts: Evolving cooperation in agent-based systems},
   author={Vostinar, Anya E},
   year={2017},
-  url={https://proxy.lib.umich.edu/login?url=https://www.proquest.com/dissertations-theses/suicide-signals-symbionts-evolving-cooperation/docview/1929231148/se-2?accountid=14667},
+  url={https://www.proquest.com/docview/1929231148},
   isbn={978-0-355-07992-0},
-  publisher={Michigan State University}
+  publisher={Doctoral dissertation, Michigan State University}
 }
 
 @article{gillespie2018changing,
