### 2024-06-01 da2a53f
Bump version: 1.1.1 → 1.1.2
