### 2024-06-01 c5464f4
Remove redundant doctoral dissertation note
