+++ paper.bib
@@ -138,7 +138,7 @@
   year={2017},
   url={https://www.proquest.com/docview/1929231148},
   isbn={978-0-355-07992-0},
-  publisher={Doctoral dissertation, Michigan State University}
+  publisher={Michigan State University}
 }
 
 @article{gillespie2018changing,
