+++ doc/conf.py
@@ -93,9 +93,9 @@ author = u"Charles Ofria"
 # the built documents.
 #
 # The short X.Y version.
-version = '1.1.2'
+version = '1.1.3'
 # The full version, including alpha/beta/rc tags.
-release = '1.1.2'
+release = '1.1.3'
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
 #
