### 2024-06-01 8d5a992
Make citation.cff title match JOSS paper
