### 2024-06-01 a4c14fc
Update @mmore500 affiliation in citation.cff
