+++ CITATION.cff
@@ -10,7 +10,7 @@ authors:
   affiliation: Michigan State University
 - given-names: Matthew Andres
   family-names: Moreno
-  affiliation: Michigan State University
+  affiliation: University of Michigan
   orcid: 'https://orcid.org/0000-0003-4726-4479'
 - given-names: Emily
   family-names: Dolson
