### 2024-06-01 #6617 @mmore500
Updated the Zenodo deposit to exactly match the JOSS title. Unfortunately, I wasn't able to edit the metadata (I think I might not have permissions for that in the @devosoft organization's Zenodo account). However, I did create a new Zenodo deposit.

software version: v1.1.4
doi: https://doi.org/10.5281/zenodo.11407351
