### 2024-06-01 #6617 @mmore500
I just checked the proof and the reference changes are now rendering correctly. :+1:
