### 2024-06-01 5f8295d
Added a load from span to SudokuAnalyzer, and a PrintSimple
