### 2024-06-01 #6617 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
