### 2024-06-02 #6617 @danielskatz
@mmore500 - As track editor, I've proofread the paper, which looks good, except for some minor issues in the references. Please merge https://github.com/devosoft/Empirical/pull/520 or let me know what you disagree with.
