### 2024-06-02 #6617 @danielskatz
@mmore500 - Also, I see the metadata in the zenodo repository doesn't match the paper, which we ask to happen. I see some discussion about this in https://github.com/openjournals/joss-reviews/issues/6617#issuecomment-2143293431. How should we resolve this?
