### 2024-06-02 #6617 @danielskatz
> What is the nature of the mismatch in the Zenodo repository? I think the titles should exactly match now. Is it the author list?

yes.
