### 2024-06-02 #6617 @mmore500
I will release to zenodo with an udpated author list. Should take about 10 minutes I expect.
