### 2024-06-02 #6617 @danielskatz
you should just be able to update the metadata - you don't need to make a new release or deposit

Also, I probably won't get back to this for 6-24 hours
