### 2024-06-02 a17eb7e
Update CITATION.cff authors to exactly match JOSS
