+++ CITATION.cff
@@ -5,53 +5,26 @@ message: If you use this software, please cite it using these metadata.
 title: 'Empirical: A scientific software library for research, education, and public engagement'
 doi: 10.5281/zenodo.4141943
 authors:
+- given-names: Anya
+  family-names: Vostinar
+  orcid: 'https://orcid.org/0000-0001-7216-5283'
+  affiliation: Carleton College
+- given-names: Alexander
+  family-names: Lalejini
+  orcid: 'https://orcid.org/0000-0003-0994-2718'
+  affiliation: Grand Valley State University
 - given-names: Charles
   family-names: Ofria
+  orcid: 'https://orcid.org/0000-0003-2924-1732'
+  affiliation: Michigan State University
+- given-names: Emily
+  family-names: Dolson
+  orcid: 'https://orcid.org/0000-0001-8616-4898'
   affiliation: Michigan State University
 - given-names: Matthew Andres
   family-names: Moreno
-  affiliation: University of Michigan
   orcid: 'https://orcid.org/0000-0003-4726-4479'
-- given-names: Emily
-  family-names: Dolson
-  affiliation: Michigan State University
-- given-names: Alex
-  family-names: Lalejini
-- given-names: Santiago
-  family-names: Rodriguez Papa
-- given-names: Jake
-  family-names: Fenton
-- given-names: Katherine
-  family-names: Perry
-- given-names: Steven
-  family-names: Jorgensen
-- given-names: hoffmanriley
-  family-names: ''
-- given-names: grenewode
-  family-names: ''
-- given-names: Oliver
-  family-names: Edwards
-  name-particle: Baldwin
-  affiliation: Amherst College
-- given-names: Jason
-  family-names: Stredwick
-  affiliation: Appdiff Inc. (test.ai)
-- given-names: cgnitash
-  family-names: ''
-- given-names: theycallmeHeem
-  family-names: ''
-- given-names: Anya
-  family-names: Vostinar
-- given-names: Ryan
-  family-names: Moreno
-- given-names: Jory
-  family-names: Schossau
-  affiliation: Michigan State University
-- given-names: Luis
-  family-names: Zaman
-  affiliation: University of Washington
-- given-names: djrain
-  family-names: ''
+  affiliation: University of Michigan
 version: 1.1.4
 repository-code: https://github.com/devosoft/Empirical
 license: MIT
