+++ CITATION.cff
@@ -25,6 +25,6 @@ authors:
   family-names: Moreno
   orcid: 'https://orcid.org/0000-0003-4726-4479'
   affiliation: University of Michigan
-version: 1.1.4
+version: 1.1.5
 repository-code: https://github.com/devosoft/Empirical
 license: MIT
