### 2024-06-02 #6617 @danielskatz
Congratulations to @mmore500 (Matthew Andres Moreno) and co-authors on your publication!!

JOSS depends on volunteers and wouldn't be successful without your work
